% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{addDashLines}
\alias{addDashLines}
\title{Add dash lines to \code{data.frame}}
\usage{
addDashLines(.df, .vLine = 0)
}
\arguments{
\item{.df}{Dataset}

\item{.vLine}{positional index of Vertical Line \code{|} in dataset}
}
\value{
modifided \code{data.frame}
}
\description{
This adds dash lines to the dataframe in the first and last rows.
}
