% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append.R
\name{append}
\alias{append}
\title{Append multiple datasets}
\usage{
append(data, ...)
}
\arguments{
\item{data}{Base dataset}

\item{...}{Datasets to be appended}
}
\value{
Modified Dataset
}
\description{
\code{append()} joins multiple datasets of the same column names.
}
\details{
Multiple datasets can be appended only if they have the same variables in the
same order.

\preformatted{append(data, data1, data2, data3, etc)}
}
\examples{
\dontrun{


# Example from IDRE UCLA
path <- "https://stats.idre.ucla.edu/stat/data/patient_pt1_stata_dm.dta"
hosp <- haven::read_dta(path)
codebook(hosp)



## use piping function
library(magrittr)


## appending three times
append(hosp, hosp, hosp, hosp) \%>\%
    codebook


## appending directly from web
append(hosp, haven::read_dta(path)) \%>\%
    codebook
}

}
\author{
For any feedback, please contact \code{Myo Minn Oo} via:

Email: \email{dr.myominnoo@gmail.com}

Website: \url{https://myominnoo.github.io/}
}
