% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nscore.R
\name{backtr}
\alias{backtr}
\title{Back-transform normal score transformed data.}
\usage{
backtr(model, tails = "none", draw = FALSE)
}
\arguments{
\item{model}{Dataframe with column of raw data matched to column of
normal score data. Is output of `nscore` function.}

\item{tails}{Treatment of distribution tails: "none", "separate", or "equal".}

\item{draw}{Plot the distributions.}
}
\value{
Numeric vector of back-transformed values.
}
\description{
Given a vector of normal scores and a normal score object (from nscore), the
function returns a vector of back-transformed values. One major issue is how
to extrapolate to the tails. Options  other than none may result in
dramatically incorrect tail estimates!
}
\details{
Tails options:
    'none': No extrapolation; more extreme score values will revert to the
    original min and max values.
    'equal': Calculate magnitude in std deviations of the scores about
    initial data mean. Extrapolation is linear to these deviations.  Will be
    based upon deviations from the mean of the original hard data - possibly
    quite dangerous!
    'separate': This calculates a separate sd for values above and below the
    mean.
}
\author{
Alex M Trueman
Modified from Ashton Shortridge, May/June, 2008.
https://msu.edu/~ashton/research/code/nscore.R
Originally based on the GSLIB code for backtr function.
}
