% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot-strap.R
\name{boot_wmean_ci}
\alias{boot_wmean_ci}
\title{Calculate weighted mean and bootstrapped confidence limits using the bias
corrected and accelerated (BCa) method, which is more robust with low sample
numbers and non-normal distributions of bootstrapped means.}
\usage{
boot_wmean_ci(df, value, weight, reps, conf, ...)
}
\arguments{
\item{df}{Data frame containing grouping, value, and weighting fields.}

\item{value}{Value field for statistics.}

\item{weight}{Weighting field.}

\item{reps}{Number of bootstrapping repetitions. May need to increase if
BCa error occurs.}

\item{conf}{Confidence limits probability (e.g., 0.95).}

\item{...}{One or more grouping columns.}
}
\value{
Data frame with weighted mean and confidence limits.
}
\description{
Based on https://www.painblogr.org/2017-10-18-purrring-through-bootstraps
by Peter Kamerman (@painblogR)
}
\examples{
df <- data.frame(
x = rlnorm(n = 100, meanlog = 1.5, sdlog = 1.1),
w = runif(n = 100, min = 0.1, max = 5),
d = sample(c("a", "b"), 100, replace = TRUE)
)

boot_wmean_ci(df, x, w, reps = 100, conf = 0.9, d)

}
\author{
Alex M Trueman
}
