% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{codebook}
\alias{codebook}
\alias{codebook.default}
\alias{codebook.data.frame}
\title{Codebook: Detailed information about variables}
\usage{
codebook(data)

\method{codebook}{default}(data)

\method{codebook}{data.frame}(data)
}
\arguments{
\item{data}{dataframe}
}
\description{
\code{codebook()} provides detailed information about each variable
within the dataset.
}
\details{
\code{codebook} generates the report of data structure with names,
data lables, types,
number of observations, number of observations with missing values and
percentage of observations with missing values.

\strong{ANNOTATIONS}:

VARS_NAME - Names of variables

LABEL     - Labels of variables

TYPE      - Types of

OBS_COUNT - Counts of valid observations

NA_COUNT  - Counts of observations with missing value

NA_(\%)    - Percentage of observations with missing value
}
\examples{
\dontrun{
codebook(infert)
codebook(iris)

# if something else
codebook(iris$Species)



# Example from IDRE UCLA
path <- "https://stats.idre.ucla.edu/stat/data/patient_pt1_stata_dm.dta"
hosp <- haven::read_dta(path)
codebook(hosp)

library(magrittr)
hosp \%>\% labelData("Hospital Dataset Example") \%>\% codebook()
}
}
\references{
STATA DATA MANAGEMENT. UCLA: Statistical Consulting Group.
from https://stats.idre.ucla.edu/stata/seminars/stata-data-management/
(accessed Febrary 25, 2020).
}
\seealso{
\code{\link{ilog}}, \code{\link{listView}}
}
\author{
For any feedback, please contact \code{Myo Minn Oo} via:

Email: \email{dr.myominnoo@gmail.com}

Website: \url{https://myominnoo.github.io/}
}
\keyword{codebook,}
\keyword{layout}
\keyword{structure,}
\keyword{summary,}
