% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode_decode.R
\name{coder}
\alias{coder}
\title{Apply functions to multiple variables in a data frame}
\usage{
coder(.data, .fns)
}
\arguments{
\item{.data}{The input data frame.}

\item{.fns}{A set of functions to be applied to the variables in the data frame.}
}
\value{
A modified data frame with the functions applied to the variables.
}
\description{
The \code{coder()} function applies a set of functions to multiple variables in a data frame using \code{dplyr}'s \code{mutate()} and \code{across()} functions.
}
\examples{
# Apply the `mean()` function to multiple variables in the `mtcars` data frame
coder(mtcars, mean)

}
