% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect}
\alias{collect}
\alias{collect.tab1}
\alias{collect.tab2}
\title{Collect tables from functions of \code{mStats}}
\usage{
collect(data)

\method{collect}{tab1}(data)

\method{collect}{tab2}(data)
}
\arguments{
\item{data}{An output from \code{mStats}functions}
}
\description{
\Sexpr[results=rd]{lifecycle::badge("experimental")}

\code{collect()} retrieves data from outputs of \code{mStats} functions and
creates a well-formatted output that can be:
\itemize{
\item exported to \code{csv} or \code{xlsx} format for further processing.
\item wraped in rmarkdown with \code{flextable} package to produce
publication-ready tables.
}
}
\section{Examples}{

\subsection{Summary measures for categorical variables}{

Create summary table for frequency measure:\preformatted{table1 <- tab(infert, education, parity, case)
collect(table1)
}

Create summary table stratified by a stratification variable \code{by}:\preformatted{table2 <- tab(infert, education, parity, by = case)
collect(table2)
}
}
}

\concept{reporting}
