% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compact.R
\name{compact}
\alias{compact}
\title{Compact publication-ready tables}
\usage{
compact(x, y)
}
\arguments{
\item{x}{data frames, or objects to be coerced to one.}

\item{y}{data frames, or objects to be coerced to one.}
}
\description{
\code{mStats} provides several S3 functions for \code{\link[=summary]{summary()}} to format the
output and prepare for publication-ready tables.
}
\section{Examples}{

\subsection{Summary measures for categorical variables}{

Create summary table for frequency measure:\if{html}{\out{<div class="r">}}\preformatted{table1 <- tab(infert, education, parity)
#>  One-way tabulation
#>  + --------- ------- + ----- ------- ----- +
#>  |  Variable   Level | Freq. Percent  Cum. |
#>  + --------- ------- + ----- ------- ----- +
#>  | education         |                     |
#>  |            0-5yrs |    12     4.8   4.8 |
#>  |           6-11yrs |   120    48.4  53.2 |
#>  |           12+ yrs |   116    46.8 100.0 |
#>  +           ------- + ----- ------- ----- +
#>  |             Total |   248   100.0 100.0 |
#>  + --------- ------- + ----- ------- ----- +
#>  |    parity         |                     |
#>  |                 1 |    99    39.9  39.9 |
#>  |                 2 |    81    32.7  72.6 |
#>  |                 3 |    36    14.5  87.1 |
#>  |                 4 |    18     7.3  94.4 |
#>  |                 5 |     6     2.4  96.8 |
#>  |                 6 |     8     3.2 100.0 |
#>  +           ------- + ----- ------- ----- +
#>  |             Total |   248   100.0 100.0 |
#>  + --------- ------- + ----- ------- ----- +

table1_sum <- summary(table1)

table2 <- tab(infert, education, parity, by = case)
#>  Two-way tabulation
#>                               case
#>  + --------- ------- + --- ---- -- ---- ----- ----- ----- ----- +
#>  |  Variable   Level |   0 r(\%)  1 r(\%) Total  r(\%)  chi2 exact |
#>  + --------- ------- + --- ---- -- ---- ----- ----- ----- ----- +
#>  | education         |                                          |
#>  |            0-5yrs |   8 66.7  4 33.3    12 100.0 0.999 1.000 |
#>  |           6-11yrs |  80 66.7 40 33.3   120 100.0             |
#>  |           12+ yrs |  77 66.4 39 33.6   116 100.0             |
#>  +           ------- + --- ---- -- ---- ----- ----- ----- ----- +
#>  |             Total | 165 66.5 83 33.5   248 100.0             |
#>  + --------- ------- + --- ---- -- ---- ----- ----- ----- ----- +
#>  |    parity         |                                          |
#>  |                 1 |  66 66.7 33 33.3    99 100.0 1.000 1.000 |
#>  |                 2 |  54 66.7 27 33.3    81 100.0             |
#>  |                 3 |  24 66.7 12 33.3    36 100.0             |
#>  |                 4 |  12 66.7  6 33.3    18 100.0             |
#>  |                 5 |   4 66.7  2 33.3     6 100.0             |
#>  |                 6 |   5 62.5  3 37.5     8 100.0             |
#>  +           ------- + --- ---- -- ---- ----- ----- ----- ----- +
#>  |             Total | 165 66.5 83 33.5   248 100.0             |
#>  + --------- ------- + --- ---- -- ---- ----- ----- ----- ----- +

table2_sum <- summary(table2)

compact(table1_sum, table2_sum)
#>     Variable   Level Total\\n n (\%)  0\\n n (\%) 1\\n n (\%) p-value\\n (Chi2)
#> 1      Total           248 (100.0) 165 (66.5) 83 (33.5)                 
#> 2  education                                                       0.999
#> 3             0-5yrs      12 (4.8)   8 (66.7)  4 (33.3)                 
#> 4            6-11yrs    120 (48.4)  80 (66.7) 40 (33.3)                 
#> 5            12+ yrs    116 (46.8)  77 (66.4) 39 (33.6)                 
#> 6     parity                                                       1.000
#> 7                  1     99 (39.9)  66 (66.7) 33 (33.3)                 
#> 8                  2     81 (32.7)  54 (66.7) 27 (33.3)                 
#> 9                  3     36 (14.5)  24 (66.7) 12 (33.3)                 
#> 10                 4      18 (7.3)  12 (66.7)  6 (33.3)                 
#> 11                 5       6 (2.4)   4 (66.7)  2 (33.3)                 
#> 12                 6       8 (3.2)   5 (62.5)  3 (37.5)                 
#>    p-value\\n (Exact)
#> 1                   
#> 2              1.000
#> 3                   
#> 4                   
#> 5                   
#> 6              1.000
#> 7                   
#> 8                   
#> 9                   
#> 10                  
#> 11                  
#> 12
}\if{html}{\out{</div>}}
}
}

\seealso{
Other reporting: 
\code{\link{summary}()}
}
\concept{reporting}
