% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countBy.R
\name{countBy}
\alias{countBy}
\title{Count from N to N}
\usage{
countBy(data, ..., var_name = NULL, N = FALSE)
}
\arguments{
\item{data}{Dataset}

\item{...}{Variables to find duplications.}

\item{var_name}{name of variable for serial counting}

\item{N}{if \code{TRUE}, it generates total number of observations for
each observation. If \code{FALSE}, it produces sets of serial numbers
from 1 to total number of observations per group.}
}
\value{
Modified ordered dataset with one additional variable
}
\description{
\code{countBy()} produces serial counting numbers or
total number of observations per each observation.
}
\details{
If variables are not specified within \code{...} arguments, all variables
are used to order first and create unique identifier for each
observation.

If variables are specified, unique identifiers are generated based
on the combination of variables, and grouped. Serial numbers are then
produced based on each group.

If \code{N} is set to \code{TRUE}, total number of observations is generated
per each observation for the whole dataset or per group. A new variable
\code{.N_} is created in the dataset.

If \code{FALSE}, serial numbers are generated per each observation for
the whole dataset or per group. A new variable \code{.n_} is created in the
dataset.

Variable's name can be set by specifying \code{var_name}.
}
\examples{

## Example from UCLA Website
## https://stats.idre.ucla.edu/stata/seminars/notes/counting-from-_n-to-_n/
df <- data.frame(matrix(
    c(72, 1, 84, 2, 76, 1, 89, 3, 82, 2, 90, 1, 85, 1),
    byrow = TRUE, ncol = 2))
names(df) <- c("score", "group")

countBy(df)
countBy(df, N = TRUE)

## count per group
countBy(df, group)
countBy(df, group, var_name = "n1")

## count per group (total observation)
countBy(df, group, N = TRUE)
countBy(df, group, var_name = "n2", N = TRUE)

}
\author{
Email: \email{dr.myominnoo@gmail.com}

Website: \url{https://myominnoo.github.io/}
}
