% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox.output.R
\name{cox.output}
\alias{cox.output}
\title{Process Cox survival model and Display Hazard Ratios}
\usage{
cox.output(covar, time, event, data, aHR = FALSE, write.csv = FALSE,
  file = NULL, print.to.console = FALSE)
}
\arguments{
\item{covar}{a vector containing all the covariates to be fitted}

\item{time}{time variable for right censored data, this is the follow up time. For interval data, the first argument is the starting time for the interval.}

\item{event}{The status indicator, normally 0=alive, 1=dead. Other choices are TRUE/FALSE (TRUE = death) or 1/2 (2=death). For interval censored data, the status indicator is 0=right censored, 1=event at time, 2=left censored, 3=interval censored. For multiple enpoint data the event variable will be a factor, whose first level is treated as censoring. Although unusual, the event indicator can be omitted, in which case all subjects are assumed to have an event.}

\item{data}{a data frame containing the variables in the model.}

\item{aHR}{a logical value indicating either unadjusted or adjusted model}

\item{write.csv}{a logical value to save the cox table output to CSV file.}

\item{file}{either a character string naming a file.}

\item{print.to.console}{print the table to the console even if the output is assigned to an object.}
}
\description{
Takes several covar and generates Hazard Ratios outputs at the same time. Outputs include unadjusted / adjusted HR, 95% CI and P-values.
}
\seealso{
itab, isum, inumsum
}
