% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateSum.R
\name{dateSum}
\alias{dateSum}
\title{Summarise a Date variable}
\usage{
dateSum(x, groupBy = "y", decreasing = FALSE, rnd = 1,
  plot.show = TRUE, xlab = NULL)
}
\arguments{
\item{x}{a Date vector}

\item{groupBy}{grouping by one of three date components: 
\enumerate{
  \item \strong{y}: Year alone
  \item \strong{ym}: Year and Month
  \item \strong{ymd}: Year, Month and Day
}}

\item{decreasing}{order the data by year, month and day sequence}

\item{rnd}{specify rounding of numbers}

\item{plot.show}{display bar plots of the data}

\item{xlab}{x-axis label}
}
\description{
\code{dateSum} tabulates dates as factor, based on date components: Day, Month, Year. 
The frequency distribution can be also displayed.
}
\examples{
# random dates generation
set.seed(1)
dates <- toDate(rnorm(100, mean = 4000, sd = 100), "e")
dateSum(dates) # by default, pattern is set to Year alone.
dateSum(dates, "y")
dateSum(dates, "ym")
dateSum(dates, "ym", xlab = "Random Dates")
dateSum(dates, "ym", xlab = "Random Dates", plot.show = FALSE)
dateSum(dates, "ymd", plot.show = FALSE)
dateSum(dates, "ymd") # this can take times.
}
\seealso{
\code{\link{itab}}, \code{\link{ibarplot}}, \code{\link{isum}}, 
\code{\link{iboxplot}}
}
\author{
Myo Minn Oo (Email: \email{dr.myominnoo@gmail.com} |
Website: \url{https://myominnoo.github.io/})
}
\keyword{aggregate}
\keyword{date}
\keyword{summary,}
