% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatDate.R
\name{day}
\alias{day}
\title{Day from a Date}
\usage{
day(x)
}
\arguments{
\item{x}{Date Object}
}
\description{
\code{day} retrieves day from a Date
}
\examples{

dates <- formatDate(c("2019-01-15", "2019-01-20", "2019-01-21", "2019-01-22"),
                    "Ymd", "-")
day(dates)
}
\seealso{
\code{\link{formatDate}}, \code{\link{year}}, \code{\link{day}}
}
\author{
For any feedback, please contact \code{Myo Minn Oo} via:

Email: \email{dr.myominnoo@gmail.com}

Website: \url{https://myominnoo.github.io/}
}
\keyword{date}
\keyword{day,}
