% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode_decode.R
\name{decode}
\alias{decode}
\title{Decode variables in a data frame}
\usage{
decode(.data, ...)
}
\arguments{
\item{.data}{The input data frame.}

\item{...}{The variables to be decoded.}
}
\value{
A modified data frame with the selected variables decoded.
}
\description{
The \code{decode()} function decodes selected variables in a data frame. It converts the variables to character type, sets labels from the original data, and returns the modified data frame.
}
\examples{
# Decode selected variables in the `mtcars` data frame
decode(mtcars, mpg, cyl)

}
