% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{describe}
\alias{describe}
\alias{des}
\title{Describe a summary of data contents}
\usage{
describe(data)

des(data)
}
\arguments{
\item{data}{A data.frame}
}
\description{
\Sexpr[results=rd]{lifecycle::badge("stable")}

\code{describe()} displays a summary of data contents.
It includes important characteristics including
number of variables, variable names, label, number of unique values,
missing values, mean, minimum and maximum.
}
\section{Examples}{
\preformatted{describe(iris)
describe(infert)
}
}

\seealso{
Other data management: 
\code{\link{append}()},
\code{\link{codebook}()},
\code{\link{duplicates}()},
\code{\link{egen}()},
\code{\link{encode}()},
\code{\link{expand2}()},
\code{\link{formatDate}()},
\code{\link{generate}()},
\code{\link{label}()},
\code{\link{n_}()},
\code{\link{recode}()},
\code{\link{replace}()},
\code{\link{rowmiss}()}
}
\concept{data management}
