% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumstats.R
\name{descstat}
\alias{descstat}
\title{Descriptive statistics for a series of columns.}
\usage{
descstat(df, ..., sep = "_(?=[^_]+$)")
}
\arguments{
\item{df}{Dataframe}

\item{...}{List of column names to be evaluated.}

\item{sep}{Character string separating original column name and statistic.}
}
\value{
Dataframe of statistics for each column input.
}
\description{
Descriptive statistics for a series of columns.
}
\author{
Alex M Trueman
}
