% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detachAll.R
\name{detachAll}
\alias{detachAll}
\title{Detach all objects in the search pathway}
\usage{
detachAll()
}
\arguments{
\item{NA}{No parameter required.}
}
\description{
detachAll() lets you detach two types of objects in the search pathway:
1) user-installed packages
2) attached data objects
}
\examples{
detachAll()
}
\keyword{data,}
\keyword{detach,}
\keyword{packages,}
\keyword{search}
