% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{dup.rm}
\alias{dup.rm}
\title{Remove Duplicated Elements}
\usage{
dup.rm(x, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric object}

\item{na.rm}{A logical value to specify missing values}
}
\description{
\code{dup.rm} determines which elements of a vector are duplicates and removes them.
}
\details{
This becomes handy when a return value with duplicates removed is need. It checks if
any elements of vector are duplicated and remove them at the same time.
}
\examples{
## Example 1
a <- rep(c("A", "B", "C"), c(10, 20, 30))
a
dup.rm(a)

## Example 2
b <- rep(c("A", "B", "C", NA), c(10, 20, 30, 10))
b
dup.rm(b)
dup.rm(b, na.rm = TRUE)
}
\seealso{
\code{\link{label}}
}
\author{
Myo Minn Oo (Email: \email{dr.myominnoo@gmail.com} |
Website: \url{https://myominnoo.github.io/})
}
\keyword{duplicate}
\keyword{duplicate,}
\keyword{remove}
