% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egen.R
\name{egen}
\alias{egen}
\alias{egen.default}
\alias{egen.data.frame}
\title{An extension to generate function}
\usage{
egen(data = NULL, old.var, cut = NULL, lbl = NULL, new.var = NULL,
  na.rm = FALSE)

\method{egen}{default}(data = NULL, old.var, cut = NULL, lbl = NULL,
  new.var = NULL, na.rm = FALSE)

\method{egen}{data.frame}(data = NULL, old.var, cut = NULL,
  lbl = NULL, new.var = NULL, na.rm = FALSE)
}
\arguments{
\item{data}{dataframe}

\item{old.var}{variable to perform cut}

\item{cut}{either a single number or a numeric vector.}

\item{lbl}{specify to name the factor levels.}

\item{new.var}{name of new variable generated}

\item{na.rm}{A logical value to specify missing values}
}
\description{
\code{egen} transforms a numeric vector to a factor vector.
}
\details{
\code{egen} allows easy conversion of a numeric vector to factor.

\strong{Cut-off Intervals}
If the interval is not specified, it is cut at an interval of 10. Otherwise,
it is divided into equal cut-off points by specified number.

\strong{Labelling}
If not specified, the labels are constructed in the format:
\strong{variable name} + "." + \strong{"cut-off intervals"}.
}
\examples{
\dontrun{
## variable with dataframe
# automatically categorized into interval of 10
infert.new <- egen(infert, age)
str(infert.new)
tab(age.cat, infert.new)

infert.new <- egen(infert, age, c(30, 35, 40))
str(infert.new)
tab(age.cat, infert.new)

infert.new <- egen(infert, age, c(30, 35, 40),
                c("young", "mid", "old", "oldest"))
str(infert.new)
tab(age.cat, infert.new)

# give variable name as age.grp
infert.new <- egen(infert, age, c(30, 35, 40),
                c("young", "mid", "old", "oldest"), age.grp)
str(infert.new)
tab(age.grp, infert.new)

# age.grp duplicates
infert.new1 <- egen(infert.new, age, c(30, 35, 40),
                c("young", "mid", "old", "oldest"), age.grp)
}
}
\seealso{
\code{\link{generate}}
}
\author{
Myo Minn Oo (Email: \email{dr.myominnoo@gmail.com} |
Website: \url{https://myominnoo.github.io/})
}
\keyword{correlation}
\keyword{distribution,}
\keyword{number}
\keyword{summary,}
