% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode.R
\name{encode}
\alias{encode}
\alias{decode}
\title{Encode variables into numeric type and vice versa}
\usage{
encode(data, ...)

decode(data, ...)
}
\arguments{
\item{data}{A data.frame}

\item{...}{One or more unquoted variables separated by commas.
Variable names can be used as if they were positions in the data
frame, so expressions like \code{x:y} can be used to select a range of
variables.}
}
\description{
\Sexpr[results=rd]{lifecycle::badge("stable")}

\code{encode()} converts variables into \code{numeric} type.

\code{decode()} converts variables into \code{character} type.
}
\section{Examples}{
\preformatted{infert <- encode(infert, education, stratum)
codebook(infert)

iris <- decode(iris)
codebook(iris)
}
}

\seealso{
Other data management: 
\code{\link{append}()},
\code{\link{codebook}()},
\code{\link{describe}()},
\code{\link{duplicates}()},
\code{\link{egen}()},
\code{\link{expand2}()},
\code{\link{formatDate}()},
\code{\link{generate}()},
\code{\link{label}()},
\code{\link{n_}()},
\code{\link{recode}()},
\code{\link{replace}()},
\code{\link{rowmiss}()}
}
\concept{data management}
