% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{enquos}
\alias{enquos}
\alias{checkEnquos}
\title{Get names within three dots in a function}
\usage{
enquos(.args, .argsName)

checkEnquos(.data, .vars, .types = "tab")
}
\arguments{
\item{.args}{argument lists from \code{as.list(match.call())}}

\item{.argsName}{Names of arguments to be removed.}

\item{.data}{Dataset}

\item{.vars}{.vars obtained with \code{as.character(enquos())}}

\item{.types}{"tab" to get categorical data; otherwise, numerical data}
}
\value{
A character vector
}
\description{
\code{enquos} retrieves names of arguments from \code{as.list(match.call())} and
returns names within three dots \code{...}.

\code{checkEnquos()} checks colon separators as well as variables whether they
are specified or not. if not, it gets all variables within the dataset.
}
\section{Functions}{
\itemize{
\item \code{checkEnquos}: 
}}

