% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esttab.R
\name{esttab}
\alias{esttab}
\title{Display formatted regression table}
\usage{
esttab(...)
}
\arguments{
\item{...}{regression functions from \code{mStats}}
}
\description{
\code{esttab()} produces publication-quality
regression tables which can be exported.
}
\details{
It is an wrapper function after running regression models.
Supported functions include \code{\link{regress}} and
\code{\link{logit}}. It processes their outputs and generates
a publication quality tabulation for model comparisons.
}
\examples{

## use infert data
data(infert)

## running different logistic regression models
m1 <- logit(infert, case, education)
m2 <- logit(infert, case, education, age)
m3 <- logit(infert, case, education, age, parity, induced)

## creating regression tables
esttab(m1, m2, m3)

}
\author{
For any feedback, please contact \code{Myo Minn Oo} via:

Email: \email{dr.myominnoo@gmail.com}

Website: \url{https://myominnoo.github.io/}
}
