% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand2.R
\name{expand2}
\alias{expand2}
\title{Duplicate observations}
\usage{
expand2(data, copies = 2, n_n = 1:nrow(data))
}
\arguments{
\item{data}{A data.frame}

\item{copies}{\code{n} copies of observations}

\item{n_n}{Row index to duplicate. By default, the value is \code{1:nrow(data)}.}
}
\description{
\Sexpr[results=rd]{lifecycle::badge("stable")}

\code{expand2()} duplicates observations to have \code{n} copies of observations.
}
\section{Examples}{
\preformatted{set.seed(123)
test <- data.frame(id = 1:5, age = rnorm(5, 30, 10))
test <- expand2(test, copies = 2, n_n = 1:3)

## check the duplicates
duplicates(test)
}
}

\seealso{
Other data management: 
\code{\link{append}()},
\code{\link{codebook}()},
\code{\link{describe}()},
\code{\link{duplicates}()},
\code{\link{egen}()},
\code{\link{encode}()},
\code{\link{formatDate}()},
\code{\link{generate}()},
\code{\link{label}()},
\code{\link{n_}()},
\code{\link{recode}()},
\code{\link{replace}()},
\code{\link{rowmiss}()}
}
\concept{data management}
