% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expandTables.R
\name{expandTables}
\alias{expandTables}
\title{Convert \code{2x2 tables} to \code{data.frame}}
\usage{
expandTables(
  ...,
  exp_name = "exposure",
  exp_lvl = c("exposed", "unexposed"),
  case_name = "outcome",
  case_lvl = c("disease", "healthy"),
  strata_name = "strata",
  stringsAsFactors = FALSE
)
}
\arguments{
\item{...}{vectors of 2x2 tables}

\item{exp_name}{Name of \code{Exposure} Variable}

\item{exp_lvl}{Names of two categories in the order of
Exposed and non-exposed}

\item{case_name}{Name of \code{Case} variable}

\item{case_lvl}{names of two categories in the order of}

\item{strata_name}{Name of stratified variable}

\item{stringsAsFactors}{\code{TRUE} or \code{FALSE}
If \code{TRUE}, character vector is converted to a factor when
\code{data.frame} is constructed.}
}
\value{
\code{data.frame}
}
\description{
\code{expandTables} generates a data frame and supports two levels.
}
\details{
\code{expandTables} uses the vectors of \code{2x2} tables and
generates a data frame of at least two columns:
Exposure and Outcome.

\preformatted{expandTables(c(100, 200, 100, 200))}

\strong{Strata}

Strata can be included using multiple named vectors. The
generated tables can be used for further calculation such as
Mantel Haenszel methods.

\preformatted{
expandTables(
             strata1 = c(100, 200, 100, 200),
             strata2 = c(100, 200, 100, 200),
             strata3 = c(100, 200, 100, 200),
             exp_name = "Exposure",
             exp_lvl = c("exposed", "unexposed"),
             case_name = "Outcome",
             case_lvl = c("case", "control"),
             strata_name = "Strata"
)
}

\strong{Labels}

If variables' names or levels are not specified, the followings are
applied.

\enumerate{
    \item Exposure Name: \code{exposure}
    \item Exposure levels: \code{exposed} and \code{unexposed}
    \item Outcome Name: \code{outcome}
    \item Outcome levels: \code{disease} and \code{healthy}
    \item Strata Name: \code{strata}
    \item Note: Strata levels are not considered as vectors must
    be named.
}
}
\examples{
\dontrun{
## creating a table
df <- expandTables(c(60, 140, 60, 40))
tab(df)
tab(df, exposure)
tab(df, exposure, by = outcome)



## one table with full labels
df <- expandTables(
    c(60, 140, 60, 40),
    exp_name = "areaType",
    exp_lvl = c("Rural", "urban"),
    case_name = "Antibodies",
    case_lvl = c("Yes", "No"))
tab(df)
tab(df, areaType, by = Antibodies)



## TWo tables
df <- expandTables(
    male = c(36, 14, 50, 50),
    female = c(24, 126, 10, 90),
    exp_name = "areaType", exp_lvl = c("Rural", "Urban"),
    case_name = "Antibodies", case_lvl = c("Yes", "No"),
    strata_name = "gender")
tab(df)
tab(df, areaType, by = Antibodies)



### checking numbers
df \%>\%
    pick(gender == "male") \%>\%
    tab(areaType, Antibodies)

df \%>\%
    pick(gender == "female") \%>\%
    tab(areaType, Antibodies)
}
}
\seealso{
\code{\link{mhodds}}, \code{\link{stmh}}

\code{\link{tab}}
}
\author{
For any feedback, please contact \code{Myo Minn Oo} via:

Email: \email{dr.myominnoo@gmail.com}

Website: \url{https://myominnoo.github.io/}
}
\keyword{contigency}
\keyword{expand}
\keyword{tables}
\keyword{tables,}
\keyword{two-by-two}
