% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalize.R
\name{finalize}
\alias{finalize}
\title{Finalizing summary measures}
\usage{
finalize(x)
}
\arguments{
\item{x}{output from summary measure functions}
}
\description{
\code{finalize()} organizes summary measures into table format
to be used in manuscript writing.
}
\details{
Functions allowed:
\code{tab} shows freq (percentage). The function can handle both one-way and
cross-tabulation.

\code{summ} shows mean (SD), median (IQR), Normality.
For grouped measures, p-values from comparison tests are reported.
}
\examples{

## one-way tabulation
result <- tab(infert, education, parity, case)
finalize(result)

## cross-tabulation
result <- tab(infert, education, parity, by = case)
finalize(result)

## summary measures
result <- summ(infert)
finalize(result)

result <- summ(infert, by = education)
finalize(result)


}
\author{
Email: \email{dr.myominnoo@gmail.com}

Website: \url{https://myominnoo.github.io/}
}
