% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fModelOutput.R
\name{fModelOutput}
\alias{fModelOutput}
\alias{fModelOutput.default}
\alias{fModelOutput.unAdjusted}
\alias{fModelOutput.adjusted}
\title{Report parameters from regression models}
\usage{
fModelOutput(model, rnd = 1, print.table = TRUE)

\method{fModelOutput}{default}(...)

\method{fModelOutput}{unAdjusted}(model, rnd = 1, print.table = TRUE)

\method{fModelOutput}{adjusted}(model, rnd = 1, print.table = TRUE)
}
\arguments{
\item{model}{object glm or lm type}

\item{rnd}{rounding digits}

\item{print.table}{logical value to display formatted outputs}

\item{...}{optional arguments}
}
\description{
\code{fModelOutput} display corresponding estimates of different models along
with 95% confidence intervals and p-values.
}
\details{
\code{fModelOutput}

reports parameters from regression models. Currently supporting model is
logistic models. Other types will be incoporated in future works.

In unadjusted analysis, it reports unajdusted odds ratios, 95% Confidence
interval and Wald's p-value.

In adjusted analysis, both unadjusted and adjusted parameters are reported.

The reports are in well-formatted texts which can be readily copied into
spreadsheet programs or format in word file.

\strong{Please be reminded:}

This is an ongoing work. In case of comments or suggestions, please contact me
at \email{dr.myominnoo@gmail.com}.

\strong{Reference: }
\enumerate{
  \item LOGIT REGRESSION | R DATA ANALYSIS EXAMPLES. UCLA: Statistical Consulting
  Group. from https://stats.idre.ucla.edu/r/dae/logit-regression/ (accessed
  September 27, 2019)
}
}
\examples{
\dontrun{
## example from IRDE website:
## https://stats.idre.ucla.edu/r/dae/logit-regression/
mydata <- read.csv("https://stats.idre.ucla.edu/stat/data/binary.csv")
codebook(mydata)
tab(admit, mydata)
tab(rank, mydata)

logit.gre <- glm(admit ~ gre, data = mydata, family = "binomial")
summary(logit.gre)
fModelOutput(logit.gre)
testModelFit(logit.gre)

logit.multi <- glm(admit ~ gre + gpa + factor(rank), data = mydata, family = "binomial")
summary(logit.multi)

fModelOutput(logit.multi) # generates parameters
testModelFit(logit.multi) # test overall significant of the model
}
}
\seealso{
\code{\link{testModelFit}}
}
\author{
Myo Minn Oo (Email: \email{dr.myominnoo@gmail.com} |
Website: \url{https://myominnoo.github.io/})
}
\keyword{coefficients}
\keyword{display}
\keyword{model}
\keyword{output,}
