% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatDate.R
\name{formatDate}
\alias{formatDate}
\alias{is.Date}
\alias{year}
\alias{month}
\alias{day}
\title{Format Dates}
\usage{
formatDate(x, format = "dmY", sep = "/", century = NULL)

is.Date(x)

year(x)

month(x)

day(x)
}
\arguments{
\item{x}{a character or numeric object}

\item{format}{only for character vectors:}

\item{sep}{separator character for date components}

\item{century}{specify either 2000 or 1900 for two-digit years}
}
\description{
\Sexpr[results=rd]{lifecycle::badge("stable")}
\itemize{
\item \code{formatDate()} converts characters or numbers to \code{Dates}.
\item \code{is.Date()} indicates which elements are \code{Dates}.
\item \code{year()} extract the numeric \code{year} corresponding to \code{Dates}.
\item \code{month()} extract the numeric \code{month} corresponding to \code{Dates}.
\item \code{day()} extract the numeric \code{day} corresponding to \code{Dates}.
}

\code{dmy} represents \code{dd mm YYYY} format.
In combination with separators from \code{sep}, this can change to
several date formats.
For example, \code{dmy} + \code{-} convert to
\code{dd-mm-yyyy} format.

\strong{Possible conversions}

\enumerate{
\item \code{dmy} + \code{-} >>> \code{dd-mm-yyyy}
\item \code{dmy} + \code{/} >>> \code{dd/mm/yyyy}
\item \code{mdy} + \code{/} >>> \code{mm/dd/yyyy}
\item \code{ymd} + \code{/} >>> \code{yyyy/mm/dd}
\item \code{dby} + \code{-} >>> \code{dd-JAN-yy}
\item \code{dby} + \code{/} >>> \code{dd/JAN/yy}
}

\strong{Numeric conversions}
Origin is set at \code{1899-12-30}.
}
\section{Examples}{
\preformatted{## convert strings to dates
x <- c("2019-01-15", "2019-01-20", "2019-01-21", "2019-01-22")

# check if it is a Date format
is.Date(x)


y <- formatDate(x, "Ymd", "-")

# check if it is a Date format
is.Date(y)
y


## another format
x <- c("22-JAN-19", "24-MAR-20")
y <- formatDate(x, "dby", "-")
is.Date(y)
y


## convert numbers to dates
x <- 42705:42710
y <- formatDate(x)
is.Date(y)
y


## get day, month or year
day(y)
month(y)
year(y)
}
}

\seealso{
Other data management: 
\code{\link{append}()},
\code{\link{codebook}()},
\code{\link{describe}()},
\code{\link{duplicates}()},
\code{\link{egen}()},
\code{\link{encode}()},
\code{\link{expand2}()},
\code{\link{generate}()},
\code{\link{label}()},
\code{\link{n_}()},
\code{\link{recode}()},
\code{\link{replace}()},
\code{\link{rowmiss}()}
}
\concept{data management}
