% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.R
\name{gen}
\alias{gen}
\title{Create a new variable within the data frame}
\usage{
gen(data, var, expr = NULL)
}
\arguments{
\item{data}{data frame}

\item{var}{name of a new variable}

\item{expr}{value or Expression: See examples below.}
}
\description{
\code{gen} function allows R user to easily create a new variable within the data frame
}
\details{
\code{gen} (generate) creates a new variable within the data frame. The value of the
variable are specified by \code{expr} argument. If \code{expr} is not specified,
\code{gen} generates \code{NA} by default.
}
\examples{
data(infert)
df <- infert
str(df)
df <- gen(df, age.grp, ifelse(age < 25, 1, ifelse(age >= 25 & age < 35, 2, 3)))
table(df$age.grp)
df$age.grp <- label(df$age.grp, c("young", "middle", "old"))
table(df$age.grp)

## create new variable with mean value
gen(infert, age.mean, mean(age))
}
\seealso{
\code{\link{label}}
}
\author{
Myo Minn Oo (Email: \email{dr.myominnoo@gmail.com} |
Website: \url{https://myominnoo.github.io/})
}
\keyword{label}
\keyword{label,}
\keyword{levels}
\keyword{value}
