% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sentiment.R
\name{get_sentiment}
\alias{get_sentiment}
\title{Analyze the theme of lyrics}
\usage{
get_sentiment(lyrics_df, song_title)
}
\arguments{
\item{lyrics_df}{A dataframe containing song lyrics, including columns named 'song' and 'lyrics'}

\item{song_title}{The title of the song to analyze}
}
\value{
A data frame with sentiment score, and number of positive and negative words
}
\description{
Perform sentiment analysis using the Bing lexicon from the tidytext package
}
\examples{
get_sentiment(downward_lyrics, "Line")
}
