% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{helpers}
\alias{helpers}
\alias{getRDataTypes}
\alias{enquotes}
\alias{checkEnquos}
\alias{checkVarName}
\alias{splitByColon}
\alias{addDashLines}
\alias{findLineWidth}
\alias{wrapText}
\alias{printLines}
\alias{printText}
\alias{printLabel}
\alias{printDF}
\alias{rowColOrder}
\alias{splitTables}
\title{Helper functions for \code{mStats}}
\usage{
helpers(...)

getRDataTypes()

enquotes(.args, .args_name)

checkEnquos(.data, .vars, .return = NULL)

checkVarName(.data, .var)

splitByColon(.data, .vars, .colon)

addDashLines(.data, .vline = NULL)

findLineWidth(.data)

wrapText(.txt, .width = 70, .sep = "\\n")

printLines(.x = "_", .width = 70)

printText(.txt = NULL)

printLabel(.data, .var_name = NULL)

printDF(.data, .txt, .split = NULL)

rowColOrder(.tbl, .exp.value = NULL, .case.value = NULL)

splitTables(.tbl, .exp.value = NULL)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=addDashLines]{addDashLines}}, \code{\link[=addDashLines]{addDashLines}}, \code{\link[=addDashLines]{addDashLines}}, \code{\link[=addDashLines]{addDashLines}}, \code{\link[=printLines]{printLines}}, \code{\link[=wrapText]{wrapText}}, \code{\link[=addDashLines]{addDashLines}}, \code{\link[=addDashLines]{addDashLines}}, \code{\link[=wrapText]{wrapText}}
  \describe{
    \item{\code{}}{}
  }}

\item{.args}{argument lists from \code{as.list(match.call())}}

\item{.args_name}{Names of arguments to be removed.}

\item{.data}{Dataset}

\item{.vars}{Variables}

\item{.return}{used when no variables are specified within three dots:
If it is set to "tab", it retrieves variables for tabulation.
(factors, ordered factor, characters, logical)

If to "summ", it gets the ones for summary measures.
("numeric", "double", "integer", "logical")}

\item{.var}{variable name to be checked}

\item{.colon}{Logical. \code{TRUE} indicates containing colon.}

\item{.vline}{positional index of Vertical Line:
to add \code{|} in dataset. If \code{NULL}, it adds dash lines without
vertical lines}

\item{.txt}{texts}

\item{.width}{desired character length to display}

\item{.sep}{separator for line break}

\item{.x}{vector, matrix, dataframe or separator (in case of printLines)}

\item{.var_name}{character vector indicating names of variables}

\item{.split}{a text that separate between two lines}

\item{.tbl}{table}

\item{.exp.value}{exposure values
modifided \code{data.frame}}

\item{.case.value}{case values}
}
\value{
modifided \code{data.frame}
}
\description{
This is a collection of all helper functions for the package.
}
\section{Functions}{
\itemize{
\item \code{getRDataTypes}: This creates four character vectors of names of data types in R.

\item \code{enquotes}: \code{enquotes} retrieves names of arguments from \code{as.list(match.call())} and
returns names within three dots \code{...}.

\item \code{checkEnquos}: \code{checkEnquos} checks colon separators within three dots.
if yes, it gets ranges of corresponding variables within the dataset.
if not, return the same variables.

\item \code{checkVarName}: \code{checkEnquos} checks colon separators within three dots.
if yes, it gets ranges of corresponding variables within the dataset.
if not, return the same variables.

\item \code{splitByColon}: Split variable names separated by colon

\item \code{addDashLines}: This adds dash lines to the dataframe in the first and last rows.

\item \code{findLineWidth}: It calculates line width for printing lines

\item \code{wrapText}: It wraps texts or messages within
a specified length.

\item \code{printLines}: It produces lines as strings for a
specified length

\item \code{printText}: It print texts within round brackets.

\item \code{printLabel}: It retrieves and print labels.
If only dataset is specified, it prints dataset label.
Otherwise, it prints variable or variables' labels.

\item \code{printDF}: This prints \code{data.frame} in a well-formatted style.
It also adds \code{printLines} for better visualization.

\item \code{rowColOrder}: This reorders table format by exposure and case values.

\item \code{splitTables}: This separates one tables into several by a reference value.
}}

