% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{histogram}
\alias{histogram}
\title{Histograms with overlay normal curve}
\usage{
histogram(
  data,
  var,
  breaks = NULL,
  xlab = NULL,
  main = NULL,
  sub = NULL,
  labels = TRUE,
  freq = TRUE,
  curve = TRUE,
  ...
)
}
\arguments{
\item{data}{Dataset}

\item{var}{variable}

\item{breaks}{one of:
    \itemize{
      \item a vector giving the breakpoints between histogram cells,
      \item a function to compute the vector of breakpoints,
      \item a single number giving the number of cells for the histogram,
      \item a character string naming an algorithm to compute the
      number of cells (see \sQuote{Details}),
      \item a function to compute the number of cells.
    }
    In the last three cases the number is a suggestion only; as the
    breakpoints will be set to \code{\link{pretty}} values, the number
    is limited to \code{1e6} (with a warning if it was larger).  If
    \code{breaks} is a function, the \code{x} vector is supplied to it
    as the only argument (and the number of breaks is only limited by
    the amount of available memory).
  }

\item{xlab}{main title and axis labels: these arguments to
    \code{\link[graphics]{title}()} get \dQuote{smart} defaults here, e.g., the default
    \code{ylab} is \code{"Frequency"} iff \code{freq} is true.}

\item{main}{main title and axis labels: these arguments to
    \code{\link[graphics]{title}()} get \dQuote{smart} defaults here, e.g., the default
    \code{ylab} is \code{"Frequency"} iff \code{freq} is true.}

\item{sub}{Sub-title (at bottom) using font, size and color 
    \code{par(c("font.sub", "cex.sub", "col.sub"))}.}

\item{labels}{logical or character string.  Additionally draw labels on top
    of bars, if not \code{FALSE}; see \code{\link[graphics]{plot.histogram}}.}

\item{freq}{logical; if \code{TRUE}, the histogram graphic is a
    representation of frequencies, the \code{counts} component of
    the result; if \code{FALSE}, probability densities, component
    \code{density}, are plotted (so that the histogram has a total area
    of one).  Defaults to \code{TRUE} \emph{if and only if} \code{breaks} are
    equidistant (and \code{probability} is not specified).}

\item{curve}{logical. If \code{TRUE} (default), a normal curve is
overlaid over the histogram.}

\item{...}{further arguments and \link[graphics]{graphical parameters} passed to
    \code{\link[graphics]{plot.histogram}} and thence to \code{\link[graphics]{title}} and
    \code{\link[graphics]{axis}} (if \code{plot = TRUE}).}
}
\description{
\code{summ()} draws a histogram with formatted texts and
adds a normal curve over the histogram.
}
\details{
If \code{freq} is set to \code{FALSE}, probability densities,
component density, are plotted (so that the histogram has
a total area of one). In this case, normal curve will not be
generated.
}
\examples{

## use infert data

data(infert)

# histogram(infert, age)
# histogram(infert, age, labels = FALSE)
# histogram(infert, age, freq = FALSE)

}
\author{
Email: \email{dr.myominnoo@gmail.com}

Website: \url{https://myominnoo.github.io/}
}
