% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iclear.R
\name{iclear}
\alias{iclear}
\title{Clear Global Environment, Plots and Console}
\usage{
iclear()
}
\description{
iclear is a quick way of clearing messy global environment, plots and console without touching
the search path.
}
\examples{
x <- rnorm(100)
y <- rnorm(100)
z <- x + y
hist(x)
hist(y)
hist(z)
iclear()
}
\keyword{all,}
\keyword{clear}
\keyword{console}
\keyword{environment,}
\keyword{global}
\keyword{plots,}
\keyword{remove}
