% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idatesum.R
\name{idatesum}
\alias{idatesum}
\title{Date Summary}
\usage{
idatesum(x, groupby = "y", decreasing = FALSE, rnd = 1,
  plot.display = TRUE, xlab = NULL, show.legend = FALSE)
}
\arguments{
\item{x}{a vector describing the bars which make up the plot. It is usually on x axis.}

\item{groupby}{a character value, indicating patterns of aggregation either by one of the followings:

"y" - Year along

"ym" - Year + Month

"ymd" - Year + Month + Day}
}
\description{
idatesum conceptuzlizes dates as factor variable and tabulate the data. The tabulation can be grouped by combinations of the three date components: Year, Month, Day. It also generates a barplot corresponding to the dates split by Year, Year + Month or a combination of three components.
}
\examples{
# random dates generation
dates <- to.Date(rnorm(100, mean = 4000, sd = 100), "e")
idatesum(dates) # by default, pattern is set to Year alone.
idatesum(dates, "y")
idatesum(dates, "ym")
idatesum(dates, "ym", xlab = "Random Dates")
idatesum(dates, "ym", xlab = "Random Dates", plot.display = FALSE)
idatesum(dates, "ymd") # this can take times.
}
\seealso{
itab, ibarplot, inumsum, isum, ikdplot
}
