% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idetach.R
\name{idetach}
\alias{idetach}
\title{Detach all objects in search pathway}
\usage{
idetach()
}
\description{
idetach removes two types of user-loaded objects in the search pathway:

1) user-loaded packages

2) attached data objects
}
\examples{
library(foreign)
library(MASS)
search()
idetach()
search()
}
\keyword{all,}
\keyword{data,}
\keyword{detach}
\keyword{packages,}
\keyword{search}
