% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifactor.R
\name{ifactor}
\alias{ifactor}
\title{Factoring several variables simultaenously}
\usage{
ifactor(x, data = NULL, lbl.list = NA)
}
\arguments{
\item{x}{An R Object: This must be a numeric vector indicating the poistions
of each variable}

\item{data}{An R Oject to specify dataset. This is not optional.}

\item{lbl.list}{A list Object containing labels as strings. This must correspond
to the levels in each variable.}
}
\description{
The function ifactor() is used to encode several variables into factors into
pre-specified lables and replace the variables from the data frame
}
\examples{
str(infert)
names(infert)
lbl <- list(induced = c("1in", "2in", "3+in"),
            case = c("Yes", "No"),
            spontaneous = c("<1 spon", "2 spon", "3+ spon"))
infert1 <- ifactor(c(4, 5, 6), data = infert)
infert1 <- ifactor(c(4, 5, 6), data = infert, lbl.list = lbl)
str(infert1)
isum(infert1)
}
\seealso{
itab, igroup, isum
}
\keyword{basic}
\keyword{isum,}
\keyword{quick}
\keyword{statistics,}
\keyword{summarize,}
\keyword{summary}
