% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igroup.R
\name{igroup}
\alias{igroup}
\title{A Quick Summary function with Graph Display}
\usage{
igroup(x, y, data = NULL, rnd = 1, na.rm = TRUE, plot.title = NULL)
}
\arguments{
\item{x}{An R Object: This must be numeric}

\item{y}{An R Object: This must be factor.}

\item{data}{An Option to specify dataset.}

\item{rnd}{an integer indicating the number of decimal places:}

\item{na.rm}{A logical value indicating whether "NA" missing values should be
removed before the computation proceeds.}

\item{plot.title}{An optional input to indicate the plottitle: This value passes
through to isum() function along with the title name for the plot.}
}
\description{
igroup() lets you do a quick summary on your data
}
\examples{
str(iris)
igroup(x = "Sepal.Length", y = "Species", data = iris)
igroup(iris$Sepal.Length, iris$Species)
str(infert)
igroup(x = "age", y = "case", data = infert)
igroup(infert$age, infert$case)
}
\seealso{
isum.factor, isum.numeric, isum.data.frame
}
\keyword{basic}
\keyword{isum,}
\keyword{quick}
\keyword{statistics,}
\keyword{summarize,}
\keyword{summary}
