% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ikdplot.R
\name{ikdplot}
\alias{ikdplot}
\title{Kernel Density Plot, Grouped plotting using ggplot2}
\usage{
ikdplot(x, y = NULL, by = NULL, data = NULL, rnd = 1,
  na.rm = FALSE, main = NULL, xlab = NULL, ylab = NULL,
  alpha = 0.1, show.legend = TRUE, legend.text = NULL,
  facet.ncol = 2, plot.save = FALSE, plot.name = "ikdplot.tiff",
  width = 5, height = 4, dpi = 150)
}
\arguments{
\item{x}{a vector describing the bars which make up the plot. It is usually on x axis.}

\item{y}{a vector describing the y axis or second variable in cross-tabulation or data relationship.}

\item{by}{a vector describing the grouping of x. By default, the plot generates a faceted barplot.}

\item{data}{an optional data frame (or object coercible by as.data.frame to a data frame) containing the variables for contigency table.}

\item{rnd}{an integer indicating the number of decimal places:}

\item{na.rm}{A logical value indicating to remove NA values in the table or not. By Default, the value is TRUE}

\item{main}{a main title for the plot.}

\item{xlab}{a label for the x axis, defaults to a description of x.}

\item{ylab}{a label for the y axis, defaults to a description of y.}

\item{alpha}{set transparency of the density}

\item{show.legend}{show or hide the legend. Hide the legend by default}

\item{legend.text}{Legend title}

\item{facet.ncol}{number of columns to be faceted}

\item{plot.save}{a logical value. If TRUE, it saves the plot generated in the current working directory.}

\item{plot.name}{a text for plot filename. Suffix can be ".png", ".tiff" and ".pdf"}

\item{width}{a value in inches}

\item{height}{a value in inches}

\item{dpi}{a value for resolution of the plot saved.}
}
\description{
generate kernel density curves.
}
\examples{
ikdplot(infert$age)
ikdplot(age, data = infert)
ikdplot(infert$age, infert$education)
infert$case <- factor(infert$case)
ikdplot(age, case, data = infert) # NA is included as one of the categories
ikdplot(age, case, data = infert, na.rm = TRUE)
ikdplot(age, pooled.stratum, data = infert)
ikdplot(age, pooled.stratum, education, data = infert)
str(iris)
ikdplot(Sepal.Length, data = iris)
ikdplot(Sepal.Length, Sepal.Width, data = iris)
ikdplot(Sepal.Length, Sepal.Width, Species, data = iris)
}
\seealso{
ibarplot, iboxplot, inumsum, isum, itab
}
\keyword{density}
\keyword{kdensity,}
\keyword{kernel}
\keyword{plot,}
