% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ilog.R
\name{ilog}
\alias{ilog}
\alias{ilog.close}
\title{Echo a copy of console to log file}
\usage{
ilog(logfile = "<unnamed>.txt", append = FALSE)

ilog.close()
}
\arguments{
\item{logfile}{Name of desired log file in \code{.txt} format}

\item{append}{specify if the log file will be appended or not}
}
\description{
\code{ilog} allows to echo a copy of your console to log file.
}
\details{
\code{ilog} is a two-step function that allows you a record of your console. 
A log is a file containing what you type and console output. If a name is not 
specified, then \code{ilog} will use the name \code{<unnamed>.txt}.

\code{ilog} opens a log file and \code{ilog.close} close the file. 

\strong{Warnings:}

Be reminded that clearing the environments while logging distrubs the 
process. If such happens, console prompt will be stuck at \code{log> }. 
Run the \code{commands} below to revert the console prompt back to normal.

options(prompt = \code{"> "}, continue = \code{"+ "})

\strong{Acknowledgement:}
\code{ilog} and \code{ilog.close} are inspriations from \code{STATA} software.
}
\examples{
ilog("myfirstlog.txt")
str(infert)
itab(infert$education)
str(iris)
isum(iris$Sepal.Length)
ilog.close()

## appending the file. 
ilog("myfirstlog.txt", append = TRUE)
summary(infert)
summary(iris)
ilog.close()
}
\seealso{
\code{\link{idetach}}, \code{\link{clear}},
}
\author{
Myo Minn Oo (Email: \email{dr.myominnoo@gmail.com} |
Website: \url{https://myominnoo.github.io/})
}
\keyword{console}
\keyword{console,}
\keyword{log}
\keyword{outputs,}
\keyword{save}
