% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inumsum.R
\name{inumsum}
\alias{inumsum}
\alias{inumsum.numeric}
\title{Number Summary and Plot using ggplot2}
\usage{
inumsum(x, by = NULL, data = NULL, rnd = 1, na.rm = TRUE,
  plot.display = TRUE, main = NULL, xlab = NULL, ylab = NULL,
  legend.text = NULL, facet = TRUE, ...)

inumsum.numeric(x, rnd = 1, na.rm = TRUE)
}
\arguments{
\item{x}{a vector describing the bars which make up the plot. It is usually on x axis.}

\item{by}{a vector describing the grouping of x. By default, the plot generates a faceted barplot.}

\item{data}{an optional data frame (or object coercible by as.data.frame to a data frame) containing the variables for contigency table.}

\item{rnd}{an integer indicating the number of decimal places:}

\item{na.rm}{A logical value indicating to remove NA values in the table or not. By Default, the value is TRUE}

\item{main}{a main title for the plot.}

\item{xlab}{a label for the x axis, defaults to a description of x.}

\item{ylab}{a label for the y axis, defaults to a description of y.}

\item{legend.text}{a text for the title of the plot}

\item{facet}{a logical value. If TRUE, the plot forms a matrix of panels defined by column faceting variables. It is most useful when you have two discrete variables, and all combinations of the variables exist in the data. By default, it is set to TRUE.}

\item{legend.position}{the position of legends ("none", "left", "right", "bottom", "top", or two-element numeric vector)}

\item{save.plot}{a logical value. If TRUE, it saves the plot generated in the current working directory.}

\item{plot.name}{a text for plot filename. Suffix can be ".png", ".tiff" and ".pdf"}

\item{width}{a value in inches}

\item{height}{a value in inches}

\item{dpi}{a value for resolution of the plot saved.}
}
\description{
inumsum (I number-summarize) generates seven-number summary for one variable which include number of observations, missing records, mean, standard deviation, median, interquartile range, minimum, maximum. If the x variable is summarized by a factor variable, then the output will include seven-number summary plus standard error, 95% confidence intervals for means. In addition, the function generates automated Kernel density plot with optional parameters for graphical display.
}
\examples{
inumsum(x = age, by = education, data = infert)
inumsum(age, data = infert)
inumsum(infert$age)
inumsum(infert$age, infert$education)
}
\seealso{
itab, ibarplot, ikdplot
}
\keyword{grouping}
\keyword{number}
\keyword{statistics,}
\keyword{summary}
\keyword{summary,}
