% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irestart.R
\name{irestart}
\alias{irestart}
\title{Restart a new R session without quitting}
\usage{
irestart()
}
\description{
irestart clears all in the search path and restart a new R session without having to quit R. It does
all the following:

1) detach user-loaded packages

2) detach attached data objects

3) clear Global Environment

4) clear Plots

5) clear Console

6) Restart R session
}
\examples{
x <- rnorm(100)
y <- rnorm(100)
z <- x + y
hist(x)
hist(y)
hist(z)
irestart()
}
\keyword{R}
\keyword{Restart,}
\keyword{session}
