% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.Date.R
\name{is.Date}
\alias{is.Date}
\alias{to.Date}
\title{Creates or Coerces Date Vectors}
\usage{
is.Date(x)

to.Date(x, format = "a")
}
\arguments{
\item{x}{object to be coerced or tested.}

\item{format}{a character value, representing the function of Date conversion or coercion.

"a" - "%Y-%m-%d"

"b" - "%Y-%d-%m"

"c" - "%d-%m-%Y"

"d" - "%m-%d-%Y"

"e" - MS excel number format to Date: origin is set to "1899-12-30".}
}
\description{
Creates or coerces objects of type "Date". is.Date is a general test of an object being interpretable as Dates.
}
\examples{
x <- c("2019-01-15", "2019-01-20", "2019-01-21", "2019-01-22")
is.Date(x)
to.Date(x)
y <- to.Date(x)
is.Date(y)
y
# dates in numeric format MS excel
x <- 42705:42710
y <- to.Date(x, "e")
y
}
