% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isave.isum.R
\name{isave.isum}
\alias{isave.isum}
\title{Saving outputs from isum function}
\usage{
isave.isum(x, file, append = FALSE)
}
\arguments{
\item{x}{An R object, containing the outputs of isum function}

\item{file}{A character string naming a file}

\item{append}{logical. Only relevant if file is a character string. If TRUE, the output
is appended to the file. If FALSE, any existing file of the name is destroyed.}
}
\description{
isave.isum is a specific function for the purpose of saving outputs to spreadsheet in
comma-separated value (CSV) format.
}
\examples{
result <- isum(infert)
isave.isum(result, "mysave.csv")
isave.csv(result, "mysave.csv", append = TRUE)


result <- isum(iris)
isave.csv(result, "mysave.csv")
isave.csv(result, "mysave.csv", append = TRUE)
}
\seealso{
ixtab, igroup
}
\keyword{CSV}
\keyword{isum,}
\keyword{outputs,}
\keyword{quick}
\keyword{summarize,}
\keyword{write}
