% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isum.character.R
\name{isum.character}
\alias{isum.character}
\title{A Quick Summary function with Graph Display}
\usage{
\method{isum}{character}(...)
}
\arguments{
\item{...}{Additional parameters}

\item{x}{An R object: it can be a dataframe or a vector.}

\item{l.size}{An integer value determining whether the variable should be treated
as factor}

\item{rnd}{an integer indicating the number of decimal places:}

\item{na.rm}{A logical value indicating whether "NA" missing values should be
removed before the computation proceeds.}
}
\description{
isum() lets you do a quick summary on your data

For the class "factor", it tabulates and gives frequency, percentages and
    cummulative percentages. The same is with the class "character".

For the class "integer", it provides seven summary measures:
    mean, standard deviation, median, Q1, Q3, minimum and maximum.

For the class "data.frame", it gives a list of all the variables based on
    their data types.
}
\examples{
isum(iris)
}
\seealso{
isum.factor, isum.numeric, isum.data.frame
}
\keyword{basic}
\keyword{isum,}
\keyword{quick}
\keyword{statistics,}
\keyword{summarize,}
\keyword{summary}
