% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isum.factor.R
\name{isum.factor}
\alias{isum.factor}
\title{A Quick Summary function with Graph Display}
\usage{
\method{isum}{factor}(x, rnd = 1, na.rm = TRUE)
}
\arguments{
\item{x}{An R object: it can be a dataframe or a vector.}

\item{rnd}{an integer indicating the number of decimal places:}

\item{na.rm}{A logical value indicating whether "NA" missing values should be
removed before the computation proceeds.}

\item{...}{Additional parameters}

\item{l.size}{An integer value determining whether the variable should be treated
as factor}
}
\description{
isum() lets you do a quick summary on your data type.

For "factor" data type, it tabulates and gives frequency, percentages and
    cummulative percentages.
For "integer" data type, it provides seven summary measures:
    mean, standard deviation, median, Q1, Q3, minimum and maximum.
}
\examples{
isum()
}
\seealso{
isum.factor, isum.numeric, isum.data.frame
}
\keyword{basic}
\keyword{isum,}
\keyword{quick}
\keyword{statistics,}
\keyword{summarize,}
\keyword{summary}
