% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isum.R
\name{isum}
\alias{isum}
\title{A Quick Summary function}
\usage{
isum(x, rnd = 1, na.rm = TRUE, l.size = 10L)
}
\arguments{
\item{x}{An R object: it can be a dataframe or a vector.}

\item{rnd}{an integer indicating the number of decimal places:}

\item{na.rm}{A logical value indicating whether "NA" missing values should be removed before the computation proceeds.}

\item{l.size}{An integer value determining how many levels of
unqiue values a vector must have in order to display as factor.}
}
\description{
isum() lets you do a quick summary on your data.
}
\examples{
isum()
}
\keyword{basic}
\keyword{isum,}
\keyword{quick}
\keyword{statistics,}
\keyword{summarize,}
\keyword{summary}
