% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itab.R
\name{itab}
\alias{itab}
\title{Tabulation, Cross Tabulation and Graphical Display}
\usage{
itab(x, by = NULL, data = NULL, rnd = 1, na.rm = FALSE,
  plot.display = TRUE, main = NULL, xlab = NULL, ylab = NULL,
  legend.text = NULL, facet = TRUE, ...)
}
\arguments{
\item{x}{a vector describing the bars which make up the plot. It is usually on x axis.}

\item{by}{a vector describing the grouping of x. By default, the plot generates a faceted barplot.}

\item{data}{an optional data frame (or object coercible by as.data.frame to a data frame) containing the variables for contigency table.}

\item{rnd}{an integer indicating the number of decimal places:}

\item{na.rm}{A logical value indicating to remove NA values in the table or not. By Default, the value is TRUE}

\item{main}{a main title for the plot.}

\item{xlab}{a label for the x axis, defaults to a description of x.}

\item{ylab}{a label for the y axis, defaults to a description of y.}

\item{legend.text}{a text for the title of the plot}

\item{facet}{a logical value. If TRUE, the plot forms a matrix of panels defined by column faceting variables. It is most useful when you have two discrete variables, and all combinations of the variables exist in the data. By default, it is set to TRUE.}

\item{stack}{indicate for stack bar chart in bivariate analysis}

\item{stack.pct}{indicate stacked percent bar chart}

\item{stack.vjust}{number to place value label in stacked bar chart}

\item{facet.ncol}{number of columns to be faceted}

\item{facet.strip.pos}{indicate position of facet strip: "top" by default: others include "bottom", "left", "right".}

\item{legend.show}{show or hide the legend. Hide the legend by default}

\item{legend.text}{Legend title}

\item{legend.position}{the position of legends ("none", "left", "right", "bottom", "top", or two-element numeric vector)}

\item{horizontal}{logical value, indicating the direction of bar: horizontal by default and vertical.}

\item{flip.levels}{if there are more than k categories, bar chart will be automatically flip to horizontal direction in order to generate a better visual. This is the same as setting horizontal to TRUE.}

\item{bar.color}{a text to depict the outer line of bars. You can set "black", "green", "blue" and so on.}

\item{dvsize}{font size of value label in stacked bar chart.}

\item{vjust}{a number to place the value label in percentage to the bar.}

\item{hjust}{a number to place the value label in percentage to the bar.}

\item{position.dodge}{shrink width of bars.}

\item{save.plot}{a logical value. If TRUE, it saves the plot generated in the current working directory.}

\item{plot.name}{a text for plot filename. Suffix can be ".png", ".tiff" and ".pdf"}

\item{width}{a value in inches}

\item{height}{a value in inches}

\item{dpi}{a value for resolution of the plot saved.}
}
\description{
itab is a simple function for data exploration and used to build frequency table of a factor variable or contigency table of the counts at each combination of factor levels.
}
\examples{
str(infert)
data(infert)
# tabulation of one factor variable
itab(education, data = infert)
itab(parity, data = infert)
itab(case, data = infert)
itab(spontaneous, data = infert)
# cross-tabulation
itab(case, education, infert)
itab(case, education, infert, facet = F)
itab(case, education, infert, facet = F, legend.show = TRUE)
itab(case, parity, infert)
itab(case, parity, infert, legend.show = TRUE)
itab(case, parity, infert, facet = FALSE)
itab(case, parity, infert, facet = FALSE, legend.show = TRUE)
itab(case, parity, infert, plot.display = FALSE)
}
\seealso{
ibarplot, inumsum, ikdplot
}
\keyword{contigency}
\keyword{cross-tabulation,}
\keyword{frequency}
\keyword{table}
\keyword{table,}
\keyword{tabulation,}
