% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ixtab.R
\name{ixtab}
\alias{ixtab}
\title{Cross-Tabulation with Graph Display}
\usage{
ixtab(x, y, data = NULL, rnd = 1, na.rm = TRUE,
  plot.display = TRUE, plot.title = NULL)
}
\arguments{
\item{x}{An R object, categorical data which represents Row variable}

\item{y}{An R object, categorical data which represents Column variable}

\item{data}{an optional data frame (or object coercible by as.data.frame to a data frame)
containing the variables for contigency table.}

\item{rnd}{an integer indicating the number of decimal places:}

\item{na.rm}{A logical value indicating to remove NA values in the table or not.
By Default, the value is TRUE}

\item{plot.display}{A logical value, indicating whether the data will be plotted and display}

\item{plot.title}{A character value, specifying the name of the plot in categorical or continuous data. This argument does not work in summaring data.frame object, where the plot will be automatically mapped to the names of the data.frame.}
}
\description{
ixtab is a function that produces three contingency tables: one with counts only, another one with
counts and percentages for rows and the last one with counts and percentages for columns. In
addition, it automatically calculates Chi Square test and Fisher Exact test as well as displays
mosaic plot.

Plot can be disabled by setting plot.display to FALSE.
}
\examples{
str(infert)
ixtab(induced, case, infert)
ixtab(education, case, infert)
ixtab(spontaneous, case, infert)
ixtab(spontaneous, case, infert, plot.display = FALSE) # disable plotting
}
\seealso{
isum, igroup
}
\keyword{categorical}
\keyword{contingency}
\keyword{cross-tabulation,}
\keyword{summary}
\keyword{table,}
