% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kdplot.R
\name{kdplot}
\alias{kdplot}
\title{Delightful Kernel Density Plots}
\usage{
kdplot(x, y = NULL, by = NULL, data = NULL, na.rm = TRUE,
  main = NULL, xlab = NULL, ylab = NULL, legend.text = NULL,
  alpha = 0.1, facet.ncol = 2, plot.save = FALSE,
  plot.name = "<unnamed>.tiff", width = 5, height = 4, dpi = 150)
}
\arguments{
\item{x}{a factor object}

\item{y}{a factor object; if ignored, frequency distribution on x is generated.}

\item{by}{a factor object; if ignored, cross-tabulation on x and y is generated.}

\item{data}{an optional data frame}

\item{na.rm}{A logical value to specify missing values, <NA>:}

\item{main}{title of the plot}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{legend.text}{title of legend}

\item{alpha}{set transparency of the density}

\item{facet.ncol}{number of columns to be faceted}

\item{plot.save}{save the plot}

\item{plot.name}{specify for name of the plot file to be saved.
Filename can be in either of these extension formats: ".tiff", ".png", and ".pdf"}

\item{width}{integer measured in inches}

\item{height}{integer measured in inches}

\item{dpi}{specify resolution of the plot to be saved}
}
\description{
\code{kdplot} generates optimized Kernel Density plots
for univariate, bivariate and stratified analyses.
}
\details{
Exploring data before jumping into complex analysis is always a necessity.
The first step of an analysis should be to summarize and display data.
By doing so, invaluable insight can be gained through the familiarity with the data.
}
\examples{
data(infert)
str(infert)

## univariate analysis
kdplot(infert$age)
kdplot(age, data = infert, main = "MY Plot: Age", xlab = "Age in years",
       ylab = "Density Value")

## bivariate analysis
kdplot(infert$age, infert$education)
kdplot(age, education, data = infert, main = "My Plot", xlab = "Age",
       ylab = "Density Value", legend.text = "Education")

kdplot(age, pooled.stratum, data = infert)
kdplot(age, pooled.stratum, data = infert, main = "My Plot", xlab = "Age",
       ylab = "Pooled Stratum")

## stratified analysis
kdplot(age, pooled.stratum, education, infert)
kdplot(age, pooled.stratum, education, infert, main = "My Plot", xlab = "Age",
       ylab = "Pooled Stratum", legend.text = "Education")

infert$case <- factor(infert$case)

## iris data set
data(iris)
str(iris)
kdplot(age, education, case, data = infert)
kdplot(Sepal.Length, data = iris)
kdplot(Sepal.Length, Sepal.Width, data = iris)
kdplot(Sepal.Length, Sepal.Width, Species, data = iris)
}
\seealso{
\code{\link{isum}}, \code{\link{iboxplot}}, \code{\link{itab}}
}
\author{
Myo Minn Oo (Email: \email{dr.myominnoo@gmail.com} |
Website: \url{https://myominnoo.github.io/})
}
\keyword{density}
\keyword{kdensity,}
\keyword{kernel}
\keyword{plot,}
