% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep.R
\name{keep}
\alias{keep}
\alias{lose}
\alias{pick}
\alias{arrange}
\title{Keep or lose variables and pick rows}
\usage{
keep(data, ...)

lose(data, ...)

pick(data, ...)

arrange(data, ...)
}
\arguments{
\item{data}{Dataset}

\item{...}{variables}
}
\description{
\code{keep} or \code{lose} variables within dataframe

\code{pick} observations that satisfy specified condition
}
\details{
\code{keep} remove all unspecified variables in the dataset.
Variables are also rearranged based on the order they are mentioned.

\preformatted{keep(data, var1, var2, var3, etc)}

\code{lose} works the same as \code{keep}, except that
the variables mentioned are removed from the dataset.

\preformatted{lose(data, va5, var6, var7, etc)}

\code{pick} selects the rows that meets the specified conditions.
This is the same as \strong{filter} function in spreadsheet or
subset in R.

\code{data.frame} or \code{filter} can be used with the same
epxressions.

\preformatted{pick(data, var1 == var2 | var3 > var4)}

\code{arrange} orders the observations of current dataset
into ascending order based on the names of the variables mentioned.
There is no limit to the number of variables that can be specified
in the arguments.

It will be in decreasing order if specified negative sign or "-"
in front of the desired variable(s) without spacing.

\preformatted{arrange(data, var1, var2, -var3)}
}
\note{
\code{drop} and \code{keep} are not reversible.  Once you have
eliminated variables and assigned to the same name,
deleted variables are gone. You need to go back and read the original
dataset again.

Hence, the best strategy is to assign to a different name rather than
putting it back to the original dataset.
}
\examples{
\dontrun{
## keeping variables
keep(infert, education, age, parity)
keep(infert, parity, age:spontaneous, stratum) # variables repositioned

## losing or removing variables
lose(infert, age, parity)
lose(infert, age:case, education)

## pick observations
pick(infert, age > 30)
pick(infert, parity == 6)
pick(infert, age > 30 & age < 35)
pick(infert, case == 0, age > 30 & age < 35, parity == 6)

# use vector instead of dataframe
age <- infert$age
pick(age, age > 40)
pick(infert, age > 40)

## arrange observations
arrange(infert, age)
arrange(infert, -age)
arrange(infert, age, induced)
arrange(infert, age, -induced)
arrange(infert, age, -induced, case)
arrange(infert, age, -induced, -case)

arrange(infert, education)
arrange(infert, education, -age)
}
}
\seealso{
\code{\link{generate}}, \code{\link{egen}}, \code{\link{recode}}
}
\author{
For any feedback, please contact \code{Myo Minn Oo} via:

Email: \email{dr.myominnoo@gmail.com}

Website: \url{https://myominnoo.github.io/}
}
\keyword{arrange,}
\keyword{delete,}
\keyword{drop,}
\keyword{filter,}
\keyword{keep,}
\keyword{pick,}
\keyword{select}
\keyword{subset,}
