% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.R
\name{label}
\alias{label}
\title{Label levels of a categorical variable (Value Label)}
\usage{
label(x, labels)
}
\arguments{
\item{x}{A vector}

\item{labels}{specify a character vector that correspond to the levels}
}
\description{
\code{label} function is handy to quickly change levels of either a factor, number,
logical or character vector.
}
\details{
This function allows to quickly change vector into factor with defined levels or even
levels of a factor into new ones.

\strong{Numeric or factor}
For numeric or factor vector, levels takes the order from the output of
\code{\link{table}} function.

\strong{Character}
For character, it takes alphabetical order and the levels have to be specified accordingly.

\strong{Logical}
TRUE means 1 and FALSE 2 in a logical vector. Hence, levels shoud be corresponding to the
numeric representation of the vector.
}
\examples{
str(infert)

## Example 1: Numeric vector
table(infert$case)
case <- label(infert$case, c("No", "Yes"))
table(case)

## Example 2: factor
table(infert$education)
edu <- label(infert$education, c("low", "middle", "high"))
table(edu)

## Example 3: character
sex <- rep(c("Male", "Female", NA), c(30, 60, 10))
sex; table(sex)
sex1 <- label(sex, c('f', 'm'))
sex1; table(sex1)
label(sex, c('f', 'm', 'missing'))

## Example 4: Logical
s <- sex == "Male"
s; table(s)
s1 <- label(s, c("male", "female"))
s1; table(s1)
label(s, c('male', 'female', 'missing'))
}
\seealso{
\code{\link{gen}}
}
\author{
Myo Minn Oo (Email: \email{dr.myominnoo@gmail.com} |
Website: \url{https://myominnoo.github.io/})
}
\keyword{label}
\keyword{label,}
\keyword{levels}
\keyword{value}
