% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelValue.R
\name{labelValue}
\alias{labelValue}
\alias{labelValue.default}
\alias{labelValue.list}
\alias{labelValue.data.frame}
\title{Manipulate labels}
\usage{
labelValue(x, data = NULL, lbl = NULL, useNA = FALSE)

\method{labelValue}{default}(x, data = NULL, lbl = NULL,
  useNA = FALSE)

\method{labelValue}{list}(x, data = NULL, lbl = NULL, useNA = FALSE)

\method{labelValue}{data.frame}(x, data = NULL, lbl = NULL,
  useNA = FALSE)
}
\arguments{
\item{x}{a variable or a list or a dataframe}

\item{data}{a dataframe object (Optional)}

\item{lbl}{specify a string}

\item{useNA}{force to use NA as level, if any}
}
\description{
\code{labelValue} attach value labels to variables
}
\details{
The order of the labels appearing is the same as output from
\code{\link{tab}} function.

If data is specified, it returns the whole dataframe with recoded variables.
}
\examples{
\dontrun{
str(infert)
# one variable at a time
case <- labelValue(case, infert)
str(case)
case <- labelValue(case, infert, c("no", "yes"))
str(case)
case <- labelValue(infert$case)
str(case)

multiple variables
infert1 <- labelValue(list(case, induced), infert,
             list(c("no", "yes"), c("none", "1", "2")))
str(infert1)

# for data frame, try labelData()
labelValue(infert)
}
}
\seealso{
\code{\link{labelVars}}, \code{\link{labelData}}
}
\author{
Myo Minn Oo (Email: \email{dr.myominnoo@gmail.com} |
Website: \url{https://myominnoo.github.io/})
}
\keyword{label}
\keyword{label,}
\keyword{levels,}
\keyword{value}
