% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelValues.R
\name{labelValues}
\alias{labelValues}
\alias{labelValues.default}
\alias{labelValues.list}
\title{Manipulate labels of a variable or variables}
\usage{
labelValues(data = NULL, var, lbl, useNA = FALSE)

\method{labelValues}{default}(data = NULL, var, lbl, useNA = FALSE)

\method{labelValues}{list}(data = NULL, var, lbl, useNA = FALSE)
}
\arguments{
\item{data}{dataset (optional)}

\item{var}{vectors or lists}

\item{lbl}{specify a string}

\item{useNA}{force to use NA as level, if any}
}
\value{
\code{data.frame} if \code{data} is specified.

\code{vector} if \code{data} is not specified.
}
\description{
\code{labelValues()} attach value labels to variables and
convert to factor
}
\details{
\code{labelValues()} basically recodes the values of variables
as factor. Hence, factored variables can also be used.

\strong{Order of labels}

The order of the labels appearing is the same as the output from
\code{\link{tab}} or \code{table} function.

\strong{Multiple variables and labels}

Multiple variables and labels must be specified as \code{list()}.

\preformatted{
labelValues(data,
            list(var1, var2, var3),
            list(c(lbl1, lbl2, lbl3),
                 c(lbl1, lbl2, lbl3),
                 c(lbl1, lbl2, lbl3)))
}
}
\examples{
\dontrun{
# piping function from magrittr package
library(magrittr)

# one variable
infert \%>\% tab(case)
infert \%>\%
    labelValues(case, c("no", "yes")) \%>\%
    tab(case)

# multiple variables
infert \%>\% tab(induced)
labelValues(infert,
            list(case, induced),
            list(c("no", "yes"), c("none", "one", "two"))) \%>\%
tab(case, by = induced)




## labelling factored variables
infert \%>\% tab(education)
infert \%>\%
    labelValues(education, c("low", "mid", "high")) \%>\%
    tab(education)

infert \%>\% tab(education, by = induced)
infert \%>\%
    labelValues(list(education, induced),
                list(c("low", "mid", "high"),
                     c("none", "one", "two"))) \%>\%
    tab(education, by = induced)




# vector
case.new <- infert$case
labelValues(NULL, case.new, c("no", "yes"))
}
}
\seealso{
\code{\link{labelVars}}, \code{\link{labelData}}
}
\author{
For any feedback, please contact \code{Myo Minn Oo} via:

Email: \email{dr.myominnoo@gmail.com}

Website: \url{https://myominnoo.github.io/}
}
\keyword{change}
\keyword{label}
\keyword{label,}
\keyword{levels,}
\keyword{value}
