% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelVar.R
\name{labelVar}
\alias{labelVar}
\alias{labelData}
\title{ADD, REPLACE or REMOVE labels of variables and dataset}
\usage{
labelVar(data, ...)

labelData(data, lbl = NULL)
}
\arguments{
\item{data}{dataset}

\item{...}{variable = "label"}

\item{lbl}{label for dataset}
}
\description{
\code{labelVar()} labels \code{variables}.

\code{labelData()} labels \code{Dataset}.
}
\details{
Labels are useful to provide more detailed information about
variables or dataset. Many functions in \code{mStats} package
extract label information and display them as footnote.
Single or multiple variables can be labelled.

\strong{Label single variables}

\preformatted{labelVar(data, var = "lbl")}

\strong{Label multiple variables}

\preformatted{labelVar(data,
      var1 = "lbl1", var2 = "lbl2", var3 = "lbl3", etc)}

\strong{Label Dataset}

Dataset can also be labelled and displayed in \code{codebook}.

\preformatted{labelData(data, lbl)}
}
\examples{

## use infert data
data(infert)
codebook(infert)


## add labels
infert.new <- labelVar(infert,
                       education = "EDUCATION",
                       age = "AGE IN YEARS",
                       parity = "count",
                       case = "case status",
                       induced = "# of prior induced abortions",
                       spontaneous = "# of prior spon. abortions",
                       stratum = "Matched set number",
                       pooled.stratum = "Stratum number")

infert.new <- labelData(infert.new,
    "Infertility after Spontaneous and Induced Abortion")
codebook(infert.new)


}
\author{
Email: \email{dr.myominnoo@gmail.com}

Website: \url{https://myominnoo.github.io/}
}
