% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelVars.R
\name{labelVars}
\alias{labelVars}
\alias{labelVars.default}
\alias{labelVars.character}
\alias{labelVars.list}
\alias{labelData}
\title{Manipulate labels: Label variables or dataframe}
\usage{
labelVars(data = NULL, vars, lbl)

\method{labelVars}{default}(data = NULL, vars, lbl)

\method{labelVars}{character}(data = NULL, vars, lbl)

\method{labelVars}{list}(data = NULL, vars, lbl)

labelData(data, lbl)
}
\arguments{
\item{data}{dataset}

\item{vars}{one variable or variables}

\item{lbl}{label or labels}
}
\description{
\code{labelVars} names \code{variables}.

\code{labelData} names \code{Dataset}.
}
\details{
Labels are useful to provide more detailed information about
variables or dataset. Many functions in \code{mStats} package
extract label information and display them as footnote.

\strong{single or multiple variables}

Single or multiple variables can be labelled.

Example: single variable \preformatted{labelVars(data, var, lble)}

Example: multiple variable

\preformatted{labelVars(data,
      c(var1, var2, var3),
      c(lbl1, lbl2, lbl3))}

\strong{Labelling Dataset}

Dataset can also be labelled and displayed in \code{codebook}.

\preformatted{labelData(data, lbl)}
}
\examples{
\dontrun{
# piping function from magrittr package
library(magrittr)

# label variables
infert \%>\%
    labelVars(age, "AGE OF PATIENT") \%>\%
    codebook()

infert \%>\%
    labelVars(c(age, parity, induced, case),
              c("Patient's AGE", "No of pregancy",
                "Induced Abortion", "case: Yes or no")) \%>\%
    codebook()

# label dataframe
infert \%>\%
    labelData("Infertility after Spontaneous and Induced Abortion") \%>\%
    codebook()
}
}
\seealso{
\code{\link{codebook}}, \code{\link{labelData}},

\code{\link{labelValues}}
}
\author{
For any feedback, please contact \code{Myo Minn Oo} via:

Email: \email{dr.myominnoo@gmail.com}

Website: \url{https://myominnoo.github.io/}
}
\keyword{data}
\keyword{data.frame}
\keyword{frame,}
\keyword{label}
\keyword{label,}
