% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listView.R
\name{listView}
\alias{listView}
\alias{listView.default}
\alias{listView.character}
\alias{listView.list}
\alias{listView.data.frame}
\title{List and view the observations of variables from a dataset}
\usage{
listView(data, vars = NULL, head = TRUE, nrow = 6)

\method{listView}{default}(...)

\method{listView}{character}(data, vars = NULL, head = TRUE, nrow = 6)

\method{listView}{list}(data, vars = NULL, head = TRUE, nrow = 6)

\method{listView}{data.frame}(data, vars = NULL, head = TRUE, nrow = 6)
}
\arguments{
\item{data}{a data frame object}

\item{vars}{either one variable or a vector of variables or two
variables separated by colon (x:y)}

\item{head}{A logical value to specify heading or tailing values}

\item{nrow}{specify the number of observations}

\item{...}{optional arguments}
}
\description{
\code{listView} display a list of observations of variables from
a dataset. If \code{vars} is not specified, the observations of
all the variables will be displayed.
}
\examples{
\dontrun{
listView(infert, education)
listView(infert, education:stratum)
listView(infert, case:stratum)
listView(infert, case:stratum, head = FALSE)
listView(infert, c(education, case, parity))
listView(infert)
}
}
\seealso{
\code{\link{codebook}}
}
\author{
Myo Minn Oo (Email: \email{dr.myominnoo@gmail.com} |
Website: \url{https://myominnoo.github.io/})
}
\keyword{display,}
\keyword{head,}
\keyword{list,}
\keyword{tail}
\keyword{view,}
