% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.output.R
\name{logistic.output}
\alias{logistic.output}
\title{Process Logistic Regression model and Display Odds Ratios}
\usage{
logistic.output(covar, outcome, data, aOR = FALSE, write.csv = FALSE,
  file = NULL, print.to.console = FALSE)
}
\arguments{
\item{covar}{a vector containing all the covariates to be fitted}

\item{outcome}{a symbolic description of the dependent variable to be fitted.}

\item{data}{a data frame containing the variables in the model.}

\item{aOR}{a logical value indicating either unadjusted or adjusted model}

\item{write.csv}{a logical value to save the cox table output to CSV file.}

\item{file}{either a character string naming a file.}

\item{print.to.console}{print the table to the console even if the output is assigned to an object.}
}
\description{
Takes several covariates and generates Odds Ratios outputs at the same time. Outputs include unadjusted / adjusted OR, 95% CI and P-values.
}
\seealso{
cox.output, itab, isum, inumsum
}
