% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhodds.R
\name{mhodds}
\alias{mhodds}
\alias{mhodds.default}
\alias{mhodds.bi}
\alias{mhodds.mh}
\alias{calc_OR}
\alias{calc_RawMHOR}
\alias{calc_mhor}
\alias{splitTables}
\title{Calculating Odds Ratios}
\usage{
mhodds(
  data = NULL,
  var_exp,
  var_case,
  strata = NULL,
  ref_exp = NULL,
  ref_case = NULL,
  na.rm = FALSE,
  rnd = 3,
  plot = TRUE,
  print.table = TRUE
)

\method{mhodds}{default}(
  data = NULL,
  var_exp,
  var_case,
  strata = NULL,
  ref_exp = NULL,
  ref_case = NULL,
  na.rm = FALSE,
  rnd = 3,
  plot = TRUE,
  print.table = TRUE
)

\method{mhodds}{bi}(
  data = NULL,
  var_exp,
  var_case,
  strata = NULL,
  ref_exp = NULL,
  ref_case = NULL,
  na.rm = FALSE,
  rnd = 3,
  plot = TRUE,
  print.table = TRUE
)

\method{mhodds}{mh}(
  data = NULL,
  var_exp,
  var_case,
  strata,
  ref_exp = NULL,
  ref_case = NULL,
  na.rm = FALSE,
  rnd = 3,
  plot = TRUE,
  print.table = TRUE
)

calc_OR(t, ref_lvl, rnd)

calc_RawMHOR(t, ref_exp, strata_lvl)

calc_mhor(
  data,
  strata,
  strata_lvl,
  exp_name,
  case_name,
  ref_exp,
  ref_case,
  na.rm,
  rnd
)

splitTables(t, ref_exp = NULL)
}
\arguments{
\item{data}{a data frame object (Optional)}

\item{var_exp}{Exposure variable.}

\item{var_case}{Case or outcome variable should be binary vector.}

\item{strata}{variable for stratified (Mantel-Haenszel Method)
analysis}

\item{ref_exp}{Exposed level}

\item{ref_case}{outcome}

\item{na.rm}{A logical value to specify missing values, <NA> in the table}

\item{rnd}{specify rounding of numbers. See \code{\link{round}}.}

\item{plot}{logical value to display plots of rates across a categorical
variable}

\item{print.table}{logical value to display formatted outputs}

\item{t}{2x2 table input to calaculate odds ratio of non-reference}

\item{ref_lvl}{logical vector to indicate where reference category is}

\item{strata_lvl}{levels in strata variable}

\item{exp_name}{Name of \code{Exposure} Variable
Exposed and non-exposed}

\item{case_name}{Name of \code{Case} variable}
}
\description{
\code{mhodds()} generates cross-tabulation between two variables and
display odds as well as Odds Ratios of failure \code{var_case}
against a categorical explanatory variable \code{var_exp}.

It is used with cross-sectional as well as case-control data.
}
\details{
The variable \code{var_case} should coded 1 for case and 0 for non-case.
The arugment \code{exposed} can be used to set the reference value for
comparison of odds ratios.

It produces a simple table with odds, Odds Ratio,95% CI as well as
p-value. If \code{strata} is specified, \code{Mantel-Haenzsel\'s Odds
Ratio} is generated along with chi-sqaured test for heterogeneity.

\strong{Odds Ratio, OR}

\deqn{OR = (D1 x H0) / (D0 x H1)}

\strong{Error Factor, EF using Woolf's formula}

\deqn{95\% CI = OR / EF or OR x EF}

\deqn{EF = exp(1.96 x SE(log(OR)))}

\deqn{SE(log(OR)) = \sqrt{1/D1 + 1/H1 + 1/D0 + 1/H0}}

\strong{Mantel-Haenszel's OR}

\deqn{ORMH = Q / R}

\deqn{Q = \sum{(D1i x H0i) / ni}}

\deqn{R = \sum{(D0i x H1i) / ni}}

\strong{95\% CI of ORMH}

\deqn{95\% CI = OR / EF or OR x EF}

\deqn{SE(ORMH) = \sqrt{V / (Q x R)}}

\deqn{V = \sum{(Di x Hi x n0i x n1i) / ( (ni)^2 x (ni - 1))}}
}
\examples{
\dontrun{
### Simpson's Paradox: Example from Burt Gerstman's Epidemiology
### Chapter 14, Page 326

library(magrittr)
simpson <- expandTables(
    "1" = c(1000, 9000, 50, 950),
    "2" = c(95, 5, 5000, 5000),
    exp_name = "treatment",
    exp_lvl = c("new", "standard"),
    case_name = "outcome",
    case_lvl = c("alive", "dead"),
    strata_name = "clinic") \%>\%
    labelVars(c(treatment, outcome, clinic),
              c("Treatment: new or standard",
                "Outcome: alive or dead", "clinic: 1 or 2")) \%>\%
    labelData("Example of Simpson's Paradox")

## checking structure
codebook(simpson)

## Calculating Odds ratio
mhodds(simpson, treatment, outcome)

## Mantel-Haenzsel's Odds Ratio
mhodds(simpson, treatment, outcome, clinic)

## Checking OR for Clinic 1
simpson \%>\%
    pick(clinic == 1) \%>\%
    mhodds(treatment, outcome)

## Checking OR for Clinic  2
simpson \%>\%
    pick(clinic == 2) \%>\%
    mhodds(treatment, outcome)




### Example from Essential Medical Statistics
# Page 178, Chapter 18: Controlling for confounding: Stratification
library(magrittr)
lepto <- expandTables(
    male = c(36, 14, 50, 50), female = c(24, 126, 10, 90),
    exp_name = "area", exp_lvl = c("Rural", "Urban"),
    case_name = "ab", case_lvl = c("Yes", "No"),
    strata_name = "gender"
) \%>\%
    labelData("Prevalence survey of leptospirosis in West Indies") \%>\%
    labelVars(c(area, ab, gender),
              c("Type of area", "Leptospirosis Antibodies",
                "Gender: Male or female"))

codebook(lepto)

## OR
mhodds(lepto, area, ab, ref_case = "Yes")

## MH OR
mhodds(lepto, area, ab, gender, ref_case = "Yes")




## Asthma Example from Essential Medical Statistics
## page 160

library(magrittr)
asthma <- expandTables(c(81, 995, 57, 867),
              exp_name = "sex",
              exp_lvl = c("Women", "Man"),
              case_name = "asthma",
              case_lvl = c("Yes", "No")) \%>\%
          labelData("Hypothetical Data of Asthma Prevalence") \%>\%
          labelVars(c(sex, asthma),
              c("Man or Woman", "Asthma or No Asthma"))

## Checking codebook
codebook(asthma)

mhodds(asthma, sex, asthma)






## The odds ratio, J Martin Bland, Douglas G Altman, BMJ 2000;320:1468
hay <- expandTables(c(141, 420, 928, 13525),
                    exp_name = "eczema",
                    exp_lvl = c("Yes", "No"),
                    case_name = "hayFever",
                    case_lvl = c("Yes", "No")) \%>\%
       labelData("hay fever and eczema in 11 year old children") \%>\%
       labelVars(c(eczema, hayFever),
                c("prevalence of eczema", "prevalence of hay fever"))

codebook(hay)

mhodds(hay, eczema, hayFever)
}
}
\references{
\enumerate{
    \item Essential Medical Statistics, Betty R. Kirkwood & Jonathan A.C.
Sterne, Second Edition. Chapter 3

    \item B. Burt Gerstman. Epidemology Kept Simple. Wiley-Blackwell, 2013:
Chapter 13, Confidence Intervals and p-values
}
}
\seealso{
\code{\link{stmh}}, \code{\link{tabOdds}}, \code{\link{tabRisks}},

\code{\link{expandTables}}
}
\author{
For any feedback, please contact \code{Myo Minn Oo} via:

Email: \email{dr.myominnoo@gmail.com}

Website: \url{https://myominnoo.github.io/}
}
\keyword{MHOR,}
\keyword{Mantel-Haenzsel's}
\keyword{Odds}
\keyword{Ratio}
\keyword{descriptive,}
\keyword{frequency}
\keyword{odds}
\keyword{odds,}
\keyword{ratio,}
\keyword{statistics,}
\keyword{table,}
