% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.output.R
\name{model.output}
\alias{model.output}
\title{Reporting coefficients of regression models}
\usage{
model.output(model, rnd = 1, raw = FALSE)
}
\arguments{
\item{model}{object glm or lm type}

\item{rnd}{rounding digits}

\item{raw}{logical value}
}
\description{
\code{model.output} display corresponding estimates of different models along
with 95% confidence intervals and p-values.
}
\details{
If raw is set to TRUE, a table containing raw estimate, standard errors, z value and
p-value is generated.
}
\examples{
\dontrun{
## example from IRDE website:
## https://stats.idre.ucla.edu/r/dae/logit-regression/
mydata <- read.csv("https://stats.idre.ucla.edu/stat/data/binary.csv")
codebook(mydata)
tab(admit, mydata)
tab(rank, mydata)

mylogit <- glm(admit ~ gre + gpa + factor(rank), data = mydata, family = "binomial")
summary(mylogit)

model.output(mylogit)
model.output(mylogit, raw = TRUE)
model.fit(mylogit)
}
}
\seealso{
\code{\link{model.fit}}
}
\author{
Myo Minn Oo (Email: \email{dr.myominnoo@gmail.com} |
Website: \url{https://myominnoo.github.io/})
}
\keyword{coefficients}
\keyword{display}
\keyword{model}
\keyword{output,}
