% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelDisplay.R
\name{modelDisplay}
\alias{modelDisplay}
\alias{modelDisplay.default}
\alias{modelDisplay.unAdjusted}
\alias{modelDisplay.adjusted}
\title{Report parameters from regression models}
\usage{
modelDisplay(model, rnd = 1)

\method{modelDisplay}{default}(model, rnd = 1)

\method{modelDisplay}{unAdjusted}(model, rnd = 1)

\method{modelDisplay}{adjusted}(model, rnd = 1)
}
\arguments{
\item{model}{object glm or lm type}

\item{rnd}{rounding digits}
}
\description{
\code{modelDisplay} display corresponding estimates of different models along
with 95% confidence intervals and p-values.
}
\details{
\code{modelDisplay}

reports parameters from regression models. Currently supporting model is
logistic models. Other types will be incoporated in future works.

In unadjusted analysis, it reports unajdusted odds ratios, 95% Confidence
interval and Wald's p-value.

In adjusted analysis, both unadjusted and adjusted parameters are reported.

The reports are in well-formatted texts which can be readily copied into
spreadsheet programs or format in word file.

\strong{Please be reminded:}

This is an ongoing work. In case of comments or suggestions, please contact me
at \email{dr.myominnoo@gmail.com}.

\strong{Reference: }
\enumerate{
  \item LOGIT REGRESSION | R DATA ANALYSIS EXAMPLES. UCLA: Statistical Consulting
  Group. from https://stats.idre.ucla.edu/r/dae/logit-regression/ (accessed
  September 27, 2019)
}
}
\examples{
\dontrun{
## example from IRDE website:
## https://stats.idre.ucla.edu/r/dae/logit-regression/
mydata <- read.csv("https://stats.idre.ucla.edu/stat/data/binary.csv")
codebook(mydata)
tab(admit, mydata)
tab(rank, mydata)

mylogit <- glm(admit ~ gre + gpa + factor(rank), data = mydata, family = "binomial")
summary(mylogit)

modelDisplay(mylogit) # generates parameters
modelFit(mylogit) # test overall significant of the model
}
}
\seealso{
\code{\link{modelFit}}
}
\author{
Myo Minn Oo (Email: \email{dr.myominnoo@gmail.com} |
Website: \url{https://myominnoo.github.io/})
}
\keyword{coefficients}
\keyword{display}
\keyword{model}
\keyword{output,}
