% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatDate.R
\name{month}
\alias{month}
\title{Month from a Date}
\usage{
month(x)
}
\arguments{
\item{x}{Date Object}
}
\description{
\code{month} retrieves month from a Date
}
\examples{
dates <- formatDate(c("2019-01-15", "2019-01-20", "2019-01-21", "2019-01-22"),
                    "Ymd", "-")
month(dates)
}
\seealso{
\code{\link{formatDate}}, \code{\link{year}}, \code{\link{day}}
}
\author{
Myo Minn Oo (Email: \email{dr.myominnoo@gmail.com} |
Website: \url{https://myominnoo.github.io/})
}
\keyword{date}
\keyword{month,}
