% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mStats-package.R
\docType{package}
\name{mStats-package}
\alias{mStats}
\alias{mStats-package}
\title{mStats: Medical Statistics & Epidemiological Analysis}
\description{
A set of tidyverse-friendly functions for data management, calculation of epidemiological measures, statistical analysis, and table creation.
}
\author{
\strong{Maintainer}: Myo Minn Oo \email{dr.myominnoo@gmail.com} (\href{https://orcid.org/0000-0003-4089-016X}{ORCID})

}
\keyword{internal}
