% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mStats_HELP.R
\docType{package}
\name{mStats}
\alias{mStats}
\alias{mStats-package}
\title{mStats: Myo's Package for Stats & Graphs}
\description{
The mStats package provides two categories of important data exploration and summarization:
statistics and graphs.
}
\seealso{
Useful links:
\enumerate{
  \item \href{https://myominnoo.github.io/pages/software.html}{Installing mStats}
  \item \href{https://myominnoo.github.io/pages/R/mStats_help.html}{Getting Started with mStats}
}
}
\author{
\strong{Maintainer}: Myo Minn Oo \email{dr.myominnoo@gmail.com} \url{https://myominnoo.github.io/}
}
