% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_N.R
\name{n_}
\alias{n_}
\alias{N_}
\alias{n_all}
\title{Count from \code{n_} to \code{N_}}
\usage{
n_(...)

N_(...)

n_all()
}
\arguments{
\item{...}{One or more unquoted variables separated by commas. \code{x:y} pattern
cannot be used.}
}
\description{
\Sexpr[results=rd]{lifecycle::badge("stable")}

\code{n_()}, \code{N_()} and \code{n_all()} are  for indexing observations
or generating sequences of numbers.
\itemize{
\item \code{n_()} generates a running counter within a group of variables, and
contains the number of the current observation.
\item \code{N_()} provide the total number within each group of variables.
\item \code{n_()} generates a sequence from 1 to number of observations in the
dataset.
}
}
\section{Examples}{


Here we show an example adopted from IRDE website.\preformatted{## Example adapted from
## https://stats.idre.ucla.edu/stata/seminars/notes/counting-from-_n-to-_n/

ex <- data.frame(rbind(c(72, 1), c(84, 2), c(76, 1), c(89, 3), c(82, 2),
                       c(90, 1), c(85, 1)))
names(ex) <- c("score", "group")

ex <- generate(ex, id_n, n_(group))
ex <- generate(ex, id_N, N_(group))
ex <- generate(ex, id_n_all, n_all())
ex
}

To counter within all variables in the dataset, you can leave the varibles
unspecified. Alternatively, \code{x:y} pattern can be used.\preformatted{iris <- generate(iris, id_n, n_())
iris <- generate(iris, id_N, N_())
iris <- generate(iris, id_n_all, n_all())
head(iris)
}
}

\seealso{
Other data management: 
\code{\link{append}()},
\code{\link{codebook}()},
\code{\link{describe}()},
\code{\link{duplicates}()},
\code{\link{egen}()},
\code{\link{encode}()},
\code{\link{expand2}()},
\code{\link{formatDate}()},
\code{\link{generate}()},
\code{\link{label}()},
\code{\link{recode}()},
\code{\link{replace}()},
\code{\link{rowmiss}()}
}
\concept{data management}
