% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nscore.R
\name{nscore}
\alias{nscore}
\title{Normal score transformation of data.}
\usage{
nscore(df, d, na.rm = TRUE)
}
\arguments{
\item{df}{Data frame.}

\item{d}{Column in data frame to be transformed.}

\item{na.rm}{Boolean, remove NAs from d. Defult is TRUE.}
}
\value{
Named list of data frames with transformed data and transformation
    table.
}
\description{
Takes a vector of values x and calculates their normal scores. Returns a list
with the scores and an ordered table of original values and scores, which is
useful as a back-transform table. See backtr().
}
\author{
Alex M Trueman
Modified from Ashton Shortridge, May/June, 2008.
https://msu.edu/~ashton/research/code/nscore.R
Originally based on the GSLIB code for nscore function.
}
