% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nscore.R
\name{nscore_model}
\alias{nscore_model}
\title{Make normal score transformation model.}
\usage{
nscore_model(x, na.rm = TRUE)
}
\arguments{
\item{x}{Numeric vector to be transformed.}

\item{na.rm}{Boolean, remove NAs from x. Defult is TRUE.}
}
\value{
Dataframe transformation model.
}
\description{
Takes a vector of values x and calculates their normal scores. Returns an
ordered table of original values and scores, which is useful as a
back-transform model. See backtr().
}
\author{
Alex M Trueman
Modified from Ashton Shortridge, May/June, 2008.
https://msu.edu/~ashton/research/code/nscore.R
Originally based on the GSLIB code for nscore function.
}
