% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOR.R
\name{plotOR}
\alias{plotOR}
\title{Odds Ratios Plot}
\usage{
plotOR(odds, cilow, cihigh, vars = NULL, main = NULL, xlab = NULL,
  ylab = NULL, ci.barsize = 0.5, ci.barheight = 0.5,
  ci.barcolor = "black", ci.linetype = "solid", vline.xintercept = 1,
  vline.size = 0.25, vline.type = "dashed", vline.color = "blue",
  or.shape = 15, or.size = odds * 2, or.color = NULL,
  xlimits = c(0, max(cihigh)), text.size = 16, save.plot = FALSE,
  plot.name = "plotOR.tiff", width = 5, height = 4, dpi = 150)
}
\arguments{
\item{odds}{a vector containing Odds Ratios (OR) values}

\item{cilow}{a vector containing lower levels of Odds Ratios (OR) values}

\item{cihigh}{a vector containing upper levels of Odds Ratios (OR) values}

\item{vars}{a vector with names of categories or sub-categories for OR}

\item{main}{a main title for the plot.}

\item{xlab}{a label for the x axis, defaults to a description of x.}

\item{ylab}{a label for the y axis, defaults to a description of y.}

\item{ci.barsize}{a double to change the size of the 95\% CI bars. 0.5 by default.}

\item{ci.barheight}{a double to change the height of the 95\% CI bars. 0.5 by default.}

\item{ci.barcolor}{a character or sequence of numeric vector to specify the color of the bar. "black" by default.}

\item{ci.linetype}{a character vector for specifying line type}

\item{vline.xintercept}{intercept value at x axis for vertical line}

\item{vline.size}{size of vertical line}

\item{vline.type}{type of vertical line}

\item{vline.color}{color of vertical line}

\item{or.shape}{shape of OR point estimate}

\item{or.size}{size of OR estimate: by default, OR x 2 (to magnify the point in display)}

\item{or.color}{color of OR estimate}

\item{xlimits}{specify two points to limit x axis. by default, it is c(0, max(cihigh))}

\item{text.size}{Text size for the axis labels}

\item{save.plot}{a logical value. If TRUE, it saves the plot generated in the current working directory.}

\item{plot.name}{a text for plot filename. Suffix can be ".png", ".tiff" and ".pdf"}

\item{width}{a value in inches}

\item{height}{a value in inches}

\item{dpi}{a value for resolution of the plot saved.}
}
\description{
produce an Odds Ratios Plot with 95% CI values.
}
\examples{
str(infert)
or <- logistic.output("education", case, data = infert)
or
infert$induced <- factor(infert$induced)
infert$spontaneous <- factor(infert$spontaneous)
or <- logistic.output(c("education", "induced", "spontaneous"), case, data = infert)
plotOR(or[,1], or[,2], or[,3])
plotOR(or[,1], or[,2], or[,3], vars = rownames(or))
plotOR(or[,1], or[,2], or[,3], vars = rownames(or), main = "Odds Ratios Plot")
plotOR(or[,1], or[,2], or[,3], vars = rownames(or), xlab = "Adjusted OR")
plotOR(or[,1], or[,2], or[,3], vars = rownames(or), xlimits = c(0, 30))
plotOR(or[,1], or[,2], or[,3], vars = rownames(or), or.color = "black")
plotOR(or[,1], or[,2], or[,3], vars = rownames(or), or.size = 3, or.color = "black")
plotOR(or[,1], or[,2], or[,3], vars = rownames(or), or.size = 3, or.color = "black")
plotOR(or[,1], or[,2], or[,3], vars = rownames(or), or.size = 3, or.color = "black", save.plot = TRUE)
}
\seealso{
iboxplot, itab, inumsum, isum
}
\keyword{OR,}
\keyword{Odds}
\keyword{Plot}
\keyword{Ratios}
