% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{plotRisks}
\alias{plotRisks}
\title{Plot estimates with 95\% Confidence Interval}
\usage{
plotRisks(.p, .ll, .ul, .x.name, .by.name, .ylab = "Risks")
}
\arguments{
\item{.p}{prevalence}

\item{.ll}{lower limit of CI}

\item{.ul}{upper limit of CI}

\item{.x.name}{xlab or x variable}

\item{.by.name}{ylab or y variable}

\item{.ylab}{ylab}
}
\description{
It creates a plot for estimates and their 95\\% CI
}
