% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyramid.R
\name{pyramid}
\alias{pyramid}
\title{Population Pyramid Graph}
\usage{
pyramid(
  data,
  age,
  sex,
  cut = NULL,
  main = NULL,
  label_left = NULL,
  label_right = NULL,
  label_center = NULL,
  center_width = NULL,
  na.rm = FALSE
)
}
\arguments{
\item{data}{Dataset}

\item{age}{age variable or a continuous variable}

\item{sex}{sex variable or a binary variable}

\item{cut}{either a single number or a numeric vector.}

\item{main}{The main title (on top) using font, size (character
    expansion) and color \code{par(c("font.main", "cex.main", "col.main"))}.}

\item{label_left}{title of left plot}

\item{label_right}{title of right plot}

\item{label_center}{title of center plot}

\item{center_width}{width between two plots: Default value is calculated
on the maximum character number of labels in the center.}

\item{na.rm}{A logical value to specify missing values}
}
\description{
\code{pyramid()} draws a population pyramid graph of age and sex. Other similar
types of variables can be used.
}
\examples{

## use infert data

data(infert)
# pyramid(infert, age, case)
# pyramid(infert, age, case, cut = 5)
# pyramid(infert, age, case, cut = 2)
# pyramid(infert, age, case, cut = 2,
#         label_left = "No", label_right = "yes")

}
\author{
Email: \email{dr.myominnoo@gmail.com}

Website: \url{https://myominnoo.github.io/}
}
