% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mStats.R
\name{recast}
\alias{recast}
\title{Converting R data types}
\usage{
recast(data, mode, ...)
}
\arguments{
\item{data}{data.farme}

\item{mode}{allows R data types from \code{\link{typeof}}}

\item{...}{variables}
}
\value{
data.frame
}
\description{
\code{recast()} converts variables into specified data type
\code{\link{typeof}}.
}
\examples{

recast(airquality, "character")
recast(airquality, "character", Ozone, Solar.R, Wind)

}
\author{
Email: \email{dr.myominnoo@gmail.com}

Website: \url{https://myominnoo.github.io/}
}
