% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode}
\alias{recode}
\title{Recode contents of a variable}
\usage{
recode(var, data = NULL, old.value, new.value)
}
\arguments{
\item{var}{name of a new variable}

\item{data}{data frame}

\item{old.value}{vector}

\item{new.value}{vector (length of 1 or same with old.value)}
}
\description{
\code{recode} easily manipulates contents of a new variable within dataframe
}
\details{
\code{recode}
changes the values of variables including categorical variables
according to the rules specified below.

In case of factor, \code{recode} first converts the vector into character,
recodes and then revert back to factor.

If data is specified, it returns the whole dataframe with recoded variables.

Common rules:

Old.value  New.value  Example             Meaning
#          #          0 >>> 1             0 recoded to 1
c(#, #)    c(#, #)    c(1,2) >>> c(3,4)   1 to 3; 2 to 4
c(#, #)    #          c(1,2) >>> 1        1 to 1; 2 to 1
#:#        #          1:5 >>> 1           1 to 1; 2 to 1; ... ; 5 to 1
}
\examples{
str(infert)
table(infert$case)
infert.new <- recode(case, infert, 0, 2)
table(infert.new$case)

infert.new <- recode(case, infert, c(0, 1), c(3, 4))
table(infert.new$case)

table(infert$parity)
infert.new <- recode(parity, infert, c(1,2,3,4), 1)
table(infert.new$parity)

table(infert$age)
infert.new <- recode(age, infert, 21:30, 1)
infert.new <- recode(age, infert.new, 31:40, 2)
infert.new <- recode(age, infert.new, 41:44, 3)
table(infert.new$age)
}
\seealso{
\code{\link{generate}}
}
\author{
Myo Minn Oo (Email: \email{dr.myominnoo@gmail.com} |
Website: \url{https://myominnoo.github.io/})
}
\keyword{change,}
\keyword{generate,}
\keyword{produce}
\keyword{recode,}
