% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ren.R
\name{ren}
\alias{ren}
\title{Clear Workspace, Plots and Console}
\usage{
ren()
}
\arguments{
\item{NA}{No parameter required.}
}
\description{
ren() is a quick way of clearing messy workspace, plots and console.
}
\examples{
ren()
}
\keyword{clear,}
\keyword{console}
\keyword{plots,}
\keyword{workspace,}
