% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace.R
\name{replace}
\alias{replace}
\title{Change contents of an existing variable}
\usage{
replace(data, var, value, condition = NULL)
}
\arguments{
\item{data}{A data.frame}

\item{var}{Name of the new variable}

\item{value}{Content of the new variable}

\item{condition}{Conditions can be expressed. By default, it is set to
no condition \code{NULL}.}
}
\description{
\Sexpr[results=rd]{lifecycle::badge("stable")}

\code{replace()} alters the contents of a variable based on the \code{condition}
you specify. If not specified, all contents of the variable is set to
\code{value}.
}
\section{Examples}{


Here we show a few examples of how to replace values in variables.\preformatted{## replacing mpg with standardized values of mpg
mtcars <- replace(mtcars, mpg, NA, mpg >= 10 & mpg <= 20)
mtcars <- replace(mtcars, mpg, mpg / mean(mpg))
codebook(mtcars)


## replacing education levels with one value
infert <- replace(infert, education, "6+yrs",
        education == "6-11yrs" | education == "12+ yrs")
table(infert$education)
}
}

\seealso{
Other data management: 
\code{\link{append}()},
\code{\link{codebook}()},
\code{\link{describe}()},
\code{\link{duplicates}()},
\code{\link{egen}()},
\code{\link{encode}()},
\code{\link{expand2}()},
\code{\link{formatDate}()},
\code{\link{generate}()},
\code{\link{label}()},
\code{\link{n_}()},
\code{\link{recode}()},
\code{\link{rowmiss}()}
}
\concept{data management}
