% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{rowColOrder}
\alias{rowColOrder}
\title{Change reference level by row or column}
\usage{
rowColOrder(.tbl, .exp.value = NULL, .case.value = NULL)
}
\arguments{
\item{.tbl}{table}

\item{.exp.value}{reference row}

\item{.case.value}{reference column}
}
\description{
\code{tblRowColOrder()} changes the order of levels in rows or
columns by specifying names of levels.
}
\details{
This function is used in calculating risks, odds and associated measures.
}
