% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_functions.R
\name{rowmiss}
\alias{rowmiss}
\title{Functions to manipulate rows}
\usage{
rowmiss(...)
}
\arguments{
\item{...}{One or more unquoted variables separated by commas. \code{x:y} pattern
cannot be used.}
}
\description{
\Sexpr[results=rd]{lifecycle::badge("stable")}

\code{rowmiss()} gives the number of \code{missing} values in specified variables
for each observation (\code{row}).
}
\section{Examples}{
\preformatted{d <- data.frame(x = c(rep(1:3, each = 10), rep(NA, 10)),
                y = rep(1:2, 20),
                z = rep(c('male', 'female', 'missing', NA), each = 10))

with(d, rowmiss(x, z))
with(d, rowmiss())
}
}

\seealso{
Other data management: 
\code{\link{append}()},
\code{\link{codebook}()},
\code{\link{describe}()},
\code{\link{duplicates}()},
\code{\link{egen}()},
\code{\link{encode}()},
\code{\link{expand2}()},
\code{\link{formatDate}()},
\code{\link{generate}()},
\code{\link{label}()},
\code{\link{n_}()},
\code{\link{recode}()},
\code{\link{replace}()}
}
\concept{data management}
