% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterPlotMatrix.R
\name{scatterPlotMatrix}
\alias{scatterPlotMatrix}
\title{Scatter plot matrices with Histograms and
Correlation Coefficients}
\usage{
scatterPlotMatrix(data, pch = 21, ...)
}
\arguments{
\item{data}{Dataset}

\item{pch}{point symbol}

\item{...}{Arguments to be passed from \code{pairs}.}
}
\description{
A matrix of scatterplots is produced with
Scatter plots with smooth regression line in lower panel,
histograms in diagonal panel and Pearson's correlation
coefficients in upper panel.
}
\examples{

## iris data
# scatterPlotMatrix(iris)

}
\author{
Email: \email{dr.myominnoo@gmail.com}

Website: \url{https://myominnoo.github.io/}
}
