% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening.R
\name{screening}
\alias{screening}
\title{Screening tests}
\usage{
screening(x, y = NULL, p = NULL)
}
\arguments{
\item{x}{a factor object or a table}

\item{y}{an optional factor object}

\item{p}{disease prevalence}
}
\description{
\code{screening} reports sensitivity, specificity, positive and negative predictive
values
}
\details{
The screening tests are based on Bayes' Theorem. These tests help clinicians to
correctly predict the presence or absence of a particular disease from the
knowlege of test results (positive or negative) and/or the status of presenting
symptoms (present or absent).

information regarding the likelihood of positive and negative test results and
the likelihood of the presence or absence of a particular symptom in patients with
and without a particular disease.

\strong{Reference:}
\enumerate{
  \item Biostatistics A Foundation for Analysis in the Health Sciences (10th Edition).
  Chapter 3.5 BAYES’ THEOREM, SCREENING TESTS, SENSITIVITY, SPECIFICITY
}
}
\examples{
# Biostatistics A Foundation for Analysis in the Health Sciences (10th Edition)
# numbers taken from Example 3.5.1
t1 <- as.table(matrix(c(436, 5, 14, 495), byrow = TRUE, ncol = 2))
screening(t1, p = .113)
}
\seealso{
\code{\link{strate}}
}
\author{
Myo Minn Oo (Email: \email{dr.myominnoo@gmail.com} |
Website: \url{https://myominnoo.github.io/})
}
\keyword{diagnostic}
\keyword{npv}
\keyword{ppv,}
\keyword{screening}
\keyword{sensitivity,}
\keyword{specificity,}
\keyword{tests,}
