% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{splitByColon}
\alias{splitByColon}
\title{Split variable names separated by colon}
\usage{
splitByColon(.data, .vars.names, .colon)
}
\arguments{
\item{.data}{Dataset}

\item{.vars.names}{Variables}

\item{.colon}{Logical. \code{TRUE} indicates containing colon.}
}
\value{
character vector
}
\description{
\code{splitByColon()} split arguments by colon separator ":"
}
