% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{splitTables}
\alias{splitTables}
\title{Split tables to make \verb{2x2} tables}
\usage{
splitTables(.tbl, .exp.value = NULL)
}
\arguments{
\item{.tbl}{table}

\item{.exp.value}{reference row}
}
\description{
\code{splitTables()} separates table with rows more than 2 into
several tables using a reference row.
}
\details{
This function is used in calculating risks, odds and associated measures.
}
