% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stmh.R
\name{stmh}
\alias{stmh}
\alias{stmh.character}
\alias{stmh.list}
\alias{calc_RR}
\title{Calculate Incidence Rate Ratios from time-to-event data}
\usage{
stmh(
  data = NULL,
  time,
  status,
  by,
  fail = 1,
  ref_value = NULL,
  strata = NULL,
  per = 1,
  rnd = 4,
  plot = TRUE,
  print.table = TRUE
)

\method{stmh}{character}(
  data = NULL,
  time,
  status,
  by,
  fail = 1,
  ref_value = NULL,
  strata = NULL,
  per = 1,
  rnd = 4,
  plot = TRUE,
  print.table = TRUE
)

\method{stmh}{list}(
  data = NULL,
  time,
  status,
  by,
  fail = 1,
  ref_value = NULL,
  strata = NULL,
  per = 1,
  rnd = 4,
  plot = TRUE,
  print.table = TRUE
)

calc_RR(t, ref_lvl, rnd)
}
\arguments{
\item{data}{an optional data frame}

\item{time}{specify the timer interval when the subject is at risk}

\item{status}{event of interest to calculate incidence: 1 for event, 0 for censored}

\item{by}{specify variable to calculate stratified rates}

\item{fail}{failure event}

\item{ref_value}{a number or character representing reference value}

\item{strata}{variable to stratify by Mantel-Haenszel method}

\item{per}{units to be used in reported rates}

\item{rnd}{Rounding of numbers}

\item{plot}{logical value to display plots of rates across a categorical
variable}

\item{print.table}{logical value to display formatted outputs}

\item{t}{2x2 table input to calaculate odds ratio of non-reference}

\item{ref_lvl}{logical vector to indicate where reference category is}
}
\description{
\code{stmh} calculates incidence rate ratios (IRR) of the cohort.
}
\details{
Rates of event occurrences, known as incidence rates are outcome measures in
longitudinal studies. In most longitudinal studies, follow-up times vary due
to logistic resasons, different periods of recruitment, delay enrolment into
the study, lost-to-follow-up, immigration or emigration and death.

\strong{Follow-up time in longitudinal studies}

Period of observation (called as follow-up time) starts when individuals join
the study and ends when they either have an outcome of interest, are lost-to-
follow-up or the follow-up period ends, whichever happens first. This period is
called \strong{person-year-at-risk}. This is denoted by \emph{PY} in \code{strate}
function's output and numer of event by \emph{D}.

\strong{Rate Ratios}

is calcluated using the following formula:
\deqn{RR = Rate of Exposured Group / Rate of Non-Exposed Group}

\strong{Confidence interval of RR}

is derived using the following formula:

\deqn{95\% CI (RR) = Rate x Error Factor}

\deqn{Error Factor (rate) = exp(1.96 / \sqrt{1/D1 + 1/D2})}

\strong{Test of null hypothesis}

is calculated as follows:

\deqn{E1 = ( D * Y1/Y)}

where D = (D0 + D1) is the total number of events, and
Y= (Y0 + Y1) is the total follow-up time.

The test is based on the difference between the observed number of
events in the exposed group and its expected value, D1 – E1 , which
we denote by U. The variance of U, this difference, is equal to:

\deqn{V = D \* (Y1/Y) \* ( 1 – (Y1/Y) )}

\strong{The square of the standardized difference}

\deqn{U^2 / V}

which is referred to the χ2 distribution with 1 degree of freedom (df).

\code{plot}, if \code{TRUE}, produces a graph of the rate against
the numerical code used for categories of \code{by}.
}
\examples{
\dontrun{
library(lung)
data(lung)
str(lung)

## incidence rates
strate(lung, time, status)

## stratified incidence rates
strate(lung, time, status, sex, per = 100000)
stmh(lung, time, status, sex, per = 100000)

## create ageband
lung1 <- egen(lung, age, c(56, 61, 71), var_new = ageband)
tab(lung1, ageband)

strate(lung1, time, status, ageband, per = 100000)
stmh(lung1, time, status, ageband, per = 100000)

## Changing reference value
stmh(lung1, time, status, ageband, per = 100000,
     ref_value = "lbl[71-82]")

data(pbc)
str(pbc)
strate(pbc, time, status, trt, per = 100000)
stmh(pbc, time, status, trt, per = 100000)

strate(pbc, time, status, sex, per = 100000)
stmh(pbc, time, status, sex, per = 100000)
}
}
\references{
\enumerate{
  \item Essential Medical Statistics, Betty R. Kirkwood & Jonathan A.C. Sterne,
  Second Edition. Chapter 22, page 229 & 239
}
}
\seealso{
\code{\link{strate}}, \code{\link{mhodds}}, \code{\link{tabOdds}},
\code{\link{tabRisks}}
}
\author{
For any feedback, please contact \code{Myo Minn Oo} via:

Email: \email{dr.myominnoo@gmail.com}

Website: \url{https://myominnoo.github.io/}
}
\keyword{Incidence}
\keyword{Rate}
\keyword{Rates,}
\keyword{Ratio,}
\keyword{rates}
