% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strate.R
\name{strate}
\alias{strate}
\title{Calculate incidence rates from time-to-event data}
\usage{
strate(data = NULL, time, status, by = NULL, fail = NULL, per = 1,
  rnd = 5)
}
\arguments{
\item{data}{an optional data frame}

\item{time}{specify the timer interval when the subject is at risk}

\item{status}{event of interest to calculate incidence: 1 for event, 0 for censored}

\item{by}{specify variable to calculate stratified rates}

\item{fail}{failure event}

\item{per}{units to be used in reported rates}

\item{rnd}{Rounding of numbers}
}
\description{
\code{strate} calculates incidence rates of the cohort.
}
\details{
Rates of event occurrences, known as incidence rates are outcome measures in
longitudinal studies. In most longitudinal studies, follow-up times vary due
to logistic resasons, different periods of recruitment, delay enrolment into
the study, lost-to-follow-up, immigration or emigration and death.

\strong{Follow-up time in longitudinal studies}

Period of observation (called as follow-up time) starts when individuals join
the study and ends when they either have an outcome of interest, are lost-to-
follow-up or the follow-up period ends, whichever happens first. This period is
called \strong{person-year-at-risk}. This is denoted by \emph{PY} in \code{strate}
function's output and numer of event by \emph{D}.

\strong{Rate}

is calcluated using the following formula:
\deqn{\lambda = D / PY}

\strong{Confidence interval of rate}

is derived using the following formula:

\deqn{95\% CI (rate) = rate x Error Factor}
\deqn{Error Factor (rate) = exp(1.96 / \sqrt{D})}

\strong{References:}
\enumerate{
  \item Essential Medical Statistics, Betty R. Kirkwood & Jonathan A.C. Sterne,
  Second Edition. Chapter 22, page 229 & 239
}
}
\examples{
\dontrun{
library(survival)
data(lung)
str(lung)

## incidence rates
strate(lung, time, status)
strate(lung, time, status, fail = 2)

## stratified incidence rates
strate(lung, time, status, by = sex)
strate(lung, time, status, by = sex, fail = 2)
strate(lung, time, status, by = sex, fail = 2, per = 1000)
strate(lung, time, status, by = sex, fail = 2, per = 100000, rnd = 1)
strate(time, status, data = lung, per = 100000)

data(pbc)
str(pbc)
strate(pbc, time, status)
strate(pbc, time, status, trt)
strate(pbc, time, status, sex)
strate(pbc, time, status, stage, per = 100000)
}
}
\seealso{
\code{\link{summ}}, \code{\link{summBy}}, \code{\link{tab}},
\code{\link{xtab}}
}
\author{
Myo Minn Oo (Email: \email{dr.myominnoo@gmail.com} |
Website: \url{https://myominnoo.github.io/})
}
\keyword{incidence}
\keyword{rate}
