% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summ.R
\name{summ}
\alias{summ}
\title{Summary statistics}
\usage{
summ(
  data,
  ...,
  by = NULL,
  na.rm = FALSE,
  test = "auto",
  paired = FALSE,
  digits = 1,
  detail = FALSE
)
}
\arguments{
\item{data}{A data.frame}

\item{...}{One or more unquoted variables separated by commas.
Variable names can be used as if they were positions in the data
frame, so expressions like \code{x:y} can be used to select a range of
variables.}

\item{by}{Stratification Variable}

\item{na.rm}{logical value: if \code{TRUE}, it removes observations with missing values.}

\item{test}{name of test
\itemize{
\item \code{ttest}   - p-value from \verb{independent t-test} using \link{t.test}
\item \code{wilcox}   - p-value from \verb{Wilcoxon Rank-Sum test} using \link{wilcox.test}
\item \code{anova}    - p-value from Analysis of variance using \link{aov}
\item \code{kwallis} - p-value from \verb{Kruskal-Wallis Rank Sum Test} using \link{kruskal.test}
}}

\item{paired}{a logical indicating whether you want a paired
    t-test.}

\item{digits}{specify rounding of numbers.}

\item{detail}{logical: \code{TRUE} displays a full spectrum of
summary statistics and includes p-values from hypothesis tests.}
}
\description{
\Sexpr[results=rd]{lifecycle::badge("stable")}

\code{summ()} calculates and displays a variety of univariate summary statistics.
\itemize{
\item \code{obs}      - Number of observations
\item \code{NA}       - Number of observations with missing value
\item \code{mean}     - Mean
\item \code{SD}       - Standard deviation
\item \code{median}   - Median
\item \code{p25}       - First quartile or percentile
\item \code{p75}       - Third quartile or percentile
}
\subsection{\code{detail}: \code{TRUE}}{
\itemize{
\item \code{min}      - Minimum value
\item \code{max}      - Maximum value
\item \code{normal}   - p-value from \verb{Shapiro-Wilk Normality Test} using \link{shapiro.test}
\item \code{ttest}   - p-value from \verb{independent t-test} using \link{t.test}
\item \code{wilcox}   - p-value from \verb{Wilcoxon Rank-Sum test} using \link{wilcox.test}
\item \code{anova}    - p-value from Analysis of variance using \link{aov}
\item \code{kwallis} - p-value from \verb{Kruskal-Wallis Rank Sum Test} using \link{kruskal.test}
}
}

\subsection{Summarizing the whole dataset}{

This is helpful when the dataset has been processed and finalized.
The final dataset can be summarized without specifying any variables.
This automatically filters and generates summary statistics
for variables that are of type \code{numeric}, \code{double}, \code{integer}, or \code{logical}.
}
}
\section{Examples}{


Summarize variable by name:\if{html}{\out{<div class="r">}}\preformatted{summ(iris, Petal.Length)
#>   Summary statistics
#>      variable | obs NA mean  sd median p25 p75
#>  ------------ + --- -- ---- --- ------ --- ---
#>  Petal.Length | 150  0  3.8 1.8    4.3 1.6 5.1
#>  ------------ + --- -- ---- --- ------ --- ---
}\if{html}{\out{</div>}}

Summarize multiple variables by separating them with commas:\if{html}{\out{<div class="r">}}\preformatted{summ(iris, Sepal.Length, Petal.Length)
#>   Summary statistics
#>      variable | obs NA mean  sd median p25 p75
#>  ------------ + --- -- ---- --- ------ --- ---
#>  Sepal.Length | 150  0  5.8 0.8    5.8 5.1 6.4
#>  Petal.Length | 150  0  3.8 1.8    4.3 1.6 5.1
#>  ------------ + --- -- ---- --- ------ --- ---
}\if{html}{\out{</div>}}

The \code{:} operator selects a range of consecutive variables:\if{html}{\out{<div class="r">}}\preformatted{summ(iris, Sepal.Length:Petal.Length)
#>   Summary statistics
#>      variable | obs NA mean  sd median p25 p75
#>  ------------ + --- -- ---- --- ------ --- ---
#>  Sepal.Length | 150  0  5.8 0.8    5.8 5.1 6.4
#>   Sepal.Width | 150  0  3.1 0.4    3.0 2.8 3.3
#>  Petal.Length | 150  0  3.8 1.8    4.3 1.6 5.1
#>  ------------ + --- -- ---- --- ------ --- ---
}\if{html}{\out{</div>}}

Summarize variables with a stratification variable \code{by}:\if{html}{\out{<div class="r">}}\preformatted{summ(iris, Sepal.Length:Petal.Length, by = Species)
#>  Summary statistics
#>       by levels of `Species`
#>      variable level | obs NA mean  sd median p25 p75 anova
#>  ------------ ----- + --- -- ---- --- ------ --- --- -----
#>  Sepal.Length       |                                0.000
#>                   1 |  50  0  5.0 0.4    5.0 4.8 5.2      
#>                   2 |  50  0  5.9 0.5    5.9 5.6 6.3      
#>                   3 |  50  0  6.6 0.6    6.5 6.2 6.9      
#>               ----- + --- -- ---- --- ------ --- --- -----
#>               Total | 150  0  5.8 0.8    5.8 5.1 6.4      
#>  ------------ ----- + --- -- ---- --- ------ --- --- -----
#>   Sepal.Width       |                                0.000
#>                   1 |  50  0  3.4 0.4    3.4 3.2 3.7      
#>                   2 |  50  0  2.8 0.3    2.8 2.5 3.0      
#>                   3 |  50  0  3.0 0.3    3.0 2.8 3.2      
#>               ----- + --- -- ---- --- ------ --- --- -----
#>               Total | 150  0  3.1 0.4    3.0 2.8 3.3      
#>  ------------ ----- + --- -- ---- --- ------ --- --- -----
#>  Petal.Length       |                                0.000
#>                   1 |  50  0  1.5 0.2    1.5 1.4 1.6      
#>                   2 |  50  0  4.3 0.5    4.3 4.0 4.6      
#>                   3 |  50  0  5.6 0.6    5.6 5.1 5.9      
#>               ----- + --- -- ---- --- ------ --- --- -----
#>               Total | 150  0  3.8 1.8    4.3 1.6 5.1      
#>  ------------ ----- + --- -- ---- --- ------ --- --- -----
}\if{html}{\out{</div>}}

Summarize all categorical variables without specifying any variables:\if{html}{\out{<div class="r">}}\preformatted{summ(iris)
#>   Summary statistics
#>      variable | obs NA mean  sd median p25 p75
#>  ------------ + --- -- ---- --- ------ --- ---
#>  Sepal.Length | 150  0  5.8 0.8    5.8 5.1 6.4
#>   Sepal.Width | 150  0  3.1 0.4    3.0 2.8 3.3
#>  Petal.Length | 150  0  3.8 1.8    4.3 1.6 5.1
#>   Petal.Width | 150  0  1.2 0.8    1.3 0.3 1.8
#>  ------------ + --- -- ---- --- ------ --- ---
summ(iris, by = Species)
#>  Summary statistics
#>       by levels of `Species`
#>      variable level | obs NA mean  sd median p25 p75 anova
#>  ------------ ----- + --- -- ---- --- ------ --- --- -----
#>  Sepal.Length       |                                0.000
#>                   1 |  50  0  5.0 0.4    5.0 4.8 5.2      
#>                   2 |  50  0  5.9 0.5    5.9 5.6 6.3      
#>                   3 |  50  0  6.6 0.6    6.5 6.2 6.9      
#>               ----- + --- -- ---- --- ------ --- --- -----
#>               Total | 150  0  5.8 0.8    5.8 5.1 6.4      
#>  ------------ ----- + --- -- ---- --- ------ --- --- -----
#>   Sepal.Width       |                                0.000
#>                   1 |  50  0  3.4 0.4    3.4 3.2 3.7      
#>                   2 |  50  0  2.8 0.3    2.8 2.5 3.0      
#>                   3 |  50  0  3.0 0.3    3.0 2.8 3.2      
#>               ----- + --- -- ---- --- ------ --- --- -----
#>               Total | 150  0  3.1 0.4    3.0 2.8 3.3      
#>  ------------ ----- + --- -- ---- --- ------ --- --- -----
#>  Petal.Length       |                                0.000
#>                   1 |  50  0  1.5 0.2    1.5 1.4 1.6      
#>                   2 |  50  0  4.3 0.5    4.3 4.0 4.6      
#>                   3 |  50  0  5.6 0.6    5.6 5.1 5.9      
#>               ----- + --- -- ---- --- ------ --- --- -----
#>               Total | 150  0  3.8 1.8    4.3 1.6 5.1      
#>  ------------ ----- + --- -- ---- --- ------ --- --- -----
#>   Petal.Width       |                                0.000
#>                   1 |  50  0  0.2 0.1    0.2 0.2 0.3      
#>                   2 |  50  0  1.3 0.2    1.3 1.2 1.5      
#>                   3 |  50  0  2.0 0.3    2.0 1.8 2.3      
#>               ----- + --- -- ---- --- ------ --- --- -----
#>               Total | 150  0  1.2 0.8    1.3 0.3 1.8      
#>  ------------ ----- + --- -- ---- --- ------ --- --- -----
}\if{html}{\out{</div>}}
}

\seealso{
Other statistical analysis: 
\code{\link{tab}()}
}
\concept{statistical analysis}
