% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isum.date.R
\name{summ_date}
\alias{summ_date}
\title{A Quick Summary function for Object of type "Date"}
\usage{
summ_date(x, rnd = 1, na.rm = TRUE)
}
\arguments{
\item{x}{An R object: it can be a dataframe or a vector.}

\item{rnd}{an integer indicating the number of decimal places:}

\item{na.rm}{A logical value indicating whether "NA" missing values should be
removed before the computation proceeds.}
}
\description{
summ_date() lets you do a quick summary on your data
}
\examples{
isum(iris)
}
\seealso{
isum.factor, isum.numeric, isum.data.frame
}
\keyword{basic}
\keyword{date,}
\keyword{quick}
\keyword{statistics,}
\keyword{summarize,}
\keyword{summary}
