% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary_mStats}
\alias{summary_mStats}
\alias{summary.tab1}
\alias{summary.tab2p}
\alias{summary.tab2}
\alias{summary.summ1}
\alias{summary.summ2}
\alias{summary.summ2t}
\title{Table Format for Publication}
\usage{
\method{summary}{tab1}(object, ...)

\method{summary}{tab2p}(object, p = 2, ...)

\method{summary}{tab2}(object, ...)

\method{summary}{summ1}(object, ...)

\method{summary}{summ2}(object, stats = "all", ...)

\method{summary}{summ2t}(object, stats = "all", test = TRUE, ...)
}
\arguments{
\item{object}{an object for which a summary is desired.}

\item{...}{additional arguments affecting the summary produced.}

\item{p}{Percentage columns present - TRUE or NOT - FALSE}

\item{stats}{Statistics to be generated.

"all" - all statistics
"mean" - mean (Standard Deviation)
"median" - median (Interquartile Range)
"range" - minimum - maximum}

\item{test}{TRUE reports p-values from statistical tests}
}
\description{
\code{summary()} gathers statistics
in table format for publication.

\code{tab1} for one-way tabulation of \code{tab()}

\code{tab2p} for two-way tabulation with percentages of \code{tab}

\code{tab2} for two-way tabulation without
percentages of \code{tab}

\code{summ1} for summary measures of \code{summ}

\code{summ2} for summary measures of \code{summ}

\code{summ2} for summary measures of \code{summ}
}
