% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumstats.R
\name{sumstats}
\alias{sumstats}
\title{Tabulation of summary statistics by group field.}
\usage{
sumstats(df, group, value, digits)
}
\arguments{
\item{df}{Data frame for generation of statistics}

\item{group}{Grouping field}

\item{value}{Numeric field to have statistics calculated}

\item{digits}{Round to this many digits}
}
\value{
Markdown format table as character string
}
\description{
Markdown formatted tabulation of summary statistics by grouping field.
}
\author{
Alex M Trueman
}
