% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableToDF.R
\name{tableTOdf}
\alias{tableTOdf}
\title{Convert Table to Data Frame (Reversing Table)}
\usage{
tableTOdf(table, vars = NULL, x.levels = NULL, y.levels = NULL)
}
\arguments{
\item{table}{a table input}

\item{vars}{specify names of two variables}

\item{x.levels}{a vector corresponding to the levels in Column variable}

\item{y.levels}{a vector corresponding to the levels in Row variable}
}
\description{
\code{\link{tableTOdf}} transforms a contigency table into individual data frame.
}
\details{
By default, it takes a table and returns a data frame with rows corresponding to the
numbers in that table. If names are not specified, it returns data that contains "x"
and "y" variables. Names and levels can be specified.
See examples below.
}
\examples{
t <- table(infert$education, infert$case)
t

## get raw data from table
head(tableTOdf(t))

## change variables' names
head(tableTOdf(t, vars = c("edu", "case")))

## specify names of levels of variables
head(tableTOdf(t, vars = c("edu", "case"), x.levels = c("No", "Yes")))

str(tableTOdf(t, vars = c("edu", "case"), x.levels = c("No", "Yes")))
table(tableTOdf(t, vars = c("edu", "case"), x.levels = c("No", "Yes")))
}
\seealso{
\code{\link{isum}}, \code{\link{itab}}
}
\author{
Myo Minn Oo (Email: \email{dr.myominnoo@gmail.com} |
Website: \url{https://myominnoo.github.io/})
}
\keyword{data}
\keyword{frame,}
\keyword{get}
\keyword{raw}
\keyword{table}
\keyword{to}
