% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabRisks.R
\name{tabRisks}
\alias{tabRisks}
\title{Calculating Risks and Relative Risks}
\usage{
tabRisks(
  data,
  ...,
  by,
  exp_value = NULL,
  case_value = NULL,
  plot = TRUE,
  na.rm = FALSE,
  rnd = 3
)
}
\arguments{
\item{data}{Dataset}

\item{...}{Variable or multiple variables
Colon separator \code{:} can be used to specify multiple variables.}

\item{by}{Varaiable for cross-tabulation}

\item{exp_value}{value for exposure as reference}

\item{case_value}{value for outcome as reference}

\item{plot}{logical value to display plots of rates across a categorical
variable}

\item{na.rm}{A logical value to specify missing values, \code{NA} in the table}

\item{rnd}{specify rounding of numbers. See \code{\link{round}}.}
}
\description{
\code{tabRisks()} cross-tabulates two variables and
reports risks of failure \code{by} among exposed and
unexposed levels of explanatory variable \code{...}.
It is used in cross-sectional studies.
}
\details{
Risks are sometimes called proportions, incidence proportions or
prevalence.

\strong{Calculating Risks}

\deqn{Risks = Outcome of Interest (A) / Sample Size (n)}

\deqn{Standard Error, SE = \sqrt(p x (1 - p) / n)}

\deqn{95\% CI = Risks +/- (1.96 x SE)}
}
\note{
This method should be avoided in small samples. Quadratic or
exact binomial methods are preferred in this regard.
}
\examples{
\dontrun{


## Example from Altman Statistics with Confidence, Chapter 7, Page 59
ulcers <- expandTables(c(6, 112, 16, 729),
                       exp_name = "pylori",
                       exp_lvl = c("Yes", "No"),
                       case_name = "history",
                       case_lvl = c("Yes", "No")) \%>\%
   labelVar(c(pylori, history),
            c("Infected with H. pylori", "History of Ulcer in mother")) \%>\%
   labelData(paste0("Prevalence of H. pylori infection in preschool children",
                    " according to mother's history of ulcer"))


## codebook
codebook(ulcers)


## tabulate risks with plot
tabRisks(ulcers, pylori, by = history)
tabRisks(ulcers, pylori, by = history, exp_value = "Yes", case_value = "Yes")

## without plot
tabRisks(ulcers, pylori, by = history, exp_value = "Yes", case_value = "Yes",
         plot = FALSE)
}

}
\references{
\enumerate{
\item Essential Medical Statistics, Betty R. Kirkwood & Jonathan A.C.
Sterne, Second Edition.
\item B. Burt Gerstman. Epidemology Kept Simple. Wiley-Blackwell, 2013:
\item Statistics with confidence, Douglas G Altman, second edition
}
}
\author{
For any feedback, please contact \code{Myo Minn Oo} via:

Email: \email{dr.myominnoo@gmail.com}

Website: \url{https://myominnoo.github.io/}
}
\concept{Risks risk ratio incidence prevalence

frequency table statistics descriptive}
