% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_duplicates.R
\name{tag_duplicates}
\alias{tag_duplicates}
\title{Tag Duplicates}
\usage{
tag_duplicates(..., .add_tags = TRUE)
}
\arguments{
\item{...}{Columns to use for identifying duplicates.}

\item{.add_tags}{logical to return three indicator columns: \code{.n_}, \code{.N_}, and \code{.dup_}.}
}
\value{
A tibble with three columns: \code{.n_}, \code{.N_}, and \code{.dup_}.
\itemize{
\item \code{.n_} represents the running counter within each group of variables,
indicating the number of the current observation.
\item \code{.N_} represents the total number of observations within each group of variables.
\item \code{.dup_} is a logical column indicating
whether the observation is a duplicate (TRUE) or not (FALSE).
}
}
\description{
\Sexpr[results=rd]{lifecycle::badge("stable")}
}
\details{
This function identifies and tags duplicate observations based on specified variables.

This function mimics the functionality of Stata's \code{duplicates} command in R.
It calculates the number of duplicates and provides a report of duplicates
based on the specified variables. The function utilizes the \link{n_} and \link[=n_]{N_} functions
for counting and grouping the observations.
}
\examples{

library(dplyr)

# Example with a custom dataset
data <- data.frame(
  x = c(1, 1, 2, 2, 3, 4, 4, 5),
  y = letters[1:8]
)

# Identify and tag duplicates based on the "x" variable
data \%>\% mutate(tag_duplicates(x))

# Identify and tag duplicates based on multiple variables
data \%>\% mutate(tag_duplicates(x, y))

# Identify and tag duplicates based on all variables
data \%>\% mutate(tag_duplicates(everything()))

\dontrun{
## STATA example
dupxmpl <- haven::read_dta("https://www.stata-press.com/data/r18/dupxmpl.dta")
dupxmpl |> mutate(tag_duplicates(everything()))
}

}
\seealso{
Other Data Management: 
\code{\link{append}()},
\code{\link{codebook}()},
\code{\link{count_functions}},
\code{\link{cut}()}
}
\concept{Data Management}
