% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testModelFit.R
\name{testModelFit}
\alias{testModelFit}
\title{Regression models, reporting overall significance of the model}
\usage{
testModelFit(model)
}
\arguments{
\item{model}{object glm or lm type}
}
\description{
\code{testModelFit} overall significance of the model and interpretation
}
\details{
This test asks whether the model with predictors fits significantly better
than a model with just an intercept (i.e., a null model).

The test statistic is the difference between the residual deviance for the
model with predictors and the null model. The test statistic is distributed
chi-squared with degrees of freedom equal to the differences in degrees of
freedom between the current and the null model (i.e., the number of predictor
variables in the model).

\strong{Reference: }
\enumerate{
  \item LOGIT REGRESSION | R DATA ANALYSIS EXAMPLES. UCLA: Statistical Consulting
  Group. from https://stats.idre.ucla.edu/r/dae/logit-regression/ (accessed
  September 27, 2019)
}
}
\examples{
\dontrun{
## example from IRDE website:
## https://stats.idre.ucla.edu/r/dae/logit-regression/
mydata <- read.csv("https://stats.idre.ucla.edu/stat/data/binary.csv")
codebook(mydata)
tab(admit, mydata)
tab(rank, mydata)

mylogit <- glm(admit ~ gre + gpa + factor(rank), data = mydata, family = "binomial")
summary(mylogit)

fModelOutput(mylogit) # generates parameters
testModelFit(mylogit) # test overall significant of the model
}
}
\seealso{
\code{\link{fModelOutput}}
}
\author{
Myo Minn Oo (Email: \email{dr.myominnoo@gmail.com} |
Website: \url{https://myominnoo.github.io/})
}
\keyword{fit,}
\keyword{model}
\keyword{of}
\keyword{significance}
