% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toDate.R
\name{toDate}
\alias{toDate}
\alias{is.Date}
\title{Coerces or test vectors to Date type}
\usage{
toDate(x, format = "a")

is.Date(x)
}
\arguments{
\item{x}{object to be coerced or tested}

\item{format}{a character value, representing the function of Date conversion or coercion.
\itemize{
  \item \strong{"a"} : Year-month-day
  \item \strong{"b"} : Year-day-month
  \item \strong{"c"} : day-month-Year
  \item \strong{"d"} : month-day-Year
  \item \strong{"e"} : Microsoft excel number whose origin is set to "1899-12-30".
}}
}
\description{
\code{toDate} coerces objects to Date type. 
\code{is.Date} indicates which vector is type of Date.
}
\examples{
x <- c("2019-01-15", "2019-01-20", "2019-01-21", "2019-01-22")
is.Date(x)
toDate(x)
y <- toDate(x)
is.Date(y)
y
# dates in numeric format MS excel
x <- 42705:42710
y <- toDate(x, "e")
y
}
\seealso{
\code{\link{dateSum}}
}
\author{
Myo Minn Oo (Email: \email{dr.myominnoo@gmail.com} |
Website: \url{https://myominnoo.github.io/})
}
\keyword{coerces,}
\keyword{date}
\keyword{test}
