% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumstats.R
\name{topcut_sumstats}
\alias{topcut_sumstats}
\title{Tabulation of top cutting statistics by group field.}
\usage{
topcut_sumstats(df, group, value, value_tc, weight, digits)
}
\arguments{
\item{df}{Data frame for generation of statistics}

\item{group}{Grouping field}

\item{value}{Numeric field to have statistics calculated}

\item{value_tc}{Numeric top cut field for statistics}

\item{weight}{Weight field for statistics}

\item{digits}{Round to this many digits}
}
\value{
Markdown format table as character string
}
\description{
Markdown formatted tabulation of weighted summary statistics by grouping field.
}
\author{
Alex M Trueman
}
