% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumstats.R
\name{weighted_sumstats}
\alias{weighted_sumstats}
\title{Tabulation of weighted summary statistics by group field.}
\usage{
weighted_sumstats(df, group, value, weight, digits)
}
\arguments{
\item{df}{Data frame for generation of statistics}

\item{group}{Grouping field}

\item{value}{Numeric field to have statistics calculated}

\item{weight}{Weight field for statistics}

\item{digits}{Round to this many digits}
}
\value{
Markdown format table as character string
}
\description{
Markdown formatted tabulation of weighted summary statistics by grouping field.
}
\author{
Alex M Trueman
}
