% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtab.R
\name{xtab}
\alias{xtab}
\alias{xtab.default}
\alias{xtab.character}
\alias{xtab.list}
\alias{xtab.data.frame}
\title{Cross tabulation, contigency table or two-by-two table}
\usage{
xtab(
  x,
  y,
  data = NULL,
  row.pct = TRUE,
  na.rm = FALSE,
  rnd = 1,
  print.table = TRUE
)

\method{xtab}{default}(...)

\method{xtab}{character}(
  x,
  y,
  data = NULL,
  row.pct = TRUE,
  na.rm = FALSE,
  rnd = 1,
  print.table = TRUE
)

\method{xtab}{list}(
  x,
  y,
  data = NULL,
  row.pct = TRUE,
  na.rm = FALSE,
  rnd = 1,
  print.table = TRUE
)

\method{xtab}{data.frame}(
  x,
  y,
  data = NULL,
  row.pct = TRUE,
  na.rm = FALSE,
  rnd = 1,
  print.table = TRUE
)
}
\arguments{
\item{x}{a character, factor or logical object OR list OR data frame}

\item{y}{a character, factor or logical object}

\item{data}{a data frame object (Optional)}

\item{row.pct}{a logical value: if NULL, a default table without any percentages
is produced. If TRUE, row percentages are shown and if FALSE, column percentages.}

\item{na.rm}{A logical value to specify missing values, <NA> in the table}

\item{rnd}{specify rounding of numbers. See \code{\link{round}}.}

\item{print.table}{logical value to display formatted outputs}

\item{...}{optional arguments}
}
\description{
\code{xtab} generates cross tabulation of two variables.
}
\details{
Exploring data before jumping into complex analysis is always a necessity.
The first step of an analysis is always to summarize and display data.

\code{xtab}
produce contigency table.

\strong{References:}
\enumerate{
  \item Essential Medical Statistics, Betty R. Kirkwood & Jonathan A.C. Sterne,
  Second Edition. Chapter 3
  \item An Introduction to MEdical Statistics, Martin Bland, Thrid Edition,
  Chapter 4
}
}
\examples{
xtab(education, induced, infert)
xtab(infert$education, infert$induced)

xtab(spontaneous, induced, infert, row.pct = FALSE) # column percentage
xtab(spontaneous, induced, infert, row.pct = NULL) # DO NOT SHOW PERCENTAGE

# multiple variables
xtab(c(spontaneous, case, induced), education, infert)

\dontrun{
# variables' labels as footnote
infert.new <- labelVars(infert, c(spontaneous, case, induced),
               c("SPONTANEOUS", "CASE YES OR NO", "INDUCED"))
xtab(c(spontaneous, case, induced), education, infert.new)
}

# whole dataset
xtab(infert, case)
xtab(infert, induced)
}
\seealso{
\code{\link{tab}}
}
\author{
Myo Minn Oo (Email: \email{dr.myominnoo@gmail.com} |
Website: \url{https://myominnoo.github.io/})
}
\keyword{2x2}
\keyword{descriptive}
\keyword{statistics,}
\keyword{table,}
\keyword{two-by-two}
\keyword{two-way}
