% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatDate.R
\name{year}
\alias{year}
\title{Year from a Date}
\usage{
year(x)
}
\arguments{
\item{x}{Date Object}
}
\description{
\code{year} retrieves year from a Date
}
\examples{
\dontrun{
dates <- formatDate(c("2019-01-15", "2019-01-20", "2019-01-21",
                      "2019-01-22"), "Ymd", "-")
year(dates)
}
}
\seealso{
\code{\link{formatDate}}, \code{\link{month}}, \code{\link{day}}
}
\author{
For any feedback, please contact \code{Myo Minn Oo} via:

Email: \email{dr.myominnoo@gmail.com}

Website: \url{https://myominnoo.github.io/}
}
\keyword{date}
\keyword{year,}
