% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append.R
\name{append}
\alias{append}
\title{Append datasets}
\usage{
append(data, ...)
}
\arguments{
\item{data}{A data.frame}

\item{...}{multiple datasets}
}
\description{
\Sexpr[results=rd]{lifecycle::badge("stable")}

\code{append()} combines multiple datasets to the end of the first dataset.
If variables are not the same or not in the same order, \code{append()}
automatically creates variables and put them into the first dataset.
}
\section{Examples}{


Mutliple datasets can be appended into \code{iris}, even if they do not share
the same variables.\preformatted{## multiple datasets without the same variables
d <- append(iris, infert, cars)
des(d)
}
}

\seealso{
Other data management: 
\code{\link{codebook}()},
\code{\link{describe}()},
\code{\link{duplicates}()},
\code{\link{egen}()},
\code{\link{encode}()},
\code{\link{expand2}()},
\code{\link{formatDate}()},
\code{\link{generate}()},
\code{\link{label}()},
\code{\link{n_}()},
\code{\link{recode}()},
\code{\link{replace}()},
\code{\link{rowmiss}()}
}
\concept{data management}
