% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append.R
\name{append}
\alias{append}
\title{APPEND datasets}
\usage{
append(data, ...)
}
\arguments{
\item{data}{Base dataset}

\item{...}{Datasets to be appended}
}
\value{
Modified dataset
}
\description{
\code{append()} row-combines multiple datasets of the same column names.
}
\details{
Single or multiple datasets can be appended.
At least one matching variable must be in the appending datasets.
If arrangement or order of variables are not the same, they are
automatically rematched to the MASTER dataset and appended.

\preformatted{append(data, data1, data2, data3, etc)}

\code{data} = MASTER DATASET
Remaining datasets: data1, data2, data3, etc = APPENDING DATASETS
}
\examples{

## use infert dataset
data(infert)
codebook(infert)

infert1 <- infert[-c(10:20), -2]
infert2 <- infert[-c(40:60), -4]

infert.new <- append(infert, infert1, infert2)
codebook(infert.new)

}
\author{
Email: \email{dr.myominnoo@gmail.com}

Website: \url{https://myominnoo.github.io/}
}
