% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append.R
\name{append}
\alias{append}
\title{Append multiple datasets}
\usage{
append(data, ...)
}
\arguments{
\item{data}{Base dataset}

\item{...}{Datasets to be appended}
}
\value{
Modified Dataset
}
\description{
\code{append()} joins multiple datasets of the same column names.
}
\details{
Multiple datasets can be appended only if they have the same variables in the
same order.

\preformatted{append(data, data1, data2, data3, etc)}
}
\examples{

## use infert dataset
data(infert)
codebook(infert)

infert.new <- append(infert, infert)
codebook(infert.new)


}
\author{
For any feedback, please contact \code{Myo Minn Oo} via:

Email: \email{dr.myominnoo@gmail.com}

Website: \url{https://myominnoo.github.io/}
}
