% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mStats.R
\name{append}
\alias{append}
\title{Append datasets}
\usage{
append(data, ...)
}
\arguments{
\item{data}{data.frame}

\item{...}{one or multiple data.frame}
}
\value{
data.frame
}
\description{
\code{append()} row-combines multiple datasets of the same column names.
}
\details{
A single or multiple datasets can be appended.

The appending datasets must have at least one variable name
which is there in the master dataset.

The order of variables of the appending datasets is automatically
set based on the variable arrangement of the master dataset.
}
\examples{

x <- append(infert[, -c(3,4)], infert[, -5], infert[, -6])
## codebook(x)

\dontrun{
## if no variables are matched, ERROR
append(infert, iris)
}

}
\author{
Email: \email{dr.myominnoo@gmail.com}

Website: \url{https://myominnoo.github.io/}
}
