% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear.R
\name{clear}
\alias{clear}
\title{Clear Global Environment, Plots and Console}
\usage{
clear()
}
\description{
\code{clear} is a quick way to clear messy Global Environment, plots and console 
without interrupting the search path. If you want to detach all user-loaded packages, 
see \code{\link{idetach}}.
}
\examples{
x <- rnorm(100)
plot(x)
\dontrun{
  clear()
}
}
\seealso{
\code{\link{idetach}}
}
\author{
Myo Minn Oo (Email: \email{dr.myominnoo@gmail.com} |
Website: \url{https://myominnoo.github.io/})
}
\keyword{all,}
\keyword{clear}
\keyword{console}
\keyword{environment,}
\keyword{global}
\keyword{plots,}
\keyword{remove}
