% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{codebook}
\alias{codebook}
\title{Detailed information of a dataset
and its variables}
\usage{
codebook(data)
}
\arguments{
\item{data}{name of a dataset}
}
\value{
\code{data.frame}
}
\description{
\code{codebook()} provides detailed information about the dataset
itself and its variables.
}
\details{
It reports data structure of the dataset specified with names of
variables, their labels, data types, number of observations,
number of observations with missing values and
percentage of observations with missing values.

\strong{ANNOTATIONS}:

Variable - Names of variables

Label     - Labels of variables

Type     - Types of variables

Obs_Num - Counts of valid observations

NA  - Counts of observations with missing value

NA(\%)    - Percentage of observations with missing value
}
\examples{

## use infert data
data(infert)
codebook(infert)


## add labels
infert.new <- labelVar(infert,
                       education = "EDUCATION",
                       age = "AGE IN YEARS",
                       parity = "count",
                       case = "case status",
                       induced = "# of prior induced abortions",
                       spontaneous = "# of prior spon. abortions",
                       stratum = "Matched set number",
                       pooled.stratum = "Stratum number")

infert.new <- labelData(infert.new,
    "Infertility after Spontaneous and Induced Abortion")
codebook(infert.new)


}
\author{
Email: \email{dr.myominnoo@gmail.com}

Website: \url{https://myominnoo.github.io/}
}
