% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{codebook}
\alias{codebook}
\alias{codebook.default}
\alias{codebook.data.frame}
\title{Codebook}
\usage{
codebook(x)

\method{codebook}{default}(x)

\method{codebook}{data.frame}(x)
}
\arguments{
\item{x}{dataframe}
}
\description{
\code{codebook} displays structure of a data frame
}
\details{
\code{codebook}

generates the report of data structure with names, data lables, types,
number of observations, number of observations with missing values and
percentage of observations with missing values.

ANNOTATIONS:

VARS_NAME - Names of variables
LABEL     - Labels of variables
TYPE      - Types of variables
OBS_COUNT - Counts of valid observations
NA_COUNT  - Counts of observations with missing value
NA_(%)    - Percentage of observations with missing value
}
\examples{
\dontrun{
codebook(infert)
codebook(iris)

# if something else
codebook(iris$Species)
}
}
\seealso{
\code{\link{ilog}}
}
\author{
Myo Minn Oo (Email: \email{dr.myominnoo@gmail.com} |
Website: \url{https://myominnoo.github.io/})
}
\keyword{codebook,}
\keyword{layout}
\keyword{structure,}
\keyword{summary,}
