% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{codebook}
\alias{codebook}
\alias{codebook.data.frame}
\alias{codebook.default}
\alias{output.format}
\alias{wrap.output}
\title{Codebook}
\usage{
codebook(x)

\method{codebook}{data.frame}(x)

\method{codebook}{default}(x)

output.format(f, title)

wrap.output(txt, width = 90, sep = "\\n")
}
\arguments{
\item{x}{Data Frame}

\item{f}{data frame}

\item{title}{label}

\item{txt}{character vector}

\item{width}{desired console width}

\item{sep}{wrapper}
}
\description{
\code{codebook} displays structure of a data frame
}
\details{
\code{codebook}

generates the report of data structure with names, data lables, types,
number of observations, number of observations with missing values and
percentage of observations with missing values.

ANNOTATIONS:

VARS_NAME - Names of variables
LABEL     - Labels of variables
TYPE      - Types of variables
OBS_COUNT - Counts of valid observations
NA_COUNT  - Counts of observations with missing value
(NA %)    - Percentage of observations with missing value
}
\examples{
str(infert)
codebook(infert)
str(iris)
codebook(iris)

# if something else
codebook(iris$Species)
}
\seealso{
\code{\link{ilog}}
}
\author{
Myo Minn Oo (Email: \email{dr.myominnoo@gmail.com} |
Website: \url{https://myominnoo.github.io/})
}
\keyword{codebook,}
\keyword{layout}
\keyword{structure,}
\keyword{summary,}
