% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{codebook}
\alias{codebook}
\title{Codebook: Detailed information about variables}
\usage{
codebook(data)
}
\arguments{
\item{data}{dataframe}
}
\value{
Codebook in data.frame format
}
\description{
\code{codebook()} provides detailed information about each variable
within the dataset.
}
\details{
\code{codebook} generates the report of data structure with names,
data lables, types,
number of observations, number of observations with missing values and
percentage of observations with missing values.

\strong{ANNOTATIONS}:

Variable - Names of variables

Label     - Labels of variables

Type     - Types of variables

_Obs - Counts of valid observations

_NA  - Counts of observations with missing value

_NA(\%)    - Percentage of observations with missing value
}
\examples{
\dontrun{
codebook(infert)
codebook(iris)

# if something else
codebook(iris$Species)



# Example from IDRE UCLA
path <- "https://stats.idre.ucla.edu/stat/data/patient_pt1_stata_dm.dta"
hosp <- haven::read_dta(path)
codebook(hosp)

library(magrittr)
hosp \%>\% labelData("Hospital Dataset Example") \%>\% codebook()
}

}
\references{
STATA DATA MANAGEMENT. UCLA: Statistical Consulting Group.
from https://stats.idre.ucla.edu/stata/seminars/stata-data-management/
(accessed Febrary 25, 2020).
}
\author{
For any feedback, please contact \code{Myo Minn Oo} via:

Email: \email{dr.myominnoo@gmail.com}

Website: \url{https://myominnoo.github.io/}
}
