% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{codebook}
\alias{codebook}
\alias{cb}
\title{Describe data contents}
\usage{
codebook(data)

cb(data)
}
\arguments{
\item{data}{A data.frame}
}
\description{
\Sexpr[results=rd]{lifecycle::badge("stable")}

\code{codebook()} displays a compact overview of the data frame.
It includes important characteristics of data contents including
number of variables, variable names, label, type and missing values.
}
\section{Examples}{
\preformatted{codebook(iris)
codebook(infert)
}
}

\seealso{
Other data management: 
\code{\link{append}()},
\code{\link{describe}()},
\code{\link{duplicates}()},
\code{\link{egen}()},
\code{\link{encode}()},
\code{\link{expand2}()},
\code{\link{formatDate}()},
\code{\link{generate}()},
\code{\link{label}()},
\code{\link{n_}()},
\code{\link{recode}()},
\code{\link{replace}()},
\code{\link{rowmiss}()}
}
\concept{data management}
