% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{codebook}
\alias{codebook}
\title{Describe data contents}
\usage{
codebook(data)
}
\arguments{
\item{data}{An object.}
}
\value{
\code{data} original \code{data} is (invisibly) returned,
allowing \code{codebook()} to be used within a data pipe line.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

\code{codebook} offers an overview of your data,
enabling easy examination of its structure; variables,
their names, labels, data types, missing counts, and unique value counts.

Apart from \link{data.frame}, it uses \link[pillar:glimpse]{pillar::glimpse} to inspect the
data structure.
}
\examples{
codebook(mtcars)
codebook(iris)

}
