% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{codebook}
\alias{codebook}
\title{Describe data contents}
\usage{
codebook(data, width = NULL)
}
\arguments{
\item{data}{A data frame.}

\item{width}{Width of output: defaults to the setting of the
\code{width} \link[=pillar_options]{option} (if finite)
or the width of the console.}
}
\value{
\code{data} original \code{data} is (invisibly) returned,
allowing \code{codebook()} to be used within a data pipe line.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

\code{codebook()} offers a concise overview of your data,
enabling easy examination of its structure, including variables,
labels, data types, missing counts, and unique value counts.
Simplify your data exploration with \code{codebook()}.
}
\examples{
codebook(mtcars)

\dontshow{if (rlang::is_installed("nycflights13")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
codebook(nycflights13::flights)
\dontshow{\}) # examplesIf}
}
