% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mStats.R
\name{codebook}
\alias{codebook}
\title{Describe the data}
\usage{
codebook(data)
}
\arguments{
\item{data}{data.frame}
}
\value{
a data.frame containing the codebook
}
\description{
\code{codebook()} examines the variable names, labels, and data
to produce a codebook for describing the dataset
}
\details{
It reports a description of the data with the following information.

\strong{ANNOTATIONS}:

\code{No} = serial number

\code{Variable} = variable name

\code{Label} = variable label

\code{Type} = type of variable

\code{Obs} = number of valid observations

\code{NA} = number of observations with missing value \code{NA}
}
\note{
For \code{haven_labelled} data.frame, data types are generated
using \code{typeof()}.
}
\examples{

codebook(infert)
codebook(iris)
codebook(mtcars)

}
\author{
Email: \email{dr.myominnoo@gmail.com}

Website: \url{https://myominnoo.github.io/}
}
