% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicates.R
\name{duplicates}
\alias{duplicates}
\alias{keepUnique}
\alias{keepDup}
\title{Tag, report, delete or keep duplicated observations}
\usage{
duplicates(data, ..., print.table = TRUE)

keepUnique(data, ..., print.table = TRUE)

keepDup(data, ..., print.table = TRUE)
}
\arguments{
\item{data}{dataframe}

\item{...}{any variables within dataframe for unique id}

\item{print.table}{logical value to display formatted outputs}
}
\value{
\code{data.frame}: original dataset plus \code{dupID_}

\code{formated_table}: display information about duplicated observations
}
\description{
\code{duplicates} reports numbers of duplicated observations
}
\details{
\code{duplicates} tags duplicate observations within dataframe
with a new variable called \code{dupID_} and reports statistics.
Duplicates are observations with identical values either on
all variables
if no variable is specified in the optional argument \code{...}
or on a specified list of variables.

\strong{ANNOTATIONS}:

\code{Copies}       - Number of duplicates

\code{Observations} - Number of records per Copies

\code{Surplus}  - Number of surplus copies

\code{dupID_} - indicates copies within the dataset. Example, 1 indicates
that there is no duplicates, 2 = two identical record, 3 = three records,
so on.

\code{keepUnique} delete all but the first occurrence of each group of
duplicated observations.

\code{keepDup} keep all but the first occurrence of each group of
duplicated observations. This function returns the opposite dataset
generated from \code{keepUnique}
}
\examples{
\dontrun{
# finding duplicates across all variables
iris_dup <- duplicates(iris)
str(iris_dup)

# finding duplicates on variables of interest
iris_dup <- duplicates(iris, Sepal.Length, Sepal.Width)
str(iris_dup)

iris_dup <- duplicates(iris, Species)
str(iris_dup)

iris_dup <- duplicates(iris, Sepal.Length, Sepal.Width, print.table = FALSE)
str(iris_dup)

# Keep Unique records
iris_dup <- keepUnique(iris, Sepal.Length)
str(iris_dup)

iris_dup <- keepUnique(iris, Species)
str(iris_dup)

iris_dup <- keepUnique(infert, case)
str(iris_dup)


# Keep duplicated records (opposite of keep unique records)
iris_dup <- keepDup(iris, Sepal.Length)
str(iris_dup)

iris_dup <- keepDup(iris, Species)
str(iris_dup)

iris_dup <- keepDup(infert, case)
str(iris_dup)
}
}
\seealso{
\code{\link{expand2}}, \code{\link{keep}}
}
\author{
For any feedback, please contact \code{Myo Minn Oo} via:

Email: \email{dr.myominnoo@gmail.com}

Website: \url{https://myominnoo.github.io/}
}
\keyword{distinct,}
\keyword{duplicates,}
\keyword{report,}
\keyword{unique}
