% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicates.R
\name{duplicates}
\alias{duplicates}
\title{Report, tag or drop duplicate observations}
\usage{
duplicates(data, ..., drop = FALSE, report = TRUE)
}
\arguments{
\item{data}{A data.frame}

\item{...}{One or more unquoted variables separated by commas.
Variable names can be used as if they were positions in the data
frame, so expressions like \code{x:y} can be used to select a range of
variables.}

\item{drop}{\code{TRUE} deletes all duplicate observations but the first
occurence of each group of duplicated observations. \code{FALSE} tags all duplicate
observations with a new variable \code{dup_}.}

\item{report}{\code{TRUE} reports a table showing the following:
\itemize{
\item Copies  - Number of observations that occur as one or more copies
\item Surplus - Indicates that they are the second (third, ...) copy of the
first of each group of duplicates.
}}
}
\description{
\Sexpr[results=rd]{lifecycle::badge("stable")}

\code{duplicates()} reports, tags or drops duplicate observations.
Duplicates are observations that have identical values.
They can be searched either on all variables if no \code{variable} is specified
or on specified \code{variables}.
}
\section{Examples}{


Here we find duplicate observations based on all variables\preformatted{iris1 <- duplicates(iris)
head(iris1)

## we remove all duplicates except the first observations.
iris2 <- duplicates(iris, drop = TRUE)
head(iris2)

## we turn off the report table
iris3 <- duplicates(iris, drop = TRUE, report = FALSE)
head(iris3)

}

Below we find duplicates based on \code{Species}.\preformatted{iris4 <- duplicates(iris, Species)
head(iris4)


## we remove all duplicates except the first observations.
iris5 <- duplicates(iris, Species, drop = TRUE)
head(iris5)


## we turn off the report table
iris6 <- duplicates(iris, Species, drop = TRUE, report = FALSE)
head(iris6)

}
}

\seealso{
Other data management: 
\code{\link{append}()},
\code{\link{codebook}()},
\code{\link{describe}()},
\code{\link{egen}()},
\code{\link{encode}()},
\code{\link{expand2}()},
\code{\link{formatDate}()},
\code{\link{generate}()},
\code{\link{label}()},
\code{\link{n_}()},
\code{\link{recode}()},
\code{\link{replace}()},
\code{\link{rowmiss}()}
}
\concept{data management}
