% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicates.R
\name{duplicates}
\alias{duplicates}
\alias{keepUnique}
\alias{keepNonUnique}
\title{Tag, report, delete or keep duplicate observations}
\usage{
duplicates(data, ...)

keepUnique(data, ...)

keepNonUnique(data, ...)
}
\arguments{
\item{data}{dataframe}

\item{...}{any variables within dataframe for unique id}
}
\description{
\code{duplicates} displays structure of a data frame
}
\details{
\code{duplicates}

tags duplicate observations within dataframe with a new variable called
\code{dupID_} and reports statistics.
Duplicates are observations with identical values either on all variables
if no variable is specified in the optional argument \code{...}
or on a specified list of variable

ANNOTATIONS:

Copies       - Number of duplicates
Observations - Count of duplicate observations
dupID Count  - Count of observations by dupID

\code{keepUnique}

delete all but the first occurrence of each group of duplicated observations.

\code{keepNonUnique}
keep all but the first occurrence of each group of duplicated observations.
This function returns the opposite dataset generated from \code{keepUnique}.
}
\examples{
\dontrun{
duplicates(infert)
duplicates(iris)

keepUnique(infert)
keepUnique(infert, education, age)
keepUnique(iris)

keepNonUnique(infert)
keepNonUnique(infert, education, age)
keepNonUnique(iris)
}
}
\seealso{
\code{\link{keep}}, \code{\link{lose}}
}
\author{
Myo Minn Oo (Email: \email{dr.myominnoo@gmail.com} |
Website: \url{https://myominnoo.github.io/})
}
\keyword{duplicates,}
\keyword{report}
