% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egen.R
\name{egen}
\alias{egen}
\title{Transform Numeric to Factor}
\usage{
egen(x, cut = NULL, labels = NULL, na.rm = FALSE)
}
\arguments{
\item{x}{numeric vector}

\item{cut}{either a single number or a numeric vector.}

\item{labels}{specify to name the factor levels.}

\item{na.rm}{A logical value to specify missing values}
}
\description{
\code{egen} allows to transform a numeric vector to a factor vector easily.
}
\details{
\code{egen} allows easy conversion of a numeric vector to factor.

\strong{Cut-off Intervals}

If a single number is specified,
the range of the data is divided into equal cut-off points by that number. The last
piece will be always the same or smaller than the rest. In case of vector, the cut-off
points will be determined accordingly.

\strong{Label Construction}
If none is specified, then the factor
level labels are constructed as \strong{variable name} plus \strong{cut-off intervals}.
}
\examples{
set.seed(1)
age <- round(c(rnorm(100, 45, 20), rep(NA, 20)),0)
summary(age)

egen(age)
egen(age, cut = 20)
egen(age, cut = c(1, 20, 40))
egen(age, cut = c(1, 20, 40), labels = c("young", "middle", "old"))
egen(age, cut = c(1, 20, 40, 60, 100))

## remove missing value
egen(age, cut = 20, na.rm = TRUE)
egen(age, cut = c(1, 20, 40), na.rm = TRUE)
egen(age, cut = c(1, 20, 40), labels = c("young", "middle", "old"), na.rm = TRUE)
egen(age, cut = c(1, 20, 40, 60, 100), na.rm = TRUE)
}
\seealso{
\code{\link{gen}}, \code{\link{label}}, \code{\link{dup.rm}},
}
\author{
Myo Minn Oo (Email: \email{dr.myominnoo@gmail.com} |
Website: \url{https://myominnoo.github.io/})
}
\keyword{correlation}
\keyword{distribution,}
\keyword{number}
\keyword{summary,}
