% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egen.R
\name{egen}
\alias{egen}
\title{An extension to \code{generate}}
\usage{
egen(data, var, cut = NULL, new_var = NULL, lbl = NULL)
}
\arguments{
\item{data}{Dataset}

\item{var}{Exisitng variable}

\item{cut}{either a single number or a numeric vector.}

\item{new_var}{name of new variable to be generated}

\item{lbl}{labels to specify}
}
\description{
\code{egen()} transforms a numeric vector to a factor vector.
}
\details{
\code{egen} allows easy conversion of a numerical variable to a categorical
variable.

\strong{Cut-off Intervals}

If the interval is not specified, it is cut at an interval of 10,
starting from the minimum value. Otherwise,
it is divided into corresponding intervals by specified cut-off points.

\strong{Automatic Labelling}

If \code{lbl} is not specified, labels are constructed in
this format: \code{lbl[##-##]}
}
\examples{
\dontrun{



# Example from IDRE UCLA
path <- "https://stats.idre.ucla.edu/stat/data/patient_pt1_stata_dm.dta"
hosp <- haven::read_dta(path)
codebook(hosp)


## to use piping function
library(magrittr)


## generate bmi.cat without cut points
egen(hosp, bmi) \%>\%
    tab(bmi.cat)


## generate bmi.cat with cut points
egen(hosp, bmi, c(0, 15, 16, 18.5, 25, 30, 35, 40, 500)) \%>\%
    tab(bmi.cat)


## generate bmi.cat with cut points
egen(hosp, bmi, c(0, 15, 16, 18.5, 25, 30, 35, 40, 500),
     lbl = c("<18.5", "18.5-24", "25-29", "30-34", "35-39", "40+"),
     new_var = BMI) \%>\%
    tab(BMI)

}

}
\author{
For any feedback, please contact \code{Myo Minn Oo} via:

Email: \email{dr.myominnoo@gmail.com}

Website: \url{https://myominnoo.github.io/}
}
\concept{new variable generate produce create transform conversion transform}
