% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egen.R
\name{egen}
\alias{egen}
\title{Convert a continuous variable into groups}
\usage{
egen(data, var, cut = NULL, label = NULL, new_var = NULL)
}
\arguments{
\item{data}{A data.frame}

\item{var}{Name of the new variable}

\item{cut}{either a number or a numeric vector}

\item{label}{Labels for the groups}

\item{new_var}{Name of the new variable}
}
\description{
\Sexpr[results=rd]{lifecycle::badge("stable")}

\code{egen()} categorize a numerical variable into a factor.
If only a number is specified in \code{cut}, it categorizes
into equal intervals based on that number. If no value is set
for \code{cut}, the default interval is \code{10}.

If \code{new_var} is not specified, a new name is automatically
assigned by appending \verb{_cat} to the original name: \code{VARNAME_cat}
If \code{label} is not specified, a set of labels are automatically
constructed in the format, \verb{##-##}.
}
\section{Examples}{


Here is how we can use \code{egen}.\preformatted{infert <- egen(infert, age, c(26, 31, 36, 41))
table(infert$age_cat)


## Add labels and give a new name
infert <- egen(infert, age, cut = c(26, 31, 36, 41),
               label = c("<=25", "26-30", "31-35", "36-40", "41+"),
               new_var = "age_grp")
table(infert$age_grp)
}
}

\seealso{
Other data management: 
\code{\link{append}()},
\code{\link{codebook}()},
\code{\link{describe}()},
\code{\link{duplicates}()},
\code{\link{encode}()},
\code{\link{expand2}()},
\code{\link{formatDate}()},
\code{\link{generate}()},
\code{\link{label}()},
\code{\link{n_}()},
\code{\link{recode}()},
\code{\link{replace}()},
\code{\link{rowmiss}()}
}
\concept{data management}
