% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egen.R
\name{egen}
\alias{egen}
\alias{egen.numeric}
\alias{egen.data.frame}
\title{An extension to \code{generate}}
\usage{
egen(
  data = NULL,
  var_old,
  cut = NULL,
  var_new = NULL,
  lbl = NULL,
  na.rm = FALSE,
  print.notes = TRUE
)

\method{egen}{numeric}(
  data = NULL,
  var_old,
  cut = NULL,
  var_new = NULL,
  lbl = NULL,
  na.rm = FALSE,
  print.notes = TRUE
)

\method{egen}{data.frame}(
  data = NULL,
  var_old,
  cut = NULL,
  var_new = NULL,
  lbl = NULL,
  na.rm = FALSE,
  print.notes = TRUE
)
}
\arguments{
\item{data}{dataset}

\item{var_old}{variable to perform cut}

\item{cut}{either a single number or a numeric vector.}

\item{var_new}{name of new variable generated}

\item{lbl}{specify to name the factor levels.}

\item{na.rm}{A logical value to specify missing values}

\item{print.notes}{logical value to indicate whether notes will be
printed}
}
\description{
\code{egen()} transforms a numeric vector to a factor vector.
}
\details{
\code{egen} allows easy conversion of a numerical variable to a categorical
variable.

\strong{Cut-off Intervals}


If the interval is not specified, it is cut at an interval of 10,
starting from the minimum value. Otherwise,
it is divided into corresponding intervals by specified cut-off points.

\strong{Automatic Labelling}


If \code{lbl} is not specified, labels are constructed in
this format: \code{lbl[##-##]}
}
\examples{
\dontrun{
## variable with dataframe
# automatically categorized into interval of 10
infert.new <- egen(infert, age)
str(infert.new)

infert.new <- egen(infert, age, c(31, 36, 41))
str(infert.new)

# give variable name as age.grp
infert.new <- egen(infert, age, c(31, 36, 41), age.grp,
                c("young", "mid", "old", "oldest"))
str(infert.new)

# age.grp duplicates
infert.new <- egen(infert.new, age, c(31, 36, 41), age.grp,
                c("young", "mid", "old", "oldest"))
str(infert.new)

# Example dataset from IDRE website
path <- "https://stats.idre.ucla.edu/stat/data/patient_pt1_stata_dm.dta"
hosp <- haven::read_dta(path)
hosp <- egen(hosp, bmi, c(16, 18.5, 25, 30, 35, 40))
str(hosp)
}
}
\seealso{
\code{\link{generate}}, \code{\link{duplicates}}, \code{\link{listView}}
}
\author{
For any feedback, please contact \code{Myo Minn Oo} via:

Email: \email{dr.myominnoo@gmail.com}

Website: \url{https://myominnoo.github.io/}
}
\keyword{conversion,}
\keyword{create,}
\keyword{generate,}
\keyword{new}
\keyword{produce,}
\keyword{transform}
\keyword{transform,}
\keyword{variable,}
