% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egen.R
\name{egen}
\alias{egen}
\title{extension to gen function}
\usage{
egen(x, cut = NULL, labels = NULL, na.rm = FALSE)
}
\arguments{
\item{x}{numeric vector}

\item{cut}{either a single number or a numeric vector.}

\item{labels}{specify to name the factor levels.}

\item{na.rm}{A logical value to specify missing values}
}
\description{
\code{egen} transforms a numeric vector to a factor vector.
}
\details{
\code{egen} allows easy conversion of a numeric vector to factor.

\strong{Cut-off Intervals}
If the interval is not specified, it is cut at an interval of 10. Otherwise,
it is divided into equal cut-off points by specified number.

\strong{Labelling}
If not specified, the labels are constructed in the format:
\strong{variable name} + "." + \strong{"cut-off intervals"}.
}
\examples{
set.seed(1)
age <- round(c(rnorm(100, 45, 20), rep(NA, 20)),0)
summary(age)

egen(age)
egen(age, cut = 20)
egen(age, cut = c(1, 20, 40))
egen(age, cut = c(1, 20, 40), labels = c("young", "middle", "old"))
egen(age, cut = c(1, 20, 40, 60, 100))

## remove missing value
egen(age, cut = 20, na.rm = TRUE)
egen(age, cut = c(1, 20, 40), na.rm = TRUE)
egen(age, cut = c(1, 20, 40), labels = c("young", "middle", "old"), na.rm = TRUE)
egen(age, cut = c(1, 20, 40, 60, 100), na.rm = TRUE)
}
\seealso{
\code{\link{generate}}
}
\author{
Myo Minn Oo (Email: \email{dr.myominnoo@gmail.com} |
Website: \url{https://myominnoo.github.io/})
}
\keyword{correlation}
\keyword{distribution,}
\keyword{number}
\keyword{summary,}
