% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egen.R
\name{egen}
\alias{egen}
\title{Convert a continuous variable into groups}
\usage{
egen(data, var, at = NULL, label = NULL, new_var = NULL, ...)
}
\arguments{
\item{data}{data.frame}

\item{var}{existing variable}

\item{at}{either a number or a numeric vector}

\item{label}{Labels for the groups}

\item{new_var}{Name of the new variable}

\item{...}{Additional arguments to be passed to \code{\link{cut}}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function was deprecated because we realized that it's
a special case of the \code{\link{cut}} function.
}
\examples{
data <- data.frame(x = 1:10)
egen(data, x, at = c(3, 7), label = c("low", "medium", "high"))
egen(data, x, at = c(3, 7), label = c("low", "medium", "high"), new_var = "group")


}
\seealso{
\code{\link{cut}}
}
\concept{data manipulation}
\keyword{deprecated}
