% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode_decode.R
\name{encode}
\alias{encode}
\title{Encode variables in a data frame}
\usage{
encode(.data, ...)
}
\arguments{
\item{.data}{The input data frame.}

\item{...}{The variables to be encoded.}
}
\value{
A modified data frame with the selected variables encoded.
}
\description{
The \code{encode()} function encodes selected variables in a data frame. It converts the variables to numeric type, sets labels from the original data, and returns the modified data frame.
}
\examples{
# Encode selected variables in the `mtcars` data frame
encode(mtcars, mpg, cyl)

}
