% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand2.R
\name{expand2}
\alias{expand2}
\title{Duplicate observations within a dataframe}
\usage{
expand2(data, n_n = NULL, copies = 2, original = TRUE)
}
\arguments{
\item{data}{a data frame object}

\item{n_n}{index or indexes specifying row numbers}

\item{copies}{desired number of copies}

\item{original}{a logical indicating whether to keep the original dataframe}
}
\value{
\code{data.frame}
}
\description{
\code{expand2} generates duplicated observations within a dataframe.
}
\details{
\code{expand2} appends observations from the dataframe
with n copies of the observations with
specified indexes of observations or all data.
}
\examples{
\dontrun{
# using IDRE UCLA Example
path <- "https://stats.idre.ucla.edu/stat/stata/notes/hsb2.dta"
hsb2 <- haven::read_dta(path)
codebook(hsb2)


## use pipe function
library(magrittr)


## keep id female ses math read write
hsb2 <- hsb2 \%>\%
    keep(id, female, ses, read:math)


# copy the first 5 rows for 4 times
hsb2 \%>\%
     expand2(1:5, 4) \%>\%
     duplicates \%>\%
     codebook

# copy the first 5 rows for 4 times and keep only duplicated records
hsb2 \%>\%
     expand2(1:5, 4, original = FALSE) \%>\%
     codebook
}
}
\references{
HOW CAN I DETECT DUPLICATE OBSERVATIONS? | STATA FAQ.
UCLA: Statistical Consulting Group.
from https://stats.idre.ucla.edu/stata/faq/how-can-i-detect-duplicate-observations-3/
(accessed September 27, 2019).
}
\seealso{
\code{\link{generate}}, \code{\link{egen}}
}
\author{
For any feedback, please contact \code{Myo Minn Oo} via:

Email: \email{dr.myominnoo@gmail.com}

Website: \url{https://myominnoo.github.io/}
}
\concept{duplicates observations expand duplicate rows}
