% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand2.R
\name{expand2}
\alias{expand2}
\title{Duplicate observations within a dataframe}
\usage{
expand2(data, n_n = NULL, copies = 2, original = TRUE)
}
\arguments{
\item{data}{a data frame object}

\item{n_n}{index or indexes specifying row numbers}

\item{copies}{desired number of copies}

\item{original}{a logical indicating whether to keep the original dataframe}
}
\description{
\code{expand2} generates duplicated observations within a dataframe.
}
\details{
\code{expand2}

appends observations from the dataframe with n copies of the observations with
specified indexes of observations or all data.
}
\examples{
\dontrun{

#### Ref:
# HOW CAN I DETECT DUPLICATE OBSERVATIONS? | STATA FAQ
# from https://stats.idre.ucla.edu/stata/faq/how-can-i-detect-duplicate-observations-3/
# (accessed September 27, 2019).

required(haven)
# install.packages("haven")
hsb2 <- read_dta("https://stats.idre.ucla.edu/stat/stata/notes/hsb2.dta")
str(hsb2)
hsb2.new <- expand2(hsb2, 1:5, 4)
hsb2.new <- expand2(hsb2, 1:5, 4, original = FALSE)
}
}
\seealso{
\code{\link{keep}}, \code{\link{generate}}, \code{\link{egen}}
}
\author{
Myo Minn Oo (Email: \email{dr.myominnoo@gmail.com} |
Website: \url{https://myominnoo.github.io/})
}
\keyword{duplicate}
\keyword{duplicates,}
\keyword{expand,}
\keyword{observations,}
\keyword{rows}
