% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mStats.R
\name{expand2}
\alias{expand2}
\title{Duplicate observations within a dataframe}
\usage{
expand2(data, n_n = NULL, copies = 2, original = TRUE)
}
\arguments{
\item{data}{a data frame object}

\item{n_n}{index or indexes specifying row numbers}

\item{copies}{desired number of copies}

\item{original}{a logical indicating whether to keep the original dataframe}
}
\value{
data.frame
}
\description{
\code{expand2} generates duplicated observations within a dataframe.
}
\details{
\code{expand2} appends observations from the dataframe
with n copies of the observations with
specified indexes of observations or all data.
}
\examples{

## create duplicates
x <- expand2(infert, 1:5, copies = 2)

## check duplicates report and rmeove dup
duplicates(x, drop = TRUE)

}
\author{
Email: \email{dr.myominnoo@gmail.com}

Website: \url{https://myominnoo.github.io/}
}
