% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expandTables.R
\name{expandTables}
\alias{expandTables}
\alias{expandFreq}
\title{Convert \code{2x2 tables} to \code{data.frame}}
\usage{
expandTables(
  ...,
  exp_name = "exp",
  exp_lvl = c("exposed", "unexposed"),
  case_name = "case",
  case_lvl = c("cases", "controls"),
  strata_name = "strata"
)

expandFreq(data, freq)
}
\arguments{
\item{...}{vectors of 2x2 tables}

\item{exp_name}{Name of \code{exp} Variable}

\item{exp_lvl}{Names of two categories in the order of
Exposed and non-exposed}

\item{case_name}{Name of \code{Case} variable}

\item{case_lvl}{names of two categories in the order of}

\item{strata_name}{Name of stratified variable}

\item{data}{frequency table in dataset format}

\item{freq}{variable for weighted frequency}
}
\value{
\code{data.frame}
}
\description{
\code{expandTables()} generates a data frame and supports two levels.
}
\details{
\strong{expandTables}

uses the vectors of \code{2x2} tables and
generates a data frame of at least two columns:
exp and case.

\preformatted{expandTables(c(100, 200, 100, 200))}

\code{Strata}

Multiple tables can be used to construct a dataset by specifying
\code{strata_name} as follow. Strata can be included
using multiple named vectors.

\preformatted{
expandTables(
             strata1 = c(100, 200, 100, 200),
             strata2 = c(100, 200, 100, 200),
             strata3 = c(100, 200, 100, 200),
             exp_name = "exp",
             exp_lvl = c("exposed", "unexposed"),
             case_name = "case",
             case_lvl = c("case", "control"),
             strata_name = "Strata"
)
}

\code{Labels for variables}

If names or lavels of variables are not specified, the followings are
applied.

\enumerate{
\item exp Name: \code{exp}
\item exp levels: \code{exposed} and \code{unexposed}
\item case Name: \code{case}
\item case levels: \code{cases} and \code{controls}
\item Strata Name: \code{strata}
\item Note: Strata levels are not considered as vectors must
be named.
}

\strong{expandFreq()} uses the data.frame and construct the dataset
based on the
frequency weight. The frequency weight must be specified by
\code{freq} argument.

\preformatted{
expandFreq(data = "data_name", freq = "freq")
}
}
\section{Functions}{
\itemize{
\item \code{expandFreq}: \code{expandFreq()} expands a frequency-weighted table into a dataset.
The table must be in dataset format.
It preserves the original dataset format.
}}

\examples{


## Asthma Example from Essential Medical Statistics
## page 160

asthma <- expandTables(c(81, 995, 57, 867),
              exp_name = "sex",
              exp_lvl = c("woman", "man"),
              case_name = "asthma",
              case_lvl = c("yes", "no"))

## label variable and dataset
asthma <- labelData(asthma, "Hypothetical Data of Asthma Prevalence")
asthma <- labelVar(asthma, sex = "Man or Woman",
                           asthma = "Asthma or No Asthma")

## Checking codebook
codebook(asthma)


## simple tabulation
tab(asthma)

## cross-tabulation
tab(asthma, sex, by = asthma)



## Example for expandFreq  <<<<<--------->>>>>

## Example from UCLA website
## you can download the dataset here:
## https://stats.idre.ucla.edu/stat/stata/examples/icda/afterlife.dta

al <- data.frame(gender = c(1, 1, 0, 0),
                 aftlife = c(1, 0, 1, 0),
                 freq = c(435, 147, 375, 134))
al.exp <- expandFreq(al, freq)

## check the numbers by tabulation
## tab(al.exp, gender, by = aftlife)

}
\author{
Email: \email{dr.myominnoo@gmail.com}

Website: \url{https://myominnoo.github.io/}
}
