% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export}
\alias{export}
\title{Export outputs}
\usage{
export(.data, file = "", append = FALSE)
}
\arguments{
\item{.data}{Outputs as \code{data.frame} or \code{list}}

\item{file}{file name}

\item{append}{logical.

If \code{TRUE}, the output is appended to the file.

If \code{FALSE}, any existing file of the name is destroyed.}
}
\description{
\code{export()} writes to \code{.csv} file
}
\details{
Outputs from the following functions can be exported.

\code{\link{tab}}

\code{\link{summ}}

\code{\link{tabOdds}}

\code{\link{tabRisks}}

\code{\link{mhor}}

\code{\link{mhrr}}

\code{\link{logistic}}

\code{\link{regress}}

\code{\link{strate}}
}
\author{
For any feedback, please contact \code{Myo Minn Oo} via:

Email: \email{dr.myominnoo@gmail.com}

Website: \url{https://myominnoo.github.io/}
}
