% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatDate.R
\name{formatDate}
\alias{formatDate}
\alias{is.Date}
\title{Coerces or test vectors to Date type}
\usage{
formatDate(x, format = "dmY", sep = "/", thisCent = NULL)

is.Date(x)
}
\arguments{
\item{x}{object to be coerced or tested}

\item{format}{for character vectors:}

\item{sep}{separator character for date components}

\item{thisCent}{specify either 2000 or 1900 for two-digit years}
}
\description{
\code{formatDate} coerces objects to Date type.
\code{is.Date} indicates which vector is type of Date.
}
\details{
Intuitively, "dmY" represents dd mm YYYY format.
In combination with value from \code{sep}, this can change to several date formats.
For example, "dmy" + "-" convert to "dd-mm-YYYY" format.

See \code{\link{as.Date}} for more date formats.
}
\examples{
dates <- c("2019-01-15", "2019-01-20", "2019-01-21", "2019-01-22")
is.Date(dates)
formatDate(dates, "Ymd", "-")
y <- formatDate(dates, "Ymd", "-")
is.Date(y)
y
# dates in numeric format MS excel
x <- 42705:42710
y <- formatDate(x)
y
}
\seealso{
\code{\link{dateSum}}
}
\author{
Myo Minn Oo (Email: \email{dr.myominnoo@gmail.com} |
Website: \url{https://myominnoo.github.io/})
}
\keyword{coerces,}
\keyword{date}
\keyword{test}
