% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatDate.R
\name{formatDate}
\alias{formatDate}
\alias{is.Date}
\title{Convert to Dates}
\usage{
formatDate(x, format = "dmY", sep = "/", century = NULL)

is.Date(x)
}
\arguments{
\item{x}{a character or numeric object}

\item{format}{only for character vectors:}

\item{sep}{separator character for date components}

\item{century}{specify either 2000 or 1900 for two-digit years}
}
\description{
\code{formatDate} converts characters or numbers to dates.
\code{is.Date} indicates which elements are Dates.
}
\details{
"dmY" represents dd mm YYYY format.
In combination with separators from \code{sep}, this can change to
several date formats.
For example, "dmy" + "-" convert to "dd-mm-YYYY" format.

See \code{\link{as.Date}} for more date formats.
}
\examples{
dates <- c("2019-01-15", "2019-01-20", "2019-01-21", "2019-01-22")
is.Date(dates)
formatDate(dates, "Ymd", "-")
y <- formatDate(dates, "Ymd", "-")
is.Date(y)
y
# dates in numeric format MS excel
x <- 42705:42710
y <- formatDate(x)
y
}
\seealso{
\code{\link{generate}}, \code{\link{egen}}, \code{\link{labelVars}}
}
\author{
Myo Minn Oo (Email: \email{dr.myominnoo@gmail.com} |
Website: \url{https://myominnoo.github.io/})
}
\keyword{conversion}
\keyword{date}
