% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate}
\alias{generate}
\title{Create new variable}
\usage{
generate(data, var, value = NA)
}
\arguments{
\item{data}{A data.frame}

\item{var}{Name of the new variable}

\item{value}{Content of the new variable}
}
\description{
\Sexpr[results=rd]{lifecycle::badge("stable")}

\code{generate()} creates a new variable.  The contents of the variable are
specified by \code{value}. If \code{value} is not specified, it is
set to \code{NA}. The \code{value} is attached as \link{label} of the \code{new} variable.
}
\section{Examples}{


Here we generate a new variable \code{age_cat}.
The value is derived from age using \link{ifelse}.\preformatted{infert <- generate(infert, age_cat, ifelse(age >= 40, "40+", "<40"))
codebook(infert)
}
}

\seealso{
Other data management: 
\code{\link{append}()},
\code{\link{codebook}()},
\code{\link{describe}()},
\code{\link{duplicates}()},
\code{\link{egen}()},
\code{\link{encode}()},
\code{\link{expand2}()},
\code{\link{formatDate}()},
\code{\link{label}()},
\code{\link{n_}()},
\code{\link{recode}()},
\code{\link{replace}()},
\code{\link{rowmiss}()}
}
\concept{data management}
