% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate}
\alias{generate}
\title{Create a new variable}
\usage{
generate(data = NULL, var, expr = NULL)
}
\arguments{
\item{data}{data frame}

\item{var}{name of a new variable}

\item{expr}{value or Expression: See examples below.}
}
\description{
\code{generate} easily create a new variable within the data frame
}
\details{
If data is specified, it returns the whole dataframe with recoded variables.

The value of the variable are specified by \code{expr} argument.
If \code{expr} is NULL, \code{generate} produce a vector
\code{NA}.
}
\examples{
\dontrun{
infert.new <- generate(infert, age)
head(infert.new)

infert.new <- generate(infert, age.cat)
head(infert.new)

infert.new <- generate(infert, age.cat,
         ifelse(age < 20, 1,
                ifelse(age >= 20 & age < 30, 2,
                       ifelse(age >= 30 & age < 40, 3, 4))))
head(infert.new)

# based on vector
age <- infert$age
age.cat <- generate(NULL, age.cat, ifelse(age < 20, 1,
                 ifelse(age >= 20 & age < 30, 2,
                    ifelse(age >= 30 & age < 40, 3, 4))))
age.cat
}
}
\seealso{
\code{\link{recode}}
}
\author{
Myo Minn Oo (Email: \email{dr.myominnoo@gmail.com} |
Website: \url{https://myominnoo.github.io/})
}
\keyword{generate,}
\keyword{new}
\keyword{produce}
\keyword{variable,}
