% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate}
\alias{generate}
\title{Create a new variable}
\usage{
generate(data, var, .expr = NULL)
}
\arguments{
\item{data}{Dataset}

\item{var}{New variable}

\item{.expr}{value or Expression for simple arithmetic or logical operations:
See examples.}
}
\description{
\code{generate()} creates a new variable within the data frame
}
\details{
If the data is specified, it returns modified dataset with
newly created variable.
The value of the variable are specified by \code{expr} argument.
}
\examples{
\dontrun{


# Example from IDRE UCLA
path <- "https://stats.idre.ucla.edu/stat/data/patient_pt1_stata_dm.dta"
hosp <- haven::read_dta(path)
codebook(hosp)


## to use piping function
library(magrittr)

## generate variable with NA
generate(hosp, average, NA) \%>\%
    keep(average)


## generate variable with other variables
generate(hosp, average, pain) \%>\%
    keep(pain, average)

generate(hosp, average, (test1 + test2) / 2) \%>\%
    keep(test1, test2, average)


## existing variable cann't be generated.
generate(hosp, pain)
}

}
\author{
For any feedback, please contact \code{Myo Minn Oo} via:

Email: \email{dr.myominnoo@gmail.com}

Website: \url{https://myominnoo.github.io/}
}
