% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mStats.R
\name{helpers}
\alias{helpers}
\alias{clear}
\title{Helper functions}
\usage{
helpers(...)

clear()
}
\arguments{
\item{...}{further arguments to be passed to or from methods}
}
\description{
These are helper functions for \code{mStats}.
}
