% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibarplot.R
\name{ibarplot}
\alias{ibarplot}
\title{Delightful Bar Plots}
\usage{
ibarplot(x, y = NULL, by = NULL, data = NULL, rnd = 1,
  na.rm = FALSE, main = NULL, xlab = NULL, ylab = NULL,
  text.size = 12, label.pct = FALSE, flip.axis = FALSE,
  legend.show = TRUE, legend.text = NULL, plot.type = "p",
  facet.ncol = 2, plot.save = FALSE, plot.name = "<unnamed>.tiff",
  width = 5, height = 4, dpi = 150)
}
\arguments{
\item{x}{a factor object}

\item{y}{a factor object; if ignored, frequency distribution on x is generated.}

\item{by}{a factor object; if ignored, cross-tabulation on x and y is generated.}

\item{data}{an optional data frame}

\item{rnd}{specify rounding of numbers. See \code{\link{round}}.}

\item{na.rm}{A logical value to specify missing values, <NA>:}

\item{main}{title of the plot}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{text.size}{size of text in the plot. Font size for data level is 14:5 ratio 
to axis font size}

\item{label.pct}{show percentage for data label}

\item{flip.axis}{logical value to flip x-axis to y-axis}

\item{legend.show}{show or hide the legend}

\item{legend.text}{title of legend}

\item{plot.type}{Only for bivariate analysis: four type of plot can be specified. 
\enumerate{
  \item \code{p}: parallel bar plot (default)
  \item \code{f}: \code{x} bar plot faceted by \code{y}
  \item \code{s}: stacked bar plot
  \item \code{fs}: full stacked or percentage bar plot
}}

\item{facet.ncol}{number of columns to be faceted in faceted plot}

\item{plot.save}{save the plot}

\item{plot.name}{specify for name of the plot file to be saved. 
Filename can be in either of these extension formats: ".tiff", ".png", and ".pdf"}

\item{width}{integer measured in inches}

\item{height}{integer measured in inches}

\item{dpi}{specify resolution of the plot to be saved}
}
\description{
\code{ibarplot} generates optimized bar plots for univariate, bivariate and 
stratified analyses.
}
\details{
Exploring data before jumping into complex analysis is always a necessity. 
The first step of an analysis should be to summarize and display data.
By doing so, invaluable insight can be gained through the familiarity with the data.

\strong{Bar Plot}

Frequency data is displayed in the form of horizontal or vertical bars. This type 
of plot is used to display single variable or the association between two or three 
categorical variables.

\strong{References:}
\enumerate{
  \item An Introduction to MEdical Statistics, Martin Bland, Thrid Edition, 
  Chapter 4, page 73
}
}
\examples{
data(infert)

## univariate display
ibarplot(infert$education)
ibarplot(education, data = infert)
ibarplot(education, data = infert, main = "Plot of eduation", 
         xlab = "Education Level", ylab = "Number")

ibarplot(education, data = infert, label.pct = TRUE)

## bivariate display
ibarplot(education, case, data = infert) # parallel bar chart
ibarplot(education, case, data = infert, plot.type = 'f') # facet plot
ibarplot(education, case, data = infert, plot.type = 's') # stacked plot
ibarplot(education, case, data = infert, plot.type = 'fs') # full stacked

## stratified display
ibarplot(education, case, parity, infert)
ibarplot(education, parity, case, infert) 
}
\seealso{
\code{\link{itab}}, \code{\link{isum}}, \code{\link{iboxplot}}
}
\author{
Myo Minn Oo (Email: \email{dr.myominnoo@gmail.com} |
Website: \url{https://myominnoo.github.io/})
}
\keyword{bar}
\keyword{bar,}
\keyword{chart,}
\keyword{faceted}
\keyword{full}
\keyword{plot,}
\keyword{stacked}
