% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibarplot.R
\name{ibarplot}
\alias{ibarplot}
\title{Create and Save beautiful Bar Plot using ggplot2}
\usage{
ibarplot(x, y = NULL, by = NULL, data = NULL, rnd = 1,
  na.rm = FALSE, main = NULL, xlab = NULL, ylab = NULL,
  show.legend = TRUE, legend.text = NULL, plot.type = "f",
  facet.ncol = 2, save.plot = FALSE, plot.name = "ibarplot.tiff",
  width = 5, height = 4, dpi = 150)
}
\arguments{
\item{x}{a vector describing the bars which make up the plot. It is usually on x axis.}

\item{y}{a vector describing the y axis or second variable in cross-tabulation or data relationship.}

\item{by}{a vector describing the grouping of x. By default, the plot generates a faceted barplot.}

\item{data}{an optional data frame (or object coercible by as.data.frame to a data frame) containing the variables for contigency table.}

\item{rnd}{an integer indicating the number of decimal places:}

\item{na.rm}{A logical value indicating to remove NA values in the table or not. By Default, the value is TRUE}

\item{main}{a main title for the plot.}

\item{xlab}{a label for the x axis, defaults to a description of x.}

\item{ylab}{a label for the y axis, defaults to a description of y.}

\item{show.legend}{show or hide the legend. Hide the legend by default}

\item{legend.text}{Legend title}

\item{plot.type}{if bivariate analysis, type of plot can be specified. By default, this generates a facted plot. 'p' for parrallel barplot, 's' for stacked barplot, 'fs' for full stacked percentage barplot.}

\item{facet.ncol}{number of columns to be faceted}

\item{save.plot}{a logical value. If TRUE, it saves the plot generated in the current working directory.}

\item{plot.name}{a text for plot filename. Suffix can be ".png", ".tiff" and ".pdf"}

\item{width}{a value in inches}

\item{height}{a value in inches}

\item{dpi}{a value for resolution of the plot saved.}
}
\description{
ibarplot is a function that generate a barplot with several options for optimal data visualization of factor variables using ggplot2 package. If one factor is provided as input, a simple barplot will be generated. If two factors, it produces a faceted barplot by default. Grouped bar plot can also be specified. Several parameters including plot title, labels of x and y axes and saving plot can be set by the user.
}
\examples{
ibarplot(infert$education)
ibarplot(x = education, data = infert)
ibarplot(x = education, y = case, data = infert)
ibarplot(x = education, y = case, data = infert, plot.type = 'f')
ibarplot(x = education, y = case, data = infert, plot.type = 'p')
ibarplot(x = education, y = case, data = infert, plot.type = 's')
ibarplot(x = education, y = case, data = infert, plot.type = 'fs')
education <- factor(rep(c("None", "Primary", "Secondary", "graduate"), c(50, 100, 25, 25)))
sex <- factor(rep(c("M", "F", NA), c(100, 75, 25)))
marital <- factor(rep(c("Single", "married", "widowed", NA), c(25, 100, 70, 5)))
ibarplot(education)
ibarplot(sex)
ibarplot(marital)
ibarplot(sex, marital)
ibarplot(sex, marital, na.rm = TRUE)
ibarplot(sex, marital, education)
}
\seealso{
iboxplot, itab, inumsum, isum
}
\keyword{barplot,}
\keyword{cross-tabulation,}
\keyword{faceted}
\keyword{plot}
