% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibarplot.R
\name{ibarplot}
\alias{ibarplot}
\title{Create and Save beautiful Bar Plot using ggplot2}
\usage{
ibarplot(x, by = NULL, data = NULL, rnd = 1, na.rm = FALSE,
  main = NULL, xlab = NULL, ylab = NULL, facet = TRUE,
  legend.text = NULL, legend.position = "right", bar.color = "white",
  vjust = -0.5, hjust = 0.5, position_dodge = 1, save.plot = FALSE,
  plot.name = "ibarplot.tiff", width = 5, height = 4, dpi = 150)
}
\arguments{
\item{x}{a vector describing the bars which make up the plot. It is usually on x axis.}

\item{by}{a vector describing the grouping of x. By default, the plot generates a faceted barplot.}

\item{data}{an optional data frame (or object coercible by as.data.frame to a data frame) containing the variables for contigency table.}

\item{rnd}{an integer indicating the number of decimal places:}

\item{na.rm}{A logical value indicating to remove NA values in the table or not. By Default, the value is TRUE}

\item{main}{a main title for the plot.}

\item{xlab}{a label for the x axis, defaults to a description of x.}

\item{ylab}{a label for the y axis, defaults to a description of y.}

\item{facet}{a logical value. If TRUE, the plot forms a matrix of panels defined by column faceting variables. It is most useful when you have two discrete variables, and all combinations of the variables exist in the data. By default, it is set to TRUE.}

\item{legend.text}{a text for the title of the plot}

\item{legend.position}{the position of legends ("none", "left", "right", "bottom", "top", or two-element numeric vector)}

\item{bar.color}{a text to depict the outer line of bars. You can set "black", "green", "blue" and so on.}

\item{vjust}{a number to place the value label in percentage to the bar.}

\item{hjust}{a number to place the value label in percentage to the bar.}

\item{save.plot}{a logical value. If TRUE, it saves the plot generated in the current working directory.}

\item{plot.name}{a text for plot filename. Suffix can be ".png", ".tiff" and ".pdf"}

\item{width}{a value in inches}

\item{height}{a value in inches}

\item{dpi}{a value for resolution of the plot saved.}
}
\description{
ibarplot is a function that generate a barplot with several options for optimal data visualization of factor variables using ggplot2 package. If one factor is provided as input, a simple barplot will be generated. If two factors, it produces a faceted barplot by default. Grouped bar plot can also be specified. Several parameters including plot title, labels of x and y axes and saving plot can be set by the user.
}
\examples{
ibarplot(x = education, data = infert)
ibarplot(x = education, by = case, data = infert)
ibarplot(x = education, by = case, data = infert, facet = FALSE)
ibarplot(induced, case, infert, facet = FALSE)
}
