% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iboxplot.R
\name{iboxplot}
\alias{iboxplot}
\title{Create and Save beautiful Box Plot using ggplot2}
\usage{
iboxplot(x, y = NULL, by = NULL, data = NULL, rnd = 1,
  na.rm = FALSE, main = NULL, xlab = NULL, ylab = NULL,
  show.legend = TRUE, legend.text = NULL, plot.save = FALSE,
  plot.name = "iboxplot.tiff", width = 5, height = 4, dpi = 150)
}
\arguments{
\item{x}{a vector describing the bars which make up the plot. It is usually on x axis.}

\item{y}{a vector describing the y axis or second variable in cross-tabulation or data relationship.}

\item{by}{a vector describing the grouping of x. By default, the plot generates a faceted barplot.}

\item{data}{an optional data frame (or object coercible by as.data.frame to a data frame) containing the variables for contigency table.}

\item{rnd}{an integer indicating the number of decimal places:}

\item{na.rm}{A logical value indicating to remove NA values in the table or not. By Default, the value is TRUE}

\item{main}{a main title for the plot.}

\item{xlab}{a label for the x axis, defaults to a description of x.}

\item{ylab}{a label for the y axis, defaults to a description of y.}

\item{show.legend}{show or hide the legend. Hide the legend by default}

\item{legend.text}{a text for the title of the plot}

\item{plot.save}{a logical value. If TRUE, it saves the plot generated in the current working directory.}

\item{plot.name}{a text for plot filename. Suffix can be ".png", ".tiff" and ".pdf"}

\item{width}{a value in inches}

\item{height}{a value in inches}

\item{dpi}{a value for resolution of the plot saved.}

\item{legend.text}{Legend title}
}
\description{
generate a boxplot for data exploration and visualization using ggplot2 package.
}
\examples{
iboxplot(infert$age)
iboxplot(age, data = infert)
iboxplot(age, education, data = infert)
iboxplot(infert$age, infert$education)
iboxplot(age, pooled.stratum, education, data = infert)
infert$case <- factor(infert$case)
# NA is included as one of the categories
iboxplot(age, case, data = infert)
iboxplot(age, case, data = infert, na.rm = TRUE)
iboxplot(age, case, education, data = infert)
iboxplot(age, case, education, data = infert, show.legend = F)
iboxplot(age, case, education, data = infert, plot.save = T)
}
\seealso{
ibarplot, itab, inumsum, isum
}
