% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iclear.R
\name{iclear}
\alias{iclear}
\title{Clear Global Environment, Plots and Console}
\usage{
iclear()
}
\description{
\code{iclear} is a quick way to clear messy Global Environment, plots and console without interrupting the search path. If you want to detach all packages, see \code{\link{idetach}}.
}
\examples{
x <- rnorm(100)
y <- rnorm(100)
z <- x + y
hist(x)
hist(y)
hist(z)
iclear()
}
\seealso{
\code{\link{idetach}}
}
\author{
Myo Minn Oo (Email: \email{dr.myominnoo@gmail.com} |
Website: \url{https://myominnoo.github.io/})
}
\keyword{all,}
\keyword{clear}
\keyword{console}
\keyword{environment,}
\keyword{global}
\keyword{plots,}
\keyword{remove}
