% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idetach.R
\name{idetach}
\alias{idetach}
\title{Detach user-loaded packges in the search pathway}
\usage{
idetach()
}
\description{
\code{idetach} searches and removes two types of user-loaded objects in the search pathway:

\enumerate{
  \item user-loaded packages
  \item attached data objects
}
}
\examples{
\dontrun{
library(foreign)
library(MASS)
search()
idetach()
search()
}
}
\seealso{
\code{\link{clear}}
}
\author{
Myo Minn Oo (Email: \email{dr.myominnoo@gmail.com} |
Website: \url{https://myominnoo.github.io/})
}
\keyword{all,}
\keyword{detach}
\keyword{packages}
\keyword{user-loaded}
