% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ikdplot.R
\name{ikdplot}
\alias{ikdplot}
\title{Kernel Density Plot, Grouped plotting using ggplot2}
\usage{
ikdplot(x, by = NULL, data = NULL, rnd = 1, na.rm = FALSE,
  main = NULL, xlab = NULL, ylab = NULL, facet = TRUE,
  legend.text = NULL, legend.position = "right",
  line.color = "black", line.type = "solid", fill = "white",
  alpha = 0, vline = TRUE, vline.color = "blue",
  vline.type = "dotted", vline.size = 0.5, save.plot = FALSE,
  plot.name = "ikdplot.tiff", width = 5, height = 4, dpi = 150)
}
\arguments{
\item{x}{a vector describing the bars which make up the plot. It is usually on x axis.}

\item{by}{a vector describing the grouping of x. By default, the plot generates a faceted barplot.}

\item{data}{an optional data frame (or object coercible by as.data.frame to a data frame) containing the variables for contigency table.}

\item{rnd}{an integer indicating the number of decimal places:}

\item{na.rm}{A logical value indicating to remove NA values in the table or not. By Default, the value is TRUE}

\item{main}{a main title for the plot.}

\item{xlab}{a label for the x axis, defaults to a description of x.}

\item{ylab}{a label for the y axis, defaults to a description of y.}

\item{facet}{a logical value. If TRUE, the plot forms a matrix of panels defined by column faceting variables. It is most useful when you have two discrete variables, and all combinations of the variables exist in the data. By default, it is set to TRUE.}

\item{legend.text}{a text for the title of the plot}

\item{legend.position}{the position of legends ("none", "left", "right", "bottom", "top", or two-element numeric vector)}

\item{save.plot}{a logical value. If TRUE, it saves the plot generated in the current working directory.}

\item{plot.name}{a text for plot filename. Suffix can be ".png", ".tiff" and ".pdf"}

\item{width}{a value in inches}

\item{height}{a value in inches}

\item{dpi}{a value for resolution of the plot saved.}
}
\description{
ikdplot (I kdplot) lets you generate k density curve for one variable and curves grouped by a factor variable. These grouped curves can be either overlapped or presented in a faceted fashion.
}
\examples{
ikdplot(infert$age)
ikdplot(age, data = infert)
ikdplot(infert$age, infert$education)
# facet option
ikdplot(age, education, infert, facet = T)
ikdplot(age, education, infert, facet = F)
ikdplot(age, education, infert, alpha = 1)
# alpha option
ikdplot(age, education, infert, facet = F, alpha = .3)
ikdplot(age, education, infert, alpha = .5)
}
