% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ilog.R
\name{ilog}
\alias{ilog}
\alias{ilog.close}
\alias{ilogtxt}
\title{Echo a copy of console to log file}
\usage{
ilog(logfile = "LOG.txt", append = FALSE)

ilog.close()

ilogtxt(cmd, res, s, vis)
}
\arguments{
\item{logfile}{Name of desired log file in \code{.txt} format}

\item{append}{a logical value}

\item{cmd}{command line}

\item{res}{result}

\item{s}{others}

\item{vis}{others}
}
\description{
\code{ilog} echo a copy of console output to log file.
}
\details{
\code{ilog} is a two-step function that allows you a record of your console.
A log is a file containing what you type and console output. If a name is not
specified, then \code{ilog} will use the name \code{<unnamed>.txt}.

\code{ilog} opens a log file and \code{ilog.close} close the file.

\strong{Warnings:}

Clearing the global environment while logging will interrupt the logging
process. In that case, console prompt will be stuck at \code{log> }.
Run the \code{code} below to get the console prompt back to normal.

options(prompt = \code{"> "}, continue = \code{"+ "})
}
\examples{
\dontrun{
## my first log
ilog("myfirstlog.txt")
str(infert)
tab(infert$education)
str(iris)
isum(iris$Sepal.Length)
ilog.close()

## appending the file.
ilog("myfirstlog.txt", append = TRUE)
summary(infert)
summary(iris)
ilog.close()
}
}
\seealso{
\code{\link{clear}}
}
\author{
Myo Minn Oo (Email: \email{dr.myominnoo@gmail.com} |
Website: \url{https://myominnoo.github.io/})
}
\keyword{console}
\keyword{console,}
\keyword{copy}
\keyword{echo,}
\keyword{log,}
\keyword{ouput,}
\keyword{save}
