% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ilog.R
\name{ilog}
\alias{ilog}
\alias{ilog.close}
\title{Simple function for logging console in R}
\usage{
ilog(logfile = "mylog.txt", append = FALSE)

ilog.close()
}
\arguments{
\item{logfile}{A character, denoted by the name of desired log file which follows by an
extension ".txt".}

\item{append}{A logical value, indicating whether the log file is appended or not.}
}
\description{
\code{\link{ilog}} is a function to log console interactively which is inspired by Stata Software.
This function captures outputs from console in a local environment which is running behind
the scene. Please be reminded that clearing the environments can make the logging stop.
In that case, your console will still be like this "log>". See the following instructions:
\enumerate{
  \item The prompt '>' and continue '+' can be restored by typing this in the console:
  \item options(prompt = "> ", continue = "+ ").
}
}
\examples{
ilog("myfirstlog.txt", append = F)
isum(infert)
str(infert)
isum(iris)
str(iris)
isum(AirPassengers) # not logging error or warning messages
ilog.close()
}
\seealso{
\code{\link{isum}}, \code{\link{itab}}, \code{\link{inumsum}}, \code{\link{ibarplot}}, \code{\link{iboxplot}}, \code{\link{ikdplot}}
}
\author{
Myo Minn Oo (Email: \email{dr.myominnoo@gmail.com} |
Website: \url{https://myominnoo.github.io/})
}
\keyword{console}
\keyword{console,}
\keyword{log}
\keyword{outputs,}
\keyword{save}
