% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ilog.R
\name{ilog}
\alias{ilog}
\alias{ilog.close}
\title{Simple function for logging console in R}
\usage{
ilog(logfile = "mylog.txt", append = FALSE)

ilog.close()
}
\arguments{
\item{logfile}{A character, denoted by the name of desired log file which follows by an
extension ".txt".}

\item{append}{A logical value, indicating whether the log file is appended or not.}
}
\description{
ilog is a function to log console interactively which is inspired by Stata Software. This
function captures outputs from console in a local environment which is running behind
the scene. Hence, be advised that clearing the workspace of removing environments can
make it stop working.

The prompt '>' and continue '+' can be restored by typing this in the console: options
(prompt = "> ", continue = "+ ").
}
\examples{
ilog("myfirstlog.txt", append = F)
isum(infert)
str(infert)
isum(iris)
str(iris)
isum(AirPassengers) # not logging the output when errors appear.
ilog.close()
}
\seealso{
isum, igroup, ixtab, irestart
}
\keyword{console}
\keyword{console,}
\keyword{log}
\keyword{outputs,}
\keyword{save}
