% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inumsum.R
\name{inumsum}
\alias{inumsum}
\title{Number Summary and Plot using ggplot2}
\usage{
inumsum(x, y = NULL, by = NULL, data = NULL, rnd = 1,
  na.rm = FALSE, x.varname = NULL, y.varname = NULL,
  by.varname = NULL, plot.display = TRUE, boxplot = TRUE,
  main = NULL, xlab = NULL, ylab = NULL, alpha = 0.1,
  show.legend = TRUE, legend.text = NULL, facet.ncol = 2,
  save.plot = FALSE, plot.name = "inumsum.tiff", width = 5,
  height = 4, dpi = 150)
}
\arguments{
\item{x}{a vector describing the bars which make up the plot. It is usually on x axis.}

\item{by}{a vector describing the grouping of x. By default, the plot generates a faceted barplot.}

\item{data}{an optional data frame (or object coercible by as.data.frame to a data frame) containing the variables for contigency table.}

\item{rnd}{an integer indicating the number of decimal places:}

\item{na.rm}{A logical value indicating to remove NA values in the table or not. By Default, the value is TRUE}

\item{main}{a main title for the plot.}

\item{xlab}{a label for the x axis, defaults to a description of x.}

\item{ylab}{a label for the y axis, defaults to a description of y.}

\item{alpha}{set transparency of the density}

\item{show.legend}{show or hide the legend. Hide the legend by default}

\item{legend.text}{Legend title}

\item{facet.ncol}{number of columns to be faceted}

\item{save.plot}{a logical value. If TRUE, it saves the plot generated in the current working directory.}

\item{plot.name}{a text for plot filename. Suffix can be ".png", ".tiff" and ".pdf"}

\item{width}{a value in inches}

\item{height}{a value in inches}

\item{dpi}{a value for resolution of the plot saved.}
}
\description{
inumsum (I number-summarize) generates seven-number summary for one variable which include number of observations, missing records, mean, standard deviation, median, interquartile range, minimum, maximum. If the x variable is summarized by a factor variable, then the output will include seven-number summary plus standard error, 95% confidence intervals for means. In addition, the function generates automated box plots or Kernel density plot with optional parameters for graphical display.
}
\examples{
inumsum(x = age, data = infert)
inumsum(x = age, data = infert, boxplot = FALSE) # calling Kernel density plot
inumsum(infert$age)
inumsum(age, data = infert)
inumsum(x = age, y = education, data = infert)
infert$case <- factor(infert$case)
inumsum(x = age, y = case, by = education, data = infert, na.rm = TRUE)
inumsum(x = age, y = case, by = education, data = infert, na.rm = TRUE, boxplot = FALSE) # calling Kernel density plot
inumsum(x = age, y = case, by = education, data = infert, na.rm = TRUE, boxplot = FALSE, facet.ncol = 3)
str(iris)
inumsum(Sepal.Length, data = iris)
inumsum(Sepal.Length, Species, data = iris)
inumsum(Sepal.Length, Species, data = iris, boxplot = FALSE)
inumsum(Sepal.Length, Sepal.Width, data = iris)
inumsum(Sepal.Length, Sepal.Width, data = iris, boxplot = FALSE)
inumsum(Sepal.Length, Sepal.Width, Species, data = iris)
inumsum(Sepal.Length, Sepal.Width, Species, data = iris, boxplot = F)
}
\seealso{
ibarplot, iboxplot, inumsum, isum, itab
}
\keyword{grouping}
\keyword{number}
\keyword{statistics,}
\keyword{summary}
\keyword{summary,}
