% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipyramid.R
\name{ipyramid}
\alias{ipyramid}
\title{The Population Pyramid}
\usage{
ipyramid(x, y, data = NULL, na.rm = FALSE, binwidth = NULL,
  labels = NULL, main = NULL, xlab = NULL, ylab = NULL,
  legend.text = NULL, text.size = 12, getdata = FALSE,
  plot.save = FALSE, plot.name = "<unnamed>.tiff", width = 5,
  height = 4, dpi = 150)
}
\arguments{
\item{x}{a numeric vector: age in a typical Population Pyramid plot.}

\item{y}{a binary vector: sex in a typical Population Pyramid plot.}

\item{data}{an optional data frame}

\item{na.rm}{A logical value indicating to remove NA values}

\item{binwidth}{specify the cut points of x:}

\item{labels}{specify to name the factor levels.}

\item{main}{title of the plot}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{legend.text}{title of legend}

\item{text.size}{size of text in the plot. Font size for data level is 14:5 ratio
to axis font size}

\item{getdata}{generate cross-tabulation, data frame in addition to plot.}

\item{plot.save}{save the plot}

\item{plot.name}{specify for name of the plot file to be saved.
Filename can be in either of these extension formats: ".tiff", ".png", and ".pdf"}

\item{width}{integer measured in inches}

\item{height}{integer measured in inches}

\item{dpi}{specify resolution of the plot to be saved}
}
\description{
\code{ipyramid} generates a population pyramid graph and data output.
}
\details{
Histogram is typically used to display the age distribution of a population. However,
it is usual to present age ditribution separately for two sexes because mortality tends to
be different. In a population pyramid, age is scaled vertically and sex horizontally.
The number or frequency scale of sex is zero at the middle and increases or decreases to
both sides.

\strong{Specifying binwidth}

By default, cut-off is set at 10 years interval from the lowest value.
A sequence of cut-off points can also be speicified. For example,
\code{binwidth = seq(1:100, 10)} or \code{binwidth = c(18, 20, 25, 30, 60, 90)}

\strong{Usage in other types of data}

It is interesting to see the application of age-sex pyramid population plot for other
types of data. I would like to hear more comments about this.

\strong{References:}
\enumerate{
  \item An Introduction to Medical Statistics, Martin Bland, Thrid Edition,
  Chapter 16, page 303
}
}
\examples{
## create random vectors
set.seed(100)
age <- sample(11:100, 1000, replace = TRUE)
sex <- factor(sample(c("male", "female"), 1000, replace = TRUE))

ipyramid(age, sex)
ipyramid(age, sex, binwidth = 5)
ipyramid(age, sex, binwidth = 20, getdata = TRUE)

data <- data.frame(age, sex)
ipyramid(age, sex, data = data)
ipyramid(age, sex, data = data, binwidth = 5)
}
\seealso{
\code{\link{ibarplot}}, \code{\link{iboxplot}}, \code{\link{isum}},
\code{\link{formatDate}}
}
\keyword{contigency}
\keyword{cross-tabulation,}
\keyword{mortality}
\keyword{population}
\keyword{pyramid,}
\keyword{statistics,}
\keyword{table}
