% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.date.R
\name{is.date}
\alias{is.date}
\title{Check if an R object is of Date class}
\usage{
is.date(x)
}
\arguments{
\item{x}{an R object}
}
\description{
This function tests if an R object is an object of class "Date" representing
calendar dates.
}
\examples{
isum(iris)
}
\seealso{
isum.factor, isum.numeric, isum.data.frame
}
\keyword{basic}
\keyword{isum,}
\keyword{quick}
\keyword{statistics,}
\keyword{summarize,}
\keyword{summary}
