% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isum.R
\name{isum}
\alias{isum}
\title{Number Summary, Correlation, Grouped and Stratified Summary}
\usage{
isum(x, y = NULL, by = NULL, data = NULL, rnd = 1, na.rm = TRUE,
  plot.show = TRUE)
}
\arguments{
\item{x}{a numeric object}

\item{y}{a numeric or factor object}

\item{by}{a factor object}

\item{data}{an optional data frame}

\item{rnd}{specify rounding of numbers}

\item{na.rm}{A logical value to specify missing values, <NA> in the table}

\item{plot.show}{display bar plots of the data}
}
\description{
\code{isum} generates tables that display univariate, bivariate and stratified
summaries.
}
\details{
Exploring data before jumping into complex analysis is always a necessity.
The first step of an analysis should be to summarize and display data.
By doing so, invaluable insight can be gained through the familiarity with the data.

\code{isum} allows to summarize the distribution of data taken by a single variable,
the association between two numeric variables as well as between a numeric variable
and a factor variable, and finally these associations stratified by a third factor
variable.

This function also generate different data visualization for better exploration by
using \code{iboxplot}. The plot display can be switched off by setting
\code{plot.show = FALSE}.

\strong{References:}
\enumerate{
  \item Essential Medical Statistics, Betty R. Kirkwood & Jonathan A.C. Sterne,
  Second Edition. Chapter 3
  \item An Introduction to MEdical Statistics, Martin Bland, Thrid Edition,
  Chapter 4
}
}
\examples{
data(infert)
str(infert)

## univariate analysis
isum(infert$age)
isum(pooled.stratum, data = infert)

## bivariate: numeric x factor
isum(age, education, data = infert)
isum(parity, education, data = infert)
isum(pooled.stratum, education, data = infert)
isum(infert$parity, factor(infert$case))

## bivariate: numeric x numeric
isum(age, parity, data = infert)
isum(age, stratum, data = infert)
isum(age, pooled.stratum, data = infert)

## stratified: numeric x factor | factor
infert$case <- factor(infert$case)
isum(age, education, case, data = infert)

## stratified: numeric x numeric | factor
isum(age, parity, education, data = infert)
isum(age, stratum, education, data = infert)
}
\seealso{
\code{\link{itab}}, \code{\link{ibarplot}}, \code{\link{strate}},
}
\author{
Myo Minn Oo (Email: \email{dr.myominnoo@gmail.com} |
Website: \url{https://myominnoo.github.io/})
}
\keyword{correlation}
\keyword{distribution,}
\keyword{number}
\keyword{summary,}
