% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isum.R
\name{isum}
\alias{isum}
\title{A Quick Summary function with Graph Display}
\usage{
isum(...)
}
\arguments{
\item{...}{Additional arguments}
}
\description{
isum() lets you do a quick summary on your data

For the class "factor", it tabulates and gives frequency, percentages and
    cummulative percentages. The same is with the class "character".

For the class "integer", it provides seven summary measures:
    mean, standard deviation, median, Q1, Q3, minimum and maximum.

For the class "data.frame", it gives a list of all the variables based on
    their data types.
}
\examples{
isum(iris)
isum(iris, 5, FALSE)
isum(iris$Sepal.Length, rnd = 3 )
isum(iris$Sepal.Length, rnd = 0)
isum(iris$Species, rnd = 3, na.rm = FALSE)
isum(iris, rnd = 3, na.rm = TRUE)
isum(iris, rnd = 3, na.rm = FALSE)
}
\seealso{
isum.factor, isum.numeric, isum.data.frame
}
\keyword{basic}
\keyword{isum,}
\keyword{quick}
\keyword{statistics,}
\keyword{summarize,}
\keyword{summary}
