% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itab.R
\name{itab}
\alias{itab}
\title{Frequency distribution, cross-tabulation and stratified tabulation}
\usage{
itab(x, y = NULL, by = NULL, data = NULL, rnd = 1, na.rm = FALSE,
  pct = "row", plot.show = TRUE)
}
\arguments{
\item{x}{a factor object}

\item{y}{a factor object; if ignored, frequency distribution on x is generated.}

\item{by}{a factor object; if ignored, cross-tabulation on x and y is generated.}

\item{data}{an optional data frame}

\item{rnd}{specify rounding of numbers. See \code{\link{round}}.}

\item{na.rm}{A logical value to specify missing values, <NA> in the table}

\item{pct}{type of percentages in cross-tabulation: by default, it's row percentage.}

\item{plot.show}{display bar plots of the data}
}
\description{
\code{itab} generates Frequency distribution, cross-tabulation and stratified 
tabulations.
}
\details{
Exploring data before jumping into complex analysis is always a necessity. 
The first step of an analysis should be to summarize and display data.
By doing so, invaluable insight can be gained through the familiarity with the data.

\code{itab} 
uses simple tabular techniques for displaying the distribution of
values taken by a single variable, the association between two variables as well as 
stratified tabulation. 

In addition, \code{itab} generates delightful bar plots of the data, using another 
function called \code{\link{ibarplot}}. This function also make use of a powerful 
and famous package, \code{\link{ggplot2}}, hence the plots generated 
by \code{itab} are just beautiful and delightful. If you want to save the plot, 
see \code{\link{ibarplot}} for full details. 

\strong{References:}
\enumerate{
  \item Essential Medical Statistics, Betty R. Kirkwood & Jonathan A.C. Sterne, 
  Second Edition. Chapter 3
  \item An Introduction to MEdical Statistics, Martin Bland, Thrid Edition, 
  Chapter 4
}
}
\examples{
str(infert)

## univariate analysis
itab(infert$education) 
itab(education, data = infert)
itab(spontaneous, data = infert)

## bivariate analysis
itab(infert$education, infert$case) 
itab(education, case, data = infert, pct = 'col') # col percentage
itab(education, case, data = infert, pct = 'all')

## stratified analysis 
itab(education, case, parity, data = infert)
itab(education, case, parity, data = infert, na.rm = TRUE)
}
\seealso{
\code{\link{isum}}, \code{\link{ibarplot}}
}
\author{
Myo Minn Oo (Email: \email{dr.myominnoo@gmail.com} |
Website: \url{https://myominnoo.github.io/})
}
\keyword{cross-tabulation,}
\keyword{distribution,}
\keyword{frequency}
\keyword{stratified}
\keyword{tabulation}
