% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itab.R
\name{itab}
\alias{itab}
\title{Cross-tabulation between two categorical variables with Chi-Squre Test}
\usage{
itab(x, y, rnd = 1, na.rm = TRUE)
}
\arguments{
\item{x}{An R Object: represents Row variable}

\item{y}{An R Object: represents Column variable}

\item{rnd}{an integer indicating the number of decimal places:}

\item{na.rm}{A logical value indicating to remove NA values in the table or not.
By Default, the value is TRUE}
}
\description{
itab() quickly provides three cross-tabulation results between two categorical variables
and quickly adds Chi-Square result as well as Fisher Exact Test result.
}
\examples{
itab(x, y)
}
\seealso{
isum, isum.factor, isum.numeric, isum.data.frame, rall
}
\keyword{basic}
\keyword{categorical,}
\keyword{cross-tabulation,}
\keyword{quick}
\keyword{statistics,}
\keyword{summary}
