% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itab.R
\name{itab}
\alias{itab}
\title{Tabulation, Cross Tabulation and Graphical Display}
\usage{
itab(x, y = NULL, by = NULL, data = NULL, rnd = 1, na.rm = FALSE,
  x.varname = NULL, y.varname = NULL, by.varname = NULL,
  pct = "row", plot.display = TRUE, main = NULL, xlab = NULL,
  ylab = NULL, show.legend = TRUE, legend.text = NULL,
  plot.type = "f", facet.ncol = 2, save.plot = FALSE,
  plot.name = "itab.tiff", width = 5, height = 4, dpi = 150)
}
\arguments{
\item{x}{a vector describing the bars which make up the plot. It is usually on x axis.}

\item{y}{a vector describing the y axis or second variable in cross-tabulation or data relationship.}

\item{by}{a vector describing the grouping of x. By default, the plot generates a faceted barplot.}

\item{data}{an optional data frame (or object coercible by as.data.frame to a data frame) containing the variables for contigency table.}

\item{rnd}{an integer indicating the number of decimal places:}

\item{na.rm}{A logical value indicating to remove NA values in the table or not. By Default, the value is TRUE}

\item{pct}{type of percentages in cross-tabulation: by default, it shows row percentages in two}

\item{plot.display}{logical value, indicating whether plot will be displayed or not}

\item{main}{a main title for the plot.}

\item{xlab}{a label for the x axis, defaults to a description of x.}

\item{ylab}{a label for the y axis, defaults to a description of y.}

\item{show.legend}{show or hide the legend. Hide the legend by default}

\item{legend.text}{Legend title}

\item{plot.type}{if bivariate analysis, type of plot can be specified. By default, this generates a facted plot. 'p' for parrallel barplot, 's' for stacked barplot, 'fs' for full stacked percentage barplot.}

\item{facet.ncol}{number of columns to be faceted}

\item{save.plot}{a logical value. If TRUE, it saves the plot generated in the current working directory.}

\item{plot.name}{a text for plot filename. Suffix can be ".png", ".tiff" and ".pdf"}

\item{width}{a value in inches}

\item{height}{a value in inches}

\item{dpi}{a value for resolution of the plot saved.}
}
\description{
generates simple tabulation, cross-tabulation and stratified cross-tabulation tables.
}
\examples{
str(infert)
# tabulation of one factor variable
itab(education, data = infert)
itab(parity, data = infert)
itab(case, data = infert)
itab(spontaneous, data = infert)
# cross-tabulation
itab(case, education, data = infert, plot.display = FALSE)
itab(case, education, data = infert)
itab(case, education, data = infert, pct = 'col') # change percentage type
itab(case, education, data = infert, pct = 'row') # default percentage type
itab(case, education, data = infert, pct = 'all')
itab(case, education, data = infert, pct = 'none')

# changing plot types
itab(induced, education, data = infert)
itab(induced, education, data = infert, plot.type = 'f') # default faceted type
itab(induced, education, data = infert, plot.type = 'p') # parallel bar
itab(induced, education, data = infert, plot.type = 's') # stacked bar
itab(induced, education, data = infert, plot.type = 'fs') # full stacked bar
itab(induced, education, data = infert, facet.ncol = 3) # change number of facets
itab(induced, education, data = infert, plot.type = 'p', show.legend = FALSE) # legend on and off
itab(induced, education, data = infert, save.plot = TRUE) # save plot to current directory
itab(induced, education, data = infert, save.plot = TRUE, plot.name = "mybarplot.tiff")
itab(induced, education, data = infert, save.plot = TRUE, width = 10, height = 8, dpi = 300)
# three variables - analysis
infert$case <- factor(infert$case)
infert$induced <- factor(infert$induced)
itab(case, education, induced, data = infert)
itab(case, education, induced, data = infert, na.rm = TRUE, pct = 'col')
}
\seealso{
ibarplot, inumsum, ikdplot, iboxplot, isum
}
\keyword{contigency}
\keyword{cross-tabulation,}
\keyword{frequency}
\keyword{table}
\keyword{table,}
\keyword{tabulation,}
