% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep.R
\name{keep}
\alias{keep}
\alias{lose}
\alias{pick}
\alias{arrange}
\title{Keep or lose variables and pick rows}
\usage{
keep(data, ...)

lose(data, ...)

pick(data, ...)

arrange(data, ...)
}
\arguments{
\item{data}{dataframe (or Vector in pick)}

\item{...}{variables}
}
\description{
\code{keep} or \code{lose} variables within dataframe

\code{pick} observations that satisfy specified condition
}
\details{
\code{keep}

It keeps variables in the data, meaning that variables not specified are
deleted.
Variables can also be repositioned based on position.

\code{lose}

It works the same as \code{keep}, except that you specify the
variables or observations to be lost from the dataframe.

Warning: drop and keep are not reversible.  Once you have
eliminated variables and assigned back, you cannot read them back in again.
You would need to go back to the original dataframe and read it in
again.

Instead of assigning back to the original dataframe, consider using multiple
copies of dataframe subsets temporarily. This is usually the best strategy.

\code{pick}

It selects the rows that meets the specified conditions.
This is the same as filter function in spreadsheet or subset in R.

Dataframe or vector can be used and epxressions to specify are almost similar.

\code{arrange}

It arranges the observations of the current data into ascending
order based on the names of the variables.
There is no limit to the number of variables that can be specified in the
arguments.
It can be in decreasing order if specified negative sign or "-" in front of
the desired variable(s) without spacing.
}
\examples{
\dontrun{
keep(infert, education, age, parity)
keep(infert, parity, age:spontaneous, stratum) # variables repositioned

lose(infert, age, parity)
lose(infert, age:case, education)

pick(infert, age > 30)
pick(infert, parity == 6)
pick(infert, age > 30 & age < 35)
pick(infert, case == 0, age > 30 & age < 35, parity == 6)

# use vector instead of dataframe
age <- infert$age
pick(age, age > 40)
pick(infert, age > 40)

arrange(infert, age)
arrange(infert, -age)
arrange(infert, age, induced)
arrange(infert, age, -induced)
arrange(infert, age, -induced, case)
arrange(infert, age, -induced, -case)

arrange(infert, education)
arrange(infert, education, -age)
}
}
\seealso{
\code{\link{generate}}, \code{\link{recode}}, \code{\link{rename}}
}
\author{
Myo Minn Oo (Email: \email{dr.myominnoo@gmail.com} |
Website: \url{https://myominnoo.github.io/})
}
\keyword{delete,}
\keyword{drop,}
\keyword{filter,}
\keyword{keep,}
\keyword{pick}
\keyword{subset,}
