% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{label}
\alias{label}
\title{Label levels of a categorical variable (Value Label)}
\usage{
label(x, labels)
}
\arguments{
\item{x}{A vector}

\item{labels}{specify a character vector that correspond to the levels}
}
\description{
\code{label} function is handy to quickly change levels of either a factor, number,
logical or character vector.
}
\details{
This function allows to quickly change vector into factor with defined levels or even
levels of a factor into new ones.
}
\examples{
data(infert)
str(infert)

## Example 1: Numeric vector
itab(case, data = infert)
case <- label(infert$case, c("No", "Yes"))
itab(case)

## Example 2: factor
itab(education, data = infert)
edu <- label(infert$education, c("low", "middle", "high"))
itab(edu)

## Example 3: character
sex <- rep(c("Male", "Female", NA), c(30, 60, 10))
str(sex)
itab(sex)
sex1 <- label(sex, c('f', 'm'))
str(sex1)
itab(sex1)

## Example 4: Logical
s <- sex == "Male"
str(s)
itab(s)
s1 <- label(s, c("male", "female"))
str(s1)
itab(s1)
}
\seealso{
\code{\link{dup.rm}}
}
\author{
Myo Minn Oo (Email: \email{dr.myominnoo@gmail.com} |
Website: \url{https://myominnoo.github.io/})
}
\keyword{label}
\keyword{label,}
\keyword{levels}
\keyword{value}
