% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.R
\name{label}
\alias{label}
\title{Label variables and dataset}
\usage{
label(data, ...)
}
\arguments{
\item{data}{A data.frame}

\item{...}{One or more expressions:
for labeling dataset, one quoted expression.

For labeling variables, one or more unquoted expressions for variables
and corresponding quoted labels separated by equal sign \code{=}.

\code{var1} = \code{"label1"}, \code{var2} = \code{"label2"}, etc.}
}
\description{
\Sexpr[results=rd]{lifecycle::badge("stable")}

\code{label()} manipulates labels for variables and dataset.
If no label is specified, any existing variable label is removed.
}
\section{Examples}{
\preformatted{## label dataset
iris <- label(iris, "Edgar Anderson's Iris Data")

## label variables
iris <- label(iris,
              Sepal.Length = "Length of Sepal",
              Sepal.Width = "Width of Sepal",
              Petal.Length = "Length of Petal",
              Petal.Width = "Width of Petal",
              Species = "Species of flower")

codebook(iris)
}
}

\seealso{
Other data management: 
\code{\link{append}()},
\code{\link{codebook}()},
\code{\link{describe}()},
\code{\link{duplicates}()},
\code{\link{egen}()},
\code{\link{encode}()},
\code{\link{expand2}()},
\code{\link{formatDate}()},
\code{\link{generate}()},
\code{\link{n_}()},
\code{\link{recode}()},
\code{\link{replace}()},
\code{\link{rowmiss}()}
}
\concept{data management}
