% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.R
\name{label}
\alias{label}
\title{Attach labels to data and variables}
\usage{
label(x, label = NULL)
}
\arguments{
\item{x}{The object to which the label will be added or modified.}

\item{label}{A character string specifying the label to be assigned to the variable.}
}
\value{
The modified object with the updated label.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function manipulates labels.
It supports different classes of objects, including default objects, data frames, and other types.
}
\details{
When used with \code{dplyr}'s \verb{[mutate]} function,
this function allows for easy labeling of variables within a data frame.

If used with a data frame, the function labels the dataset itself, and
the label can be checked using the \verb{[codebook]} function.
}
\examples{

library(dplyr)

iris |>
	 mutate(Species = label(Species, 'Species of iris flower')) |>
	 codebook()

iris |>
	 label("Iris dataset") |>
	 codebook()

}
