% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.R
\name{label}
\alias{label}
\title{Label variables and dataset}
\usage{
label(data, ...)
}
\arguments{
\item{data}{A data.frame}

\item{...}{One or more expressions separated by equal sign \code{=}}
}
\description{
\Sexpr[results=rd]{lifecycle::badge("stable")}

\code{label()} attaches a text label to each variable specified.
If a label is provided without a variable being specified, it will be
attached to the dataset.
}
\section{Examples}{


Here we show the usage for labeling dataset.\if{html}{\out{<div class="r">}}\preformatted{iris <- label(iris, "Edgar Anderson's Iris Data")
}\if{html}{\out{</div>}}\preformatted{#>   (`iris` labeled as `Edgar Anderson's Iris Data`)
}\if{html}{\out{<div class="r">}}\preformatted{codebook(iris)
}\if{html}{\out{</div>}}\preformatted{#> 	      Codebook
#>    Dataset's Name : `iris`
#>   Dataset's Label : Edgar Anderson's Iris Data
#> 	     Vars : 5
#> 	      Obs : 150
}\preformatted{#>  + -- ------------ + ----- ------- --- -- ----- +
#>  | No     Variable | Label    Type Obs NA NA(\%) |
#>  + -- ------------ + ----- ------- --- -- ----- +
#>  |  1 Sepal.Length |       numeric 150  0   0.0 |
#>  |  2  Sepal.Width |       numeric 150  0   0.0 |
#>  |  3 Petal.Length |       numeric 150  0   0.0 |
#>  |  4  Petal.Width |       numeric 150  0   0.0 |
#>  |  5      Species |        factor 150  0   0.0 |
#>  + -- ------------ + ----- ------- --- -- ----- +
}

Here is how to label multiple variables.\if{html}{\out{<div class="r">}}\preformatted{iris <- label(iris, Sepal.Length = "Length of Sepal",
              Petal.Length = "Length of Petal",
              Species = "Type of species")
#>   (`Sepal.Length` labeled as `Length of Sepal`)
#>   (`Petal.Length` labeled as `Length of Petal`)
#>   (`Species` labeled as `Type of species`)
codebook(iris)
#> 	      Codebook
#>    Dataset's Name : `iris`
#>   Dataset's Label : Edgar Anderson's Iris Data
#> 	     Vars : 5
#> 	      Obs : 150
#>  + -- ------------ + --------------- ------- --- -- ----- +
#>  | No     Variable |           Label    Type Obs NA NA(\%) |
#>  + -- ------------ + --------------- ------- --- -- ----- +
#>  |  1 Sepal.Length | Length of Sepal numeric 150  0   0.0 |
#>  |  2  Sepal.Width |                 numeric 150  0   0.0 |
#>  |  3 Petal.Length | Length of Petal numeric 150  0   0.0 |
#>  |  4  Petal.Width |                 numeric 150  0   0.0 |
#>  |  5      Species | Type of species  factor 150  0   0.0 |
#>  + -- ------------ + --------------- ------- --- -- ----- +
}\if{html}{\out{</div>}}
}

