% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelValues.R
\name{labelValues}
\alias{labelValues}
\alias{labelValues.default}
\alias{labelValues.list}
\title{Manipulate labels}
\usage{
labelValues(data = NULL, var, lbl, useNA = FALSE)

\method{labelValues}{default}(data = NULL, var, lbl, useNA = FALSE)

\method{labelValues}{list}(data = NULL, var, lbl, useNA = FALSE)
}
\arguments{
\item{data}{a dataframe object (Optional)}

\item{var}{vectors or lists}

\item{lbl}{specify a string}

\item{useNA}{force to use NA as level, if any}
}
\description{
\code{labelValues} attach value labels to variables
}
\details{
The order of the labels appearing is the same as output from
\code{\link{tab}} function.

If data is specified, it returns the whole dataframe with recoded variables.
Otherwise, only one vector can be used to label values.
}
\examples{
\dontrun{
str(infert)
infert.new <- labelValues(infert, case, c("no", "yes"))
table(infert.new$case)

infert.new <- labelValues(infert,
            list(case, induced),
            list(c("no", "yes"), c("none", "one", "two")))
table(infert.new$case)
table(infert.new$induced)

summary(labelValues(infert, education, c("low", "mid", "high")))
summary(labelValues(infert,
                list(education, induced),
                list(c("low", "mid", "high"),
                     c("none", "one", "two"))))

# vector
case.new <- infert$case
labelValues(NULL, case.new, c("no", "yes"))
}
}
\seealso{
\code{\link{labelVars}}, \code{\link{labelData}}
}
\author{
Myo Minn Oo (Email: \email{dr.myominnoo@gmail.com} |
Website: \url{https://myominnoo.github.io/})
}
\keyword{label}
\keyword{label,}
\keyword{levels,}
\keyword{value}
