% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelVars.R
\name{labelVars}
\alias{labelVars}
\alias{labelVars.default}
\alias{labelVars.character}
\alias{labelVars.list}
\alias{labelData}
\title{Manipulate labels: Label variables or dataframe}
\usage{
labelVars(data = NULL, var, lbl)

\method{labelVars}{default}(data = NULL, var, lbl)

\method{labelVars}{character}(data = NULL, var, lbl)

\method{labelVars}{list}(data = NULL, var, lbl)

labelData(data, lbl)
}
\arguments{
\item{data}{a dataframe object (Optional)}

\item{var}{vectors or list}

\item{lbl}{specify a string}
}
\description{
\code{labelVars} attaches a label to variables or a dataframe
}
\details{
Labels are displayed when you \code{\link{codebook}}
the dataframe. See the examples below.

This function works only on data.frame type. Consider changing to this type
by using as.data.frame().
}
\examples{
\dontrun{
# label variables
codebook(labelVars(infert, age, "AGE"))
codebook(labelVars(infert, c(age, parity), c("AGE", "PARITY")))

# label dataframe
codebook(labelData(infert, "INFERT"))
}
}
\seealso{
\code{\link{codebook}}, \code{\link{labelData}}, \code{\link{labelValues}}
}
\author{
Myo Minn Oo (Email: \email{dr.myominnoo@gmail.com} |
Website: \url{https://myominnoo.github.io/})
}
\keyword{data}
\keyword{dataframe}
\keyword{frame,}
\keyword{label}
\keyword{label,}
