% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelVars.R
\name{labelVars}
\alias{labelVars}
\alias{labelVars.default}
\alias{labelVars.list}
\alias{labelVars.data.frame}
\alias{labelData}
\title{Manipulate labels: Label columns or dataframe}
\usage{
labelVars(x, lbl = NULL, data = NULL)

\method{labelVars}{default}(x, lbl = NULL, data = NULL)

\method{labelVars}{list}(x, lbl = NULL, data = NULL)

\method{labelVars}{data.frame}(x, lbl = NULL, data = NULL)

labelData(x, lbl = NULL)
}
\arguments{
\item{x}{a variable or a list or a dataframe}

\item{lbl}{specify a string}

\item{data}{a dataframe object (Optional)}
}
\description{
\code{labelVars} attaches a label to variables or a dataframe
}
\details{
Dataframe and variable labels are displayed when you \code{\link{codebook}}
the dataframe.
If a label is not specified, existing label is removed.

If data is specified, it returns the whole dataframe with recoded variables.
}
\examples{
codebook(iris)

# label one variable
iris1 <- labelVars(Sepal.Length, "Length of Sepal", iris)
codebook(iris1)
# label multiple variables in a list
iris2 <- labelVars(list(Sepal.Width, Petal.Length, Petal.Width, Species),
                   list("Width of Sepal", "Length of Petal",
                        "Width of Petal", "Type of Species"), iris1)
codebook(iris2)

# label dataframe
labelVars(iris2, "Fisher's Iris data set")
iris3 <- labelData(iris2, "Fisher's Iris data set")
codebook(iris3)
}
\seealso{
\code{\link{codebook}}, \code{\link{labelData}}, \code{\link{labelValue}}
}
\author{
Myo Minn Oo (Email: \email{dr.myominnoo@gmail.com} |
Website: \url{https://myominnoo.github.io/})
}
\keyword{data}
\keyword{dataframe}
\keyword{frame,}
\keyword{label}
\keyword{label,}
