% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listView.R
\name{listView}
\alias{listView}
\alias{listView.default}
\alias{listView.character}
\alias{varsColonSplit}
\alias{listView.list}
\alias{listView.data.frame}
\title{List and view observations of \code{data.frame} in R Console}
\usage{
listView(data, ...)

\method{listView}{default}(...)

\method{listView}{character}(data, ...)

varsColonSplit(data, vars, hasColon)

\method{listView}{list}(data, ...)

\method{listView}{data.frame}(data, ...)
}
\arguments{
\item{data}{a data frame object}

\item{...}{Names of variables to be subsetted.}

\item{vars}{variable's names}

\item{hasColon}{Logical indicating where colon is present or not}
}
\description{
\code{listView()} displays a list of observations of specified or all
variables in R Console.
}
\details{
If \code{data} is specified without any variables, the whole dataset
will be displayed using \code{print.data.frame}. If variables are
specified, a corresponding subset will be printed in R Console.

If specified variables contain duplicated variables, they will all be
printed with numbered suffix. Example, if \code{age} is duplicated
three times, then \code{age.1},  \code{age.2}, and
 \code{age.3} are printed in corresponding order.

Maximum number is set at \code{1e5}.
}
\examples{
\dontrun{
# list infert dataset
listView(infert, education)

listView(infert, age, case, stratum)
listView(infert, education:stratum)

# list several columns in different order
listView(infert, education, induced:pooled.stratum)
listView(infert, age, education, case:age, age:parity)

# list the whole dataset
listView(infert)



# Example from IDRE UCLA
path <- "https://stats.idre.ucla.edu/stat/data/patient_pt1_stata_dm.dta"
hosp <- haven::read_dta(path)
listView(hosp, hospital:pain)
}
}
\references{
STATA DATA MANAGEMENT. UCLA: Statistical Consulting Group.
from https://stats.idre.ucla.edu/stata/seminars/stata-data-management/
(accessed Febrary 25, 2020).
}
\seealso{
\code{\link{codebook}}
}
\author{
For any feedback, please contact \code{Myo Minn Oo} via:

Email: \email{dr.myominnoo@gmail.com}

Website: \url{https://myominnoo.github.io/}
}
\keyword{display,}
\keyword{head,}
\keyword{list,}
\keyword{tail}
\keyword{view,}
