% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mStats.R
\name{n_}
\alias{n_}
\alias{N_}
\title{Count from \code{n_} to \code{N_}}
\usage{
n_(data, ...)

N_(data, ...)
}
\arguments{
\item{data}{data.farme}

\item{...}{variables for grouping}
}
\value{
data.frame
}
\description{
\code{n_()} generates the current observation number
per specified group. It is
regarded as grouped serial numbers.

\code{N_()} generates total number of observation per group.
It is regarded as grouped total number.
}
\details{
If no variable is set in \code{...}, all variables in the datset
is used for grouping.
}
\examples{

x <- n_(iris, Species)
\dontrun{
x
codebook(x)

x <- N_(iris, Species)
x
codebook(x)
}

}
\author{
Email: \email{dr.myominnoo@gmail.com}

Website: \url{https://myominnoo.github.io/}
}
