% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printText.R
\name{printText}
\alias{printText}
\alias{printLines}
\alias{printMsg}
\alias{warningText}
\alias{wrapText}
\title{\code{mStats}'s Printing Functions}
\usage{
printText(x, txt, split = NULL)

printLines(x = "=", width = 80)

printMsg(txt = NULL)

warningText(txt = NULL)

wrapText(txt, width = 80, sep = "\\n")
}
\arguments{
\item{x}{vector, matrix, dataframe or separator (in case of printLines)}

\item{txt}{texts}

\item{split}{separator for printText}

\item{width}{desired character length to display}

\item{sep}{separator for line break}
}
\description{
\code{printText()}, \code{printLines()}, \code{printMsg()},
\code{warningText()}, \code{wrapText()}

Printing Functions to format and display outputs from mStats package
}
\author{
For any feedback, please contact \code{Myo Minn Oo} via:

Email: \email{dr.myominnoo@gmail.com}

Website: \url{https://myominnoo.github.io/})
}
\keyword{print}
\keyword{support,}
