% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode}
\alias{recode}
\title{Recode contents of a variable}
\usage{
recode(data = NULL, var, values_old, values_new)
}
\arguments{
\item{data}{dataset}

\item{var}{name of a new variable}

\item{values_old}{vector}

\item{values_new}{vector (length of 1 or same with values_old)}
}
\description{
\code{recode} easily manipulates contents of a new variable
of a data.frame
}
\details{
\code{recode}
changes the values of variables including categorical variables
according to the rules specified below.

In case of factor, \code{recode} first converts the vector into character,
recodes and then revert back to factor.

If data is specified, it returns the whole dataframe with recoded variables.

\strong{Sample Inputs for conversion}:

Old.value    to    New.value

   #        >>>>       #

c(#, #)     >>>>    c(#, #)

c(#, #)     >>>>       #

   #:#      >>>>       #
}
\examples{
\dontrun{
table(infert$case)
infert.new <- recode(infert, case, 0, 2)
table(infert.new$case)

infert.new <- recode(infert.new, case, c(1, 2), c(3, 4))
table(infert.new$case)

table(infert$parity)
infert.new <- recode(infert, parity, c(1,2,3,4), 0)
table(infert.new$parity)

table(infert$age)
infert.new <- recode(infert, age, 21:30, 1)
infert.new <- recode(infert.new, age, 31:40, 2)
infert.new <- recode(infert.new, age, 41:44, 3)
table(infert.new$age)

# errors
recode(infert, age, 21:30, c(1, 2))
}
}
\seealso{
\code{\link{generate}}, \code{\link{egen}}, \code{\link{duplicates}}
}
\author{
For any feedback, please contact \code{Myo Minn Oo} via:

Email: \email{dr.myominnoo@gmail.com}

Website: \url{https://myominnoo.github.io/}
}
\keyword{change,}
\keyword{generate,}
\keyword{produce}
\keyword{recode,}
