% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode}
\alias{recode}
\alias{mvrecode}
\alias{mvdecode}
\title{Recode variables}
\usage{
recode(data, vars = list(), ...)

mvrecode(data, value)

mvdecode(data, ...)
}
\arguments{
\item{data}{A data.frame}

\item{vars}{Variable to be recoded.
\code{list()} can be used to recode multiple
variables.}

\item{...}{Old and new values to be recoded: \code{old_value}/\code{new_value}}

\item{value}{a value to be recoded for missing values in the data.frame}
}
\description{
\Sexpr[results=rd]{lifecycle::badge("stable")}

\code{recode()} changes the values of numeric variables
according to the values specified.
Values that does not meet any of the conditions, they are left unchanged.
Operator \code{:} can be used as \code{x:y} to indicate a range of numeric numbers.

\code{mvrecode()} converts all missing values \code{NA} in the data.frame to
a specified value.

\code{mvdecode()} converts all coded missing values in the data.frame to \code{NA}.
}
\section{Examples}{

\subsection{Recode one variable at a time.}{\preformatted{infert <- recode(infert, case, 0/"No", 1/"Yes")
tab(infert, case)
}
}

\subsection{Recode multiple variables at the same time, using \code{list}.}{\preformatted{infert <- recode(infert, list(parity, induced), 1:2/1)
tab(infert, parity, induced)
}
}

\subsection{Recode a range of numeric values using operator \code{:}.}{\preformatted{infert <- recode(infert, age, 21:28.9/1, 29:34.9/2, 35:44/3)
tab(infert, age)
}
}

\subsection{Recode value to NA or NA to a new value.}{\preformatted{infert <- recode(infert, parity, 4:6/NA)
tab(infert, parity)
}
}

\subsection{Recode \code{NA} back to 4}{\preformatted{infert <- recode(infert, parity, NA/4)
tab(infert, parity)
}
}

\subsection{Using \code{mvdecode} and \code{mvrecode}}{\preformatted{## Example from A Gentle Introduction to STATA 4th Edition
## Chapter 3 Preparing data for analysis, Page 54
relate <- haven::read_dta("http://www.stata-press.com/data/agis4/relate.dta")

## recode all values from -5 to -1 into NA
relate <- mvdecode(relate, -5:-1)

## recode multiple values of multiple variables
related <- recode(related, list(R3483700, R3483900, R3485300, R3485500),
                  0/4, 1/3, 2/2, 3/1, 4/0

## recode NA to 99
relate <- mvrecode(relate, 99)
}
}
}

\seealso{
Other data management: 
\code{\link{append}()},
\code{\link{codebook}()},
\code{\link{describe}()},
\code{\link{duplicates}()},
\code{\link{egen}()},
\code{\link{encode}()},
\code{\link{expand2}()},
\code{\link{formatDate}()},
\code{\link{generate}()},
\code{\link{label}()},
\code{\link{n_}()},
\code{\link{replace}()},
\code{\link{rowmiss}()}
}
\concept{data management}
