% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode}
\alias{recode}
\title{Recode contents of a variable}
\usage{
recode(data, var, ...)
}
\arguments{
\item{data}{dataset}

\item{var}{name of a new variable}

\item{...}{recode like this: "old value"-"new value" or 1-2 or NA-"Missing"
or "Missing"-NA.}
}
\description{
\code{recode} easily manipulates contents of a new variable
of a data.frame
}
\details{
\code{recode}
changes the values of variables including categorical variables
according to the rules specified below.

\preformatted{labelVar(data, "old value"-"new value", 1-0, NA-"Missing",
                       "Missing"-NA)}
}
\examples{

## use infert data
data(infert)

## tabulate induced to check values
tab(infert, induced)

## recode induced: 2 into 1
infert.new <- recode(infert, induced, 2-1)

## tabulate to check
tab(infert.new, induced)


## recode induced: 0 to 3, 1 to 7, 2 to 10
infert.new <- recode(infert, induced, 0-3, 1-7, 2-10)

## tabulate to check
tab(infert.new, induced)


}
\author{
For any feedback, please contact \code{Myo Minn Oo} via:

Email: \email{dr.myominnoo@gmail.com}

Website: \url{https://myominnoo.github.io/}
}
