% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress.R
\name{regress}
\alias{regress}
\alias{predict.regress}
\alias{plot.regress}
\alias{ladder}
\title{Linear Regression Model}
\usage{
regress(model, vce = FALSE, rnd = 5)

\method{predict}{regress}(object, ...)

\method{plot}{regress}(x, ...)

ladder(data, var)
}
\arguments{
\item{model}{glm or lm model}

\item{vce}{if \code{TRUE}, robust standard errors are calculated. It is
used when heteroskedasticity is detected in data.
Otherwise, OLS standard errors are estimated.}

\item{rnd}{specify rounding of numbers. See \code{\link{round}}.}

\item{object}{a model object for which prediction is desired.}

\item{...}{additional arguments affecting the predictions produced.}

\item{x}{the coordinates of points in the plot. Alternatively, a
      single plotting structure, function or \emph{any \R object with a
        \code{plot} method} can be provided.}

\item{data}{dataset}

\item{var}{variable name}
}
\value{
A list of three \code{data.frame} and \code{model}
}
\description{
\code{regress()} produces summary of the model
with coefficients and 95\% Confident Intervals.

\code{predict} S3 method to predict linear model
after running \code{regress} function from \code{mStats}.
It generates an original data with statistics for model
diagnostics:

fitted (Fitted values)

resid (Residuals)

std.resid (Studentized Residuals)

hat (leverage)

sigma

cooksd (Cook's Distance)

\code{plot} S3 method to plot linear model
after running \code{regress} function from \code{mStats}.

This is used to check model diagnostics.

\code{ladder} converts a variable into a normally
distributed one.
}
\details{
\code{regress} is based on \code{\link{lm}}. All statistics presented
in the function's output are derivates of \code{\link{lm}},
except AIC value which is obtained from \code{\link{AIC}}.

\strong{Outputs}

Outputs can be divided into three parts.
\enumerate{
\item Information about the model
}

Here provides number of observations (Obs.), F value, p-value
from F test,
R Squared value, Adjusted R Squared value, square root of mean square
error
(Root MSE) and AIC value.
\enumerate{
\item Errors
}

Outputs from \code{anova(model)} is tabulated here. SS, DF and MS indicate
sum of square of errors, degree of freedom and mean of square of errors.
\enumerate{
\item Regression Output
}

Coefficients from summary of model are tabulated here along with 95\\%
confidence interval.

\strong{using Robust Standard Errors}

if heteroskedasticity is present in our data sample,
the ordinary least square (OLS) estimator will remain unbiased
and consistent,
but not efficient. The estimated OLS standard errors
will be biased and cannot be solved with a larger sample size.
To remedy this, robust standard erros can be used to adjusted
standard errors.

\deqn{Variance of Robust = (N / N - K) (X'X)^(-1)
 \sum{Xi X'i ei^2} (X'X)^(-1)}

where N =  number of observations, and K =  the number of regressors
(including the intercept). This returns a Variance-covariance (VCV)
matrix
where the diagonal elements are the estimated heteroskedasticity-robust
coefficient
variances — the ones of interest. Estimated coefficient standard errors
are the square root of these diagonal elements.

Note: Credits to Kevin Goulding, The Tarzan Blog.
}
\examples{

## use airquality dataset
data(airquality)

## fit linear model with lm or glm
fit <- lm(Ozone ~ Wind, data = airquality)

## linear model for Ozone
regress(fit)

## run again with robust standard errors and with plots to check
## assumption
regress(fit, vce = TRUE)


## Multiple linear model fit
fit.multi <- lm(Ozone ~ Wind + Solar.R + Temp + Month + Day,
                data = airquality)
## linear model with multiple predictors
reg <- regress(fit.multi, vce = TRUE)

## predict to generate statistics for model diagnostics.
## Input the output from regress
predict(reg)

## Plot to get plots for regression diagnostics
## Input the output from regress
# plot(reg)

## Ladder to generate transformation formula
## It seems log transformation would be the most approapriate one
ladder(airquality, Ozone)

}
\author{
For any feedback, please contact \code{Myo Minn Oo} via:

Email: \email{dr.myominnoo@gmail.com}

Website: \url{https://myominnoo.github.io/}
}
