% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{rename}
\alias{rename}
\alias{rename.default}
\alias{rename.character}
\alias{rename.list}
\title{Rename variable}
\usage{
rename(old.var, new.var, data)

\method{rename}{default}(...)

\method{rename}{character}(old.var, new.var, data)

\method{rename}{list}(old.var, new.var, data)
}
\arguments{
\item{old.var}{R object or list}

\item{new.var}{R object or list}

\item{data}{Data Frame}

\item{...}{Optional arguments}
}
\description{
\code{rename} Rename variable
}
\details{
\code{rename}
changes the name of an existing variable old_varname to new_varname;
the contents of the variable are unchanged.

Groups of variables can be also renamed by specifying variables in a list.
}
\examples{
iris1 <- rename(Species, plantType, iris)
names(iris1)

iris2 <- rename(list(Sepal.Length, Sepal.Width, Petal.Length, Petal.Width),
                list(sl, sw, pl, pw), iris)
names(iris2)
}
\seealso{
\code{\link{ilog}}
}
\author{
Myo Minn Oo (Email: \email{dr.myominnoo@gmail.com} |
Website: \url{https://myominnoo.github.io/})
}
\keyword{change}
\keyword{names}
\keyword{rename,}
