% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{rename}
\alias{rename}
\alias{rename.default}
\alias{rename.character}
\alias{rename.list}
\title{Rename variable}
\usage{
rename(data, old.var, new.var)

\method{rename}{default}(...)

\method{rename}{character}(data, old.var, new.var)

\method{rename}{list}(data, old.var, new.var)
}
\arguments{
\item{data}{Data Frame}

\item{old.var}{R object or list}

\item{new.var}{R object or list}

\item{...}{Optional arguments}
}
\description{
\code{rename} Rename variable
}
\details{
\code{rename}
changes the name of an existing variable old_varname to new_varname;
the contents of the variable are unchanged.

Groups of variables can be also renamed by specifying variables in a list.
}
\examples{
\dontrun{
infert.new <- rename(infert, age, AGE)
str(infert.new)

infert.new <- rename(infert, c(age, parity, induced, case),
               c(AGE, PARITY, INDUCED, CASE))
str(infert.new)
}
}
\seealso{
\code{\link{ilog}}
}
\author{
Myo Minn Oo (Email: \email{dr.myominnoo@gmail.com} |
Website: \url{https://myominnoo.github.io/})
}
\keyword{change}
\keyword{names}
\keyword{rename,}
