% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{rename}
\alias{rename}
\alias{rename.default}
\alias{rename.character}
\alias{rename.list}
\title{Rename variable}
\usage{
rename(data, var_old, var_new)

\method{rename}{default}(...)

\method{rename}{character}(data, var_old, var_new)

\method{rename}{list}(data, var_old, var_new)
}
\arguments{
\item{data}{dataset}

\item{var_old}{name of existing variable}

\item{var_new}{new name to be changed}
}
\value{
\code{data.frame}
}
\description{
\code{rename()} changes names of variables.
}
\details{
\code{rename()} changes the name of an existing variable,
\code{var_old} to a new name, \code{var_new};
the contents of the variable are unchanged.

A group of variables can be also \code{renamed} by specifying the same
number of variables in both \code{var_old} and \code{var_new}
}
\examples{
\dontrun{
## using infert dataset
# renaming one variable
codebook(infert)
infert.new <- rename(infert, age, AGE)
codebook(infert.new)

# renaming a group of variables
infert.new <- rename(infert,
                    c(age, parity, induced, case),
                    c(AGE, PARITY, INDUCED, CASE))
codebook(infert.new)



## IDRE UCLA Example 1
path <- "https://stats.idre.ucla.edu/stat/data/patient_pt1_stata_dm.dta"
hosp <- haven::read_dta(path)
codebook(hosp)

# one variable
hosp <- rename(hosp, sex, SEX)
codebook(hosp)

# multiple variables
hosp <- rename(hosp,
              c(wbc, rbc, bmi, test1, test2),
              c(WBC, RBC, BMI, TEST1, TEST2))
codebook(hosp)
}
}
\seealso{
\code{\link{recode}}, \code{\link{generate}}, \code{\link{egen}}
}
\author{
For any feedback, please contact \code{Myo Minn Oo} via:

Email: \email{dr.myominnoo@gmail.com}

Website: \url{https://myominnoo.github.io/}
}
\keyword{change}
\keyword{names}
\keyword{rename,}
