% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv.R
\name{renv}
\alias{renv}
\title{Clear Workspace, Plots and Console}
\usage{
renv()
}
\description{
renv() is a quick way of clearing messy workspace, plots and console.
}
\examples{
renv()
}
\keyword{clear,}
\keyword{console}
\keyword{plots,}
\keyword{workspace,}
