% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace.R
\name{replace}
\alias{replace}
\title{Change contents of an existing variable}
\usage{
replace(data, var, value, ...)
}
\arguments{
\item{data}{Dataset}

\item{var}{Variable}

\item{value}{Replacement value}

\item{...}{\code{if} conditions.}
}
\value{
Modified Dataset
}
\description{
\code{replace()} alters the values of a variable when specified
conditions are met.

\preformatted{replace(data, var, value,
     var < somevalue | var > somevalue, is.na(var))}

If conditions are not specified, \code{replac()} changes the whole
variable with specified value.

\preformatted{replace(data, var, value)}
}
\details{
It is used when multiple conditions have to be met to change a value.
The function first checks whether specified value is a variable of the dataset.
If yes, then the values are replaced with those of that variables
with the conditions.
}
\examples{
\dontrun{


# Example from IDRE UCLA
path <- "https://stats.idre.ucla.edu/stat/data/patient_pt1_stata_dm.dta"
hosp <- haven::read_dta(path)
codebook(hosp)

## to use piping function
library(magrittr)

hosp \%>\% replace(pain, 0, pain < 4)
hosp \%>\% replace(pain, NA, pain < 4)

## conditions separated by comma as `AND` statement
hosp \%>\% replace(pain, 0, pain < 4, mobility < 2)

## Conditions not specified
hosp \%>\% replace(pain, 0)
hosp \%>\% replace(pain, NA)


## Mixed multiple conditions met
hosp \%>\% replace(co2, NA, mobility == 2, pain < 4)
hosp \%>\% replace(co2, NA, mobility == 2, pain < 4, tumorsize < 70)


## replace with values from another variable
hosp \%>\% replace(pain, mobility, pain < 4)
hosp \%>\% replace(co2, tumorsize, mobility == 2, pain < 4)
hosp \%>\% replace(co2, tumorsize, mobility == 2, pain < 4, tumorsize < 70)


## convert data type without expression
replace(hosp, wbc, as.numeric(wbc)) \%>\%
    codebook

replace(hosp, wbc, sex) \%>\%
    keep(sex, wbc)
}

}
\author{
For any feedback, please contact \code{Myo Minn Oo} via:

Email: \email{dr.myominnoo@gmail.com}

Website: \url{https://myominnoo.github.io/}
}
