% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mStats.R
\name{scatterPlotMatrix}
\alias{scatterPlotMatrix}
\title{Scatter plot matrices with histogram and
correlation coefficients}
\usage{
scatterPlotMatrix(data, main = NULL, pch = 21, ...)
}
\arguments{
\item{data}{data.frame.}

\item{main}{The main title (on top) using font, size (character
    expansion) and color \code{par(c("font.main", "cex.main", "col.main"))}.}

\item{pch}{numeric: point symbol}

\item{...}{further arguments to be passed to or from methods}
}
\description{
A matrix of scatter plots is produced with
Scatter plots with smooth regression line in lower panel,
histograms in diagonal panel and Pearson's correlation
coefficients in upper panel.
}
\examples{

## iris data
# scatterPlotMatrix(iris)

}
\author{
Email: \email{dr.myominnoo@gmail.com}

Website: \url{https://myominnoo.github.io/}
}
