% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strate.R
\name{strate}
\alias{strate}
\title{Tabulate and stratify incidence rates}
\usage{
strate(time, event, data = NULL, strata = NULL, fail = NULL,
  perPY = 1, rnd = 4)
}
\arguments{
\item{time}{specify the timer interval when the subject is at risk}

\item{event}{event of interest to calculate incidence}

\item{data}{an optional data frame}

\item{strata}{specify variable to calculate stratified rates}

\item{fail}{failure event}

\item{perPY}{units to be used in reported rates}

\item{rnd}{Rounding of numbers}
}
\description{
\code{\link{strate}} calculates incidence rates of the cohort. Rates of event 
occurrences, known as incidence rates are outcome measures in longitudinal studies. 
Calculation is based on the method descibed by Kirkwood & Sterne: 
"Essential Medical Statistics: Betty R. Kirkwood & Jonathan A.C. Sterne" Second 
Edition, Part D, Chapter 22, page 229 & 238.
}
\examples{
## use diet data from PaulDickman
## Source: \\url{http://www.pauldickman.com/survival/stataintro.pdf}

diet <- foreign::read.dta("http://www.pauldickman.com/survival/diet.dta")
diet$suryear <- as.numeric(diet$dox - diet$doe) / 365.25

## incidence rates
strate(suryear, chd, data = diet)
strate(suryear, chd, data = diet, perPY = 1000)

## stratified incidence rates
strate(suryear, chd, diet, job, perPY = 1000) # strata
strate(suryear, chd, diet, hieng, perPY = 1000) #strata
}
\seealso{
\code{\link{isum}}
}
\author{
Myo Minn Oo (Email: \email{dr.myominnoo@gmail.com} |
Website: \url{https://myominnoo.github.io/})
}
\keyword{contigency}
\keyword{cross-tabulation,}
\keyword{frequency}
\keyword{table}
\keyword{table,}
\keyword{tabulation,}
