% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strate.R
\name{strate}
\alias{strate}
\alias{calcRate}
\alias{calcSRate}
\title{Calculate Incidence Rates from time-to-event data}
\usage{
strate(data, time, fail, ..., fail_value = NULL, per = 1, rnd = 4)

calcRate(time, fail, fail_value, per, rnd)

calcSRate(time, fail, by, fail_value, per, rnd)
}
\arguments{
\item{data}{Dataset}

\item{time}{person-time variable}

\item{fail}{outcome variable: preferably 1 for event, 0 for censored}

\item{...}{
  Arguments passed on to \code{\link[=strate]{strate}}, \code{\link[=strate]{strate}}
  \describe{
    \item{\code{}}{}
  }}

\item{fail_value}{a value or values to specify failure event}

\item{per}{units to be used in reported rates}

\item{rnd}{Rounding of numbers}

\item{by}{a variable to specify rates}
}
\description{
\code{strate()} calculates incidence rates and Corresponding 95\\% CI.

\code{calcRate} calculates incidence rate and its
confidence intervals.

\code{calcSRate} calculates Stratified incidence rate and its
confidence intervals.
}
\details{
Rates of event occurrences, known as incidence rates are outcome measures in
longitudinal studies. In most longitudinal studies, follow-up times vary due
to logistic reasons, different periods of recruitment, delay enrollment into
the study, lost-to-follow-up, immigration or emigration and death.

\strong{Follow-up time in longitudinal studies}

Period of observation (called as follow-up time) starts when individuals join
the study and ends when they either have an outcome of interest, are lost-to-
follow-up or the follow-up period ends, whichever happens first. This period is
called \strong{person-year-at-risk}. This is denoted by \emph{PY} in \code{strate}
function's output and numer of event by \emph{D}.

\strong{Rate}

is calcluated using the following formula:
\deqn{\lambda = D / PY}

\strong{Confidence interval of rate}

is derived using the following formula:

\deqn{95\% CI (rate) = rate x Error Factor}
\deqn{Error Factor (rate) = exp(1.96 / \sqrt{D})}

\code{plot}, if \code{TRUE}, produces a graph of the rates against
the numerical code used for categories of \code{by}.
}
\section{Functions}{
\itemize{
\item \code{calcRate}: 

\item \code{calcSRate}: 
}}

\examples{


### Using the diet data (Clayton and Hills 1993) described in STATA manual
## import diet data: require haven package to read dta format.
## magrittr package for piping operation
# diet <- haven::read_dta("https://www.stata-press.com/data/r16/diet.dta")
# diet <- diet \%>\%
#    generate(time, (dox - doe) / 365.25) \%>\%
#    replace(time, as.numeric(time)) \%>\%
#      generate(age, as.numeric(doe - dob) / 365.25) \%>\%
#      egen(age, c(41, 51, 61, 71), ageband)
#
#  ## check dataset
#  codebook(diet)
#
#  ## calculate overall rates and 95\% Confidence intervals
#  strate(d, time, fail, fail_value = c(1, 3, 13))
#  ## per 100 unit
#  strate(d, time, fail, fail_value = c(1, 3, 13), per = 100)
#
#  ## calculate Stratified rates and 95\% Confidence Intervals
#  strate(d, time, fail, fail_value = c(1, 3, 13), job)
#  strate(d, time, fail, fail_value = c(1, 3, 13), job, ageband)
#  ## per 100 unit
#  strate(d, time, fail, fail_value = c(1, 3, 13), job, ageband, per = 100)


}
\references{
Betty R. Kirkwood, Jonathan A.C. Sterne (2006, ISBN:978–0–86542–871–3)
}
\author{
Email: \email{dr.myominnoo@gmail.com}

Website: \url{https://myominnoo.github.io/}
}
