% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summ.R
\name{summ}
\alias{summ}
\title{Number Summary for numerical data}
\usage{
summ(data, ..., by = NULL, na.rm = FALSE, rnd = 1)
}
\arguments{
\item{data}{Dataset}

\item{...}{Variable or multiple variables
Colon separator \code{:} can be used to specify multiple variables.}

\item{by}{Varaiable for cross-tabulation}

\item{na.rm}{A logical value to specify missing values,}

\item{rnd}{specify rounding of numbers. See \code{\link{round}}.}
}
\value{
summary measures as \code{data.frame}
}
\description{
\code{summ()} generates summary statistics for numerical data as well as
grouped summary measures.
}
\details{
\code{summ()} reports seven number summary statistics, normality and other additional
metadata.

\preformatted{summ(data, var1)}

\preformatted{summ(data, var1, var2, var3:var5, var10)}

\preformatted{summ(data)}

Normality test is perfomed by Shapiro-Wilk Normality Test. See more at
\code{\link{shapiro.test}}.

'
\strong{ANNOTATIONS}

\code{Obs.} = observation

\code{NA.} = missing data

\code{Mean} = Mean value

\code{Std.Dev} = Standard deviation

\code{Median} = Median value

\code{Q1} = First quartile or percentile

\code{Q3} = Third quartile or percentile

\code{Min} = Minimum value

\code{Max} = Maximum value

\code{Normality} = P-value from Shapiro-Wilk Normality Test

\strong{\verb{Grouped Summary Measures}}

If \code{by} is specified, grouped summary measures are calculated and
produced five number summary, excluding minimum and maximum. In addition,
if levels of \code{by} are more than 2, p-values from ANOVA and Kruskal Wallis tests
are displayed. Otherwise, Student's t-test and Wilcoxon signed rank test are
measured and their respective p-values are tabulated.

There are two parts of the final table. The first part tabulates
grouped summary measures and second part tabulates one-variable summary
measures for corresponding variables.

\preformatted{summ(data, var1, var2, by = var3)}

\preformatted{summ(data, var1, var2, var3:var5, var10, by = var11)}

\preformatted{summ(data, by = var11)}

\strong{Using colon \code{:} spearator}

Colon separator \code{:} can be used to indicate sequence of variables.

\preformatted{summ(data, var1, var2, var3:var5, var10)}
}
\examples{
\dontrun{


## UCLA IDRE Example
## Website:
path <- "https://stats.idre.ucla.edu/stat/data/hsbdemo.dta"
hsb <- haven::read_dta(path)
codebook(hsb)




## single variable
summ(hsb, math)

## multiple variables
summ(hsb, math, write)

## using colon separator
summ(hsb, write:socst, awards, cid)

## the whole dataset
summ(hsb)




# Example from IDRE UCLA
path <- "https://stats.idre.ucla.edu/stat/data/patient_pt1_stata_dm.dta"
hosp <- haven::read_dta(path)
codebook(hosp)


## to use piping function
library(magrittr)


## summary measures
summ(hosp)

## grouped summary measures by sex
summ(hosp, age, rbc:test2, by = sex)


## remove NA values from summary measures
hosp \%>\%
    replace(sex, NA, sex == 12.2) \%>\%
    summ(age, rbc:test2, by = sex, na.rm = FALSE)


hosp \%>\%
    replace(sex, NA, sex == 12.2) \%>\%
    summ(age, rbc:test2, by = sex, na.rm = TRUE)


## the whole dataset
summ(hosp, by = sex)
}

}
\references{
\enumerate{
\item Essential Medical Statistics, Betty R. Kirkwood & Jonathan
A.C. Sterne,
Second Edition.
\item An Introduction to MEdical Statistics, Martin Bland,
Thrid Edition,
\item STATA DATA MANAGEMENT. UCLA: Statistical Consulting Group.
from https://stats.idre.ucla.edu/stata/seminars/stata-data-management/
(accessed Febrary 25, 2020).
}
}
\author{
For any feedback, please contact \code{Myo Minn Oo} via:

Email: \email{dr.myominnoo@gmail.com}

Website: \url{https://myominnoo.github.io/}
}
