% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summ.R, R/summBy.R
\name{summ}
\alias{summ}
\alias{summ.default}
\alias{summ.numeric}
\alias{summ.list}
\alias{summ.data.frame}
\alias{summBy.default}
\alias{summBy.numeric}
\title{Number Summary for numerical data}
\usage{
summ(data = NULL, ..., by = NULL, na.rm = FALSE, rnd = 1, print.table = TRUE)

\method{summ}{default}(...)

\method{summ}{numeric}(data = NULL, x, by = NULL, na.rm = FALSE, rnd = 1, print.table = TRUE)

\method{summ}{list}(data = NULL, ..., by = NULL, na.rm = FALSE, rnd = 1, print.table = TRUE)

\method{summ}{data.frame}(data = NULL, ..., by = NULL, na.rm = FALSE, rnd = 1, print.table = TRUE)

\method{summBy}{default}(...)

\method{summBy}{numeric}(data = NULL, x, by = NULL, na.rm = FALSE, rnd = 1, print.table = TRUE)
}
\arguments{
\item{data}{dataset (optional)}

\item{...}{a variable or variables. Colon separator \code{:} can be
used for multiple variables. See details.}

\item{by}{outcome variable for cross-tabulation}

\item{na.rm}{A logical value to specify missing values,
<NA> in the table}

\item{rnd}{specify rounding of numbers. See \code{\link{round}}.}

\item{print.table}{logical value to display formatted outputs}

\item{x}{exposure variable}
}
\description{
\code{summ()} generates summary statistics for numerical data
}
\details{
\code{summ()} reports number of observations in the dataset,
missing data,
seven number summary statistics, coefficient of
variation (CV) and normality
test.

Normality test is perfomed by Shapiro-Wilk Normality Test. See more at
\code{\link{shapiro.test}}.

\strong{ANNOTATIONS}

\code{Obs.} = observation

\code{NA.} = missing data

\code{Mean} = Mean value

\code{Std.Dev} = Standard deviation

\code{Median} = Median value

\code{Q1} = First quartile or percentile

\code{Q3} = Third quartile or percentile

\code{Min} = Minimum value

\code{Max} = Maximum value

\code{Normality} = P-value from Shapiro-Wilk Normality Test
}
\note{
In case of using the whole dataset for grouped summary measures,
the dataset must be \code{data.frame}.
}
\examples{
\dontrun{
## Using infert dataset
# one variable
summ(infert, age)

# multiple variable
summ(infert, age, induced, case)

# multiple variable using colon separator
summ(infert, age:case)
summ(infert, age, parity:pooled.stratum)

# the whole dataset
summ(infert)
summ(iris)

# without using data argument
induced <- infert$induced
case <- infert$case
parity <- infert$parity
summ(NULL, induced, case, parity)



## IDRE UCLA Example 1
path <- "https://stats.idre.ucla.edu/stat/data/patient_pt1_stata_dm.dta"
hosp <- haven::read_dta(path)
codebook(hosp)

# one variable
summ(hosp, rbc)

# multiple variables
summ(hosp, tumorsize:age, lengthofstay, rbc:test2)

# whole dataset
summ(hosp)




## IDRE UCLA Example 2
path <- "https://stats.idre.ucla.edu/stat/stata/modules/kids.dta"
kids <- haven::read_dta(path)

# display codebook
codebook(kids)

# one variable
summ(kids, age)

# multiple variables
listView(kids, age, wt, birth)
summ(kids, age, wt, birth)

# the whole dataset
summ(kids)





### GROUPED SUMMARY MEASURES
# using infert dataset
# one variable
summ(infert, age, by = education)

# multiple variables
summ(infert, age, induced, case, by = education)
summ(infert, age:pooled.stratum, by = education)

# the whole dataset
summ(infert, by = education)




## IDRE UCLA Example 1
path <- "https://stats.idre.ucla.edu/stat/data/patient_pt1_stata_dm.dta"
hosp <- haven::read_dta(path)
codebook(hosp)

# one variable
summ(hosp, rbc, by = cancerstage)

# multiple variables
summ(hosp, tumorsize:age, lengthofstay, rbc:test2, by = cancerstage)

# whole dataset
summ(hosp, by = cancerstage) ## Error

## change to data.frame and then summ
library(magrittr)
hosp \%>\% data.frame \%>\% summ(by = cancerstage)



## IDRE UCLA Example 2
path <- "https://stats.idre.ucla.edu/stat/stata/modules/kids.dta"
kids <- haven::read_dta(path)

# display codebook
codebook(kids)

# one variable
summ(kids, age, by = sex)

# multiple variables
summ(kids, age, wt, by = sex)

# the whole dataset
summ(kids, by = sex) # Error

library(magrittr)
kids \%>\% data.frame \%>\% summ(by = sex)
}
}
\seealso{
\code{\link{summBy}}, \code{\link{tab}}, \code{\link{xtab}}
}
\author{
For any feedback, please contact \code{Myo Minn Oo} via:

Email: \email{dr.myominnoo@gmail.com}

Website: \url{https://myominnoo.github.io/}
}
\keyword{descriptive,}
\keyword{five}
\keyword{number}
\keyword{statistics,}
\keyword{summary,}
