% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summ.R, R/summBy.R
\name{summ}
\alias{summ}
\alias{summ.default}
\alias{summ.numeric}
\alias{summ.list}
\alias{summ.data.frame}
\alias{summBy.default}
\alias{summBy.numeric}
\title{Number Summary for numerical data}
\usage{
summ(x, data = NULL, rnd = 1, na.rm = FALSE)

\method{summ}{default}(x, data = NULL, rnd = 1, na.rm = FALSE)

\method{summ}{numeric}(x, data = NULL, rnd = 1, na.rm = FALSE)

\method{summ}{list}(x, data = NULL, rnd = 1, na.rm = FALSE)

\method{summ}{data.frame}(x, data = NULL, rnd = 1, na.rm = FALSE)

\method{summBy}{default}(x, y, data = NULL, rnd = 1, na.rm = FALSE)

\method{summBy}{numeric}(x, y, data = NULL, rnd = 1, na.rm = FALSE)
}
\arguments{
\item{x}{numeric object, list or data frame}

\item{data}{a data frame object (Optional)}

\item{rnd}{specify rounding of numbers. See \code{\link{round}}.}

\item{na.rm}{A logical value to specify missing values, <NA> in the table}

\item{y}{a factor or character object}
}
\description{
\code{summ} generates seven summary statistics for numerical data and tests
for normality on the fly.
}
\details{
\code{summ} reports number of observations in the dataset, missing data,
seven number summary statistics, coefficient of variation (CV.) and normality
test.
Normality test is perfomed by Shapiro-Wilk Normality Test. See more at
\code{\link{shapiro.test}}.

CV = Standard Deviation / MEAN * 100

\strong{ANNOTATIONS}

Obs. = observation

NA. = missing data

Mean = Mean value

Std.Dev = Standard deviation

Median = Median value

Q1 = First quartile or percentile

Q3 = Third quartile or percentile

Min = Minimum value

Max = Maximum value

Normal.test = test by Shapiro-Wilk Normality Test
}
\examples{
\dontrun{
# single variable
summ(age, infert)
summ(infert$age)

# multiple variable
summ(c(age, case, parity), infert)

# whole dataframe
summ(infert)
summ(iris)
summ(mtcars)

# subset: one category and one number
infert.new <- keep(infert, education, age)
summ(infert.new)
}
}
\seealso{
\code{\link{summBy}}, \code{\link{tab}}, \code{\link{xtab}}
}
\author{
Myo Minn Oo (Email: \email{dr.myominnoo@gmail.com} |
Website: \url{https://myominnoo.github.io/})
}
\keyword{descriptive,}
\keyword{five}
\keyword{number}
\keyword{statistics,}
\keyword{summary,}
