% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary}
\alias{summary}
\alias{summary.tab}
\title{Summary tables for publication}
\usage{
\method{summary}{tab}(object, ...)
}
\arguments{
\item{object}{an object for which a summary is desired.}

\item{...}{additional arguments affecting the summary produced.}
}
\description{
\code{mStats} provides S3 functions for \code{\link[=summary]{summary()}} to format the
outputs from \code{tab()} and prepare for publication-ready tables.
}
\details{
\code{summary()} supports the following functions of \code{mStats}:

to add more
}
\section{Examples}{

\subsection{Summary measures for categorical variables}{

Create summary table for frequency measure:\if{html}{\out{<div class="r">}}\preformatted{table1 <- tab(infert, education, parity, case)
#>  One-way tabulation
#>  + --------- ------- + ----- ------- ----- +
#>  |  Variable   Level | Freq. Percent  Cum. |
#>  + --------- ------- + ----- ------- ----- +
#>  | education         |                     |
#>  |            0-5yrs |    12     4.8   4.8 |
#>  |           6-11yrs |   120    48.4  53.2 |
#>  |           12+ yrs |   116    46.8 100.0 |
#>  +           ------- + ----- ------- ----- +
#>  |             Total |   248   100.0 100.0 |
#>  + --------- ------- + ----- ------- ----- +
#>  |    parity         |                     |
#>  |                 1 |    99    39.9  39.9 |
#>  |                 2 |    81    32.7  72.6 |
#>  |                 3 |    36    14.5  87.1 |
#>  |                 4 |    18     7.3  94.4 |
#>  |                 5 |     6     2.4  96.8 |
#>  |                 6 |     8     3.2 100.0 |
#>  +           ------- + ----- ------- ----- +
#>  |             Total |   248   100.0 100.0 |
#>  + --------- ------- + ----- ------- ----- +
#>  |      case         |                     |
#>  |                 0 |   165    66.5  66.5 |
#>  |                 1 |    83    33.5 100.0 |
#>  +           ------- + ----- ------- ----- +
#>  |             Total |   248   100.0 100.0 |
#>  + --------- ------- + ----- ------- ----- +

summary(table1)
#>     Variable   Level Total\\n n (\%)
#> 1      Total           248 (100.0)
#> 2  education                      
#> 3             0-5yrs      12 (4.8)
#> 4            6-11yrs    120 (48.4)
#> 5            12+ yrs    116 (46.8)
#> 6     parity                      
#> 7                  1     99 (39.9)
#> 8                  2     81 (32.7)
#> 9                  3     36 (14.5)
#> 10                 4      18 (7.3)
#> 11                 5       6 (2.4)
#> 12                 6       8 (3.2)
#> 13      case                      
#> 14                 0    165 (66.5)
#> 15                 1     83 (33.5)
}\if{html}{\out{</div>}}

Create summary table stratified by a stratification variable \code{by}:\if{html}{\out{<div class="r">}}\preformatted{table2 <- tab(infert, education, parity, by = case)
#>  Two-way tabulation
#>                               case
#>  + --------- ------- + --- ---- -- ---- ----- ----- ----- ----- +
#>  |  Variable   Level |   0 r(\%)  1 r(\%) Total  r(\%)  chi2 exact |
#>  + --------- ------- + --- ---- -- ---- ----- ----- ----- ----- +
#>  | education         |                                          |
#>  |            0-5yrs |   8 66.7  4 33.3    12 100.0 0.999 1.000 |
#>  |           6-11yrs |  80 66.7 40 33.3   120 100.0             |
#>  |           12+ yrs |  77 66.4 39 33.6   116 100.0             |
#>  +           ------- + --- ---- -- ---- ----- ----- ----- ----- +
#>  |             Total | 165 66.5 83 33.5   248 100.0             |
#>  + --------- ------- + --- ---- -- ---- ----- ----- ----- ----- +
#>  |    parity         |                                          |
#>  |                 1 |  66 66.7 33 33.3    99 100.0 1.000 1.000 |
#>  |                 2 |  54 66.7 27 33.3    81 100.0             |
#>  |                 3 |  24 66.7 12 33.3    36 100.0             |
#>  |                 4 |  12 66.7  6 33.3    18 100.0             |
#>  |                 5 |   4 66.7  2 33.3     6 100.0             |
#>  |                 6 |   5 62.5  3 37.5     8 100.0             |
#>  +           ------- + --- ---- -- ---- ----- ----- ----- ----- +
#>  |             Total | 165 66.5 83 33.5   248 100.0             |
#>  + --------- ------- + --- ---- -- ---- ----- ----- ----- ----- +

summary(table2)
#>     Variable   Level Total\\n n (\%)  0\\n n (\%) 1\\n n (\%) p-value\\n (Chi2)
#> 1      Total           248 (100.0) 165 (66.5) 83 (33.5)                 
#> 2  education                                                       0.999
#> 3             0-5yrs    12 (100.0)   8 (66.7)  4 (33.3)                 
#> 4            6-11yrs   120 (100.0)  80 (66.7) 40 (33.3)                 
#> 5            12+ yrs   116 (100.0)  77 (66.4) 39 (33.6)                 
#> 6     parity                                                       1.000
#> 7                  1    99 (100.0)  66 (66.7) 33 (33.3)                 
#> 8                  2    81 (100.0)  54 (66.7) 27 (33.3)                 
#> 9                  3    36 (100.0)  24 (66.7) 12 (33.3)                 
#> 10                 4    18 (100.0)  12 (66.7)  6 (33.3)                 
#> 11                 5     6 (100.0)   4 (66.7)  2 (33.3)                 
#> 12                 6     8 (100.0)   5 (62.5)  3 (37.5)                 
#>    p-value\\n (Exact)
#> 1                   
#> 2              1.000
#> 3                   
#> 4                   
#> 5                   
#> 6              1.000
#> 7                   
#> 8                   
#> 9                   
#> 10                  
#> 11                  
#> 12
}\if{html}{\out{</div>}}
}
}

\seealso{
Other reporting: 
\code{\link{compact}()}
}
\concept{reporting}
