% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mStats.R
\name{summary}
\alias{summary}
\alias{summary.tab}
\alias{summary.summ}
\title{Table Format for Publication}
\usage{
\method{summary}{tab}(object, ...)

\method{summary}{summ}(object, ...)
}
\arguments{
\item{object}{an object for which a summary is desired.}

\item{...}{additional arguments affecting the summary produced.}
}
\description{
\code{summary()} organizes the output and print
a favorable format
to the console, which is used with rmarkdown package
to produce publication-ready tables.
}
\examples{

\dontrun{
## Summary for tabulation
x <- tab(infert, education, parity:spontaneous)
summary(x)

x <- tab(infert, education, parity:spontaneous, by = case)
summary(x)

## Summary for summary statistics
x <- summ(iris)
summary(x)

x <- summ(iris, by = Species)
summary(x)

x <- summ(iris, by = Species, detail = TRUE)
summary(x)
}


}
\author{
Email: \email{dr.myominnoo@gmail.com}

Website: \url{https://myominnoo.github.io/}
}
