% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summBy.R
\name{summBy}
\alias{summBy}
\alias{summBy.list}
\alias{summBy.data.frame}
\title{Grouped Number Summary for Numerical Data}
\usage{
summBy(data = NULL, ..., by = NULL, na.rm = FALSE, rnd = 1, print.table = TRUE)

\method{summBy}{list}(data = NULL, ..., by, na.rm = FALSE, rnd = 1, print.table = TRUE)

\method{summBy}{data.frame}(data = NULL, ..., by, na.rm = FALSE, rnd = 1, print.table = TRUE)
}
\arguments{
\item{data}{dataset (optional)}

\item{...}{a variable or variables. Colon separator \code{:} can be
used for multiple variables. See details.}

\item{by}{outcome variable for cross-tabulation}

\item{na.rm}{A logical value to specify missing values,
<NA> in the table}

\item{rnd}{specify rounding of numbers. See \code{\link{round}}.}

\item{print.table}{logical value to display formatted outputs}
}
\value{
\code{data.frame}
}
\description{
\code{summBy()} generates summary statistics for numerical data
grouped by a categorical variable
}
\details{
\code{summBy()} produces grouped summary statistics, similar
to the outputs generated by
\code{summ()}.


\strong{ANNOTATIONS}

\code{Obs.} = observation

\code{NA.} = missing data

\code{Mean} = Mean value

\code{Std.Dev} = Standard deviation

\code{Median} = Median value

\code{Q1} = First quartile or percentile

\code{Q3} = Third quartile or percentile

\code{Normality} = P-value from Shapiro-Wilk Normality Test

\strong{P-values from Statistical Tests}:

\enumerate{
    \item Student's t-test if two levels
    \item Wilcoxon Rank Sum test if two levels (also known
    as Mann-Whitney test)
    \item ANOVA test if more than two levels
    \item Kruskal-Wallis test if more than two levels
}
}
\examples{
\dontrun{

### GROUPED SUMMARY MEASURES
# using infert dataset
# one variable
summBy(infert, age, by = education)

# multiple variables
summBy(infert, age, induced, case, by = education)
summBy(infert, age:pooled.stratum, by = education)

# the whole dataset
summBy(infert, by = education)




## IDRE UCLA Example 1
path <- "https://stats.idre.ucla.edu/stat/data/patient_pt1_stata_dm.dta"
hosp <- haven::read_dta(path)
codebook(hosp)

# one variable
summBy(hosp, rbc, by = cancerstage)

# multiple variables
summBy(hosp, tumorsize:age, lengthofstay, rbc:test2, by = cancerstage)

# whole dataset
summBy(hosp, by = cancerstage) ## Error

## change to data.frame and then summBy
library(magrittr)
hosp \%>\% data.frame \%>\% summBy(by = cancerstage)






## IDRE UCLA Example 2
path <- "https://stats.idre.ucla.edu/stat/stata/modules/kids.dta"
kids <- haven::read_dta(path)

# display codebook
codebook(kids)

# one variable
summBy(kids, age, by = sex)

# multiple variables
summBy(kids, age, wt, by = sex)

# the whole dataset
summBy(kids, by = sex) # Error

library(magrittr)
kids \%>\% data.frame \%>\% summBy(by = sex)
}
}
\seealso{
\code{\link{summ}}, \code{\link{tab}}, \code{\link{xtab}}
}
\author{
For any feedback, please contact \code{Myo Minn Oo} via:

Email: \email{dr.myominnoo@gmail.com}

Website: \url{https://myominnoo.github.io/}
}
\keyword{descriptive,}
\keyword{five}
\keyword{number}
\keyword{statistics,}
\keyword{summary,}
