% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summBy.R
\name{summBy}
\alias{summBy}
\alias{summBy.list}
\alias{summBy.data.frame}
\title{Number summary grouped by a factor}
\usage{
summBy(x, y, data = NULL, rnd = 1, na.rm = FALSE, print.table = TRUE)

\method{summBy}{list}(x, y, data = NULL, rnd = 1, na.rm = FALSE, print.table = TRUE)

\method{summBy}{data.frame}(x, y, data = NULL, rnd = 1, na.rm = FALSE, print.table = TRUE)
}
\arguments{
\item{x}{a numeric object}

\item{y}{a factor or character object}

\item{data}{a data frame object (Optional)}

\item{rnd}{specify rounding of numbers. See \code{\link{round}}.}

\item{na.rm}{A logical value to specify missing values, <NA> in the table}

\item{print.table}{logical value to display formatted outputs}
}
\description{
\code{summ.by} generates seven number summary statistics and tests normality
on the fly grouped by a categorical variable.
}
\details{
Similar to \code{summ} output, \code{summBy}
reports number of observations in the dataset, missing data, seven number
summary statistics, coefficient of variation (CV.) and normality test.
Normality test is perfomed by Shapiro-Wilk Normality Test. See more at
\code{\link{shapiro.test}}.

\strong{ANNOTATIONS}

Obs. = observation

NA. = missing data

Mean = Mean value

Std.Dev = Standard deviation

Median = Median value

Q1 = First quartile or percentile

Q3 = Third quartile or percentile

Min = Minimum value

Max = Maximum value

Normal.test = test by Shapiro-Wilk Normality Test

If the second variable has two levels of category, it performs either
Student's t-test
\code{\link{t.test}} or Wilcoxon test (Mann-Whitney's test)
\code{\link{wilcox.test}}. If more than two levels, ANOVA
\code{\link{aov}} or
Kruskal-Wallis rank sum test \code{\link{kruskal.test}}
is carried out to test the difference between different groups.
}
\examples{
\dontrun{
# single x variable
summBy(age, education, infert)
summBy(infert$age, infert$education)

# multiple variables
summBy(c(age, parity, stratum), education, infert)

# whole dataframe
summBy(infert, education)
summBy(iris, Species)

# variables labels
infert.new <- labelVars(infert,
                c(age, case, spontaneous, education),
                c("age of patient",
                  "case (yes or no)",
                  "spontaneous labor",
                  "education of patient"))
infert.new <- keep(infert.new, age, case, spontaneous, education)
codebook(infert.new)
summBy(infert.new, education)
}
}
\seealso{
\code{\link{summ}}, \code{\link{tab}}, \code{\link{xtab}}
}
\author{
Myo Minn Oo (Email: \email{dr.myominnoo@gmail.com} |
Website: \url{https://myominnoo.github.io/})
}
\keyword{descriptive,}
\keyword{five}
\keyword{number}
\keyword{statistics,}
\keyword{summary,}
