% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{tab}
\alias{tab}
\title{Tabulate or cross-tabulate}
\usage{
tab(data, ..., by = NULL, row.pct = TRUE, na.rm = FALSE, digits = 1)
}
\arguments{
\item{data}{A data.frame}

\item{...}{One or more unquoted expressions separated by commas.
Variable names can be used as if they were positions in the data
frame, so expressions like \code{x:y} can be used to select a range of
variables.}

\item{by}{Stratification Variable}

\item{row.pct}{This allows you to control how relative frequency are
calculated or none at all. The value can be:
\itemize{
\item \code{TRUE} reports relative frequency within its row.
\item \code{FALSE} reports relative frequency within its column.
\item \code{NULL} to hide relative frequency.
}}

\item{na.rm}{logical: if \code{TRUE}, it removes observations with missing values.}

\item{digits}{specify rounding of numbers.}
}
\description{
\Sexpr[results=rd]{lifecycle::badge("stable")}

\code{tab()} attaches a text label to each variable specified.
If a label is provided without a variable being specified, it will be
attached to the dataset.
}
\section{Examples}{


Tabulate variable by name:\if{html}{\out{<div class="r">}}\preformatted{tab(infert, education)
#>  + --------- ------- + ----- ------- ----- +
#>  |  Variable   Level | Freq. Percent  Cum. |
#>  + --------- ------- + ----- ------- ----- +
#>  | education         |                     |
#>  |            0-5yrs |    12     4.8   4.8 |
#>  |           6-11yrs |   120    48.4  53.2 |
#>  |           12+ yrs |   116    46.8 100.0 |
#>  +           ------- + ----- ------- ----- +
#>  |             Total |   248   100.0 100.0 |
#>  + --------- ------- + ----- ------- ----- +
}\if{html}{\out{</div>}}

Tabulate multiple variables by separating them with commas:\if{html}{\out{<div class="r">}}\preformatted{tab(infert, education, parity, induced)
#>  + --------- ------- + ----- ------- ----- +
#>  |  Variable   Level | Freq. Percent  Cum. |
#>  + --------- ------- + ----- ------- ----- +
#>  | education         |                     |
#>  |            0-5yrs |    12     4.8   4.8 |
#>  |           6-11yrs |   120    48.4  53.2 |
#>  |           12+ yrs |   116    46.8 100.0 |
#>  +           ------- + ----- ------- ----- +
#>  |             Total |   248   100.0 100.0 |
#>  + --------- ------- + ----- ------- ----- +
#>  |    parity         |                     |
#>  |                 1 |    99    39.9  39.9 |
#>  |                 2 |    81    32.7  72.6 |
#>  |                 3 |    36    14.5  87.1 |
#>  |                 4 |    18     7.3  94.4 |
#>  |                 5 |     6     2.4  96.8 |
#>  |                 6 |     8     3.2 100.0 |
#>  +           ------- + ----- ------- ----- +
#>  |             Total |   248   100.0 100.0 |
#>  + --------- ------- + ----- ------- ----- +
#>  |   induced         |                     |
#>  |                 0 |   143    57.7  57.7 |
#>  |                 1 |    68    27.4  85.1 |
#>  |                 2 |    37    14.9 100.0 |
#>  +           ------- + ----- ------- ----- +
#>  |             Total |   248   100.0 100.0 |
#>  + --------- ------- + ----- ------- ----- +
}\if{html}{\out{</div>}}

The \code{:} operator selects a range of consecutive variables:\if{html}{\out{<div class="r">}}\preformatted{tab(infert, education, parity:spontaneous)
#>  + ----------- ------- + ----- ------- ----- +
#>  |    Variable   Level | Freq. Percent  Cum. |
#>  + ----------- ------- + ----- ------- ----- +
#>  |   education         |                     |
#>  |              0-5yrs |    12     4.8   4.8 |
#>  |             6-11yrs |   120    48.4  53.2 |
#>  |             12+ yrs |   116    46.8 100.0 |
#>  +             ------- + ----- ------- ----- +
#>  |               Total |   248   100.0 100.0 |
#>  + ----------- ------- + ----- ------- ----- +
#>  |      parity         |                     |
#>  |                   1 |    99    39.9  39.9 |
#>  |                   2 |    81    32.7  72.6 |
#>  |                   3 |    36    14.5  87.1 |
#>  |                   4 |    18     7.3  94.4 |
#>  |                   5 |     6     2.4  96.8 |
#>  |                   6 |     8     3.2 100.0 |
#>  +             ------- + ----- ------- ----- +
#>  |               Total |   248   100.0 100.0 |
#>  + ----------- ------- + ----- ------- ----- +
#>  |     induced         |                     |
#>  |                   0 |   143    57.7  57.7 |
#>  |                   1 |    68    27.4  85.1 |
#>  |                   2 |    37    14.9 100.0 |
#>  +             ------- + ----- ------- ----- +
#>  |               Total |   248   100.0 100.0 |
#>  + ----------- ------- + ----- ------- ----- +
#>  |        case         |                     |
#>  |                   0 |   165    66.5  66.5 |
#>  |                   1 |    83    33.5 100.0 |
#>  +             ------- + ----- ------- ----- +
#>  |               Total |   248   100.0 100.0 |
#>  + ----------- ------- + ----- ------- ----- +
#>  | spontaneous         |                     |
#>  |                   0 |   141    56.9  56.9 |
#>  |                   1 |    71    28.6  85.5 |
#>  |                   2 |    36    14.5 100.0 |
#>  +             ------- + ----- ------- ----- +
#>  |               Total |   248   100.0 100.0 |
#>  + ----------- ------- + ----- ------- ----- +
}\if{html}{\out{</div>}}

Tabulate all categorical variables without specifying any variables:\if{html}{\out{<div class="r">}}\preformatted{tab(infert)
#>  + --------- ------- + ----- ------- ----- +
#>  |  Variable   Level | Freq. Percent  Cum. |
#>  + --------- ------- + ----- ------- ----- +
#>  | education         |                     |
#>  |            0-5yrs |    12     4.8   4.8 |
#>  |           6-11yrs |   120    48.4  53.2 |
#>  |           12+ yrs |   116    46.8 100.0 |
#>  +           ------- + ----- ------- ----- +
#>  |             Total |   248   100.0 100.0 |
#>  + --------- ------- + ----- ------- ----- +
}\if{html}{\out{</div>}}

Tabulate variables with a stratification variable \code{by}:\if{html}{\out{<div class="r">}}\preformatted{tab(infert, parity:case, by = education)
#> 
#>                                    education
#>  + -------- ----- + ------ ---- ------- ---- ------- ---- ----- ----- -----
#>  | Variable Level | 0-5yrs r(\%) 6-11yrs r(\%) 12+ yrs r(\%) Total  r(\%)  chi2
#>  + -------- ----- + ------ ---- ------- ---- ------- ---- ----- ----- -----
#>  |   parity       |                                                        
#>  |              1 |      3  3.0      42 42.4      54 54.5    99 100.0 0.000
#>  |              2 |      0  0.0      42 51.9      39 48.1    81 100.0      
#>  |              3 |      0  0.0      21 58.3      15 41.7    36 100.0      
#>  |              4 |      3 16.7      12 66.7       3 16.7    18 100.0      
#>  |              5 |      0  0.0       3 50.0       3 50.0     6 100.0      
#>  |              6 |      6 75.0       0  0.0       2 25.0     8 100.0      
#>  +          ----- + ------ ---- ------- ---- ------- ---- ----- ----- -----
#>  |          Total |     12  4.8     120 48.4     116 46.8   248 100.0      
#>  + -------- ----- + ------ ---- ------- ---- ------- ---- ----- ----- -----
#>  |  induced       |                                                        
#>  |              0 |      4  2.8      78 54.5      61 42.7   143 100.0 0.002
#>  |              1 |      2  2.9      27 39.7      39 57.4    68 100.0      
#>  |              2 |      6 16.2      15 40.5      16 43.2    37 100.0      
#>  +          ----- + ------ ---- ------- ---- ------- ---- ----- ----- -----
#>  |          Total |     12  4.8     120 48.4     116 46.8   248 100.0      
#>  + -------- ----- + ------ ---- ------- ---- ------- ---- ----- ----- -----
#>  |     case       |                                                        
#>  |              0 |      8  4.8      80 48.5      77 46.7   165 100.0 0.999
#>  |              1 |      4  4.8      40 48.2      39 47.0    83 100.0      
#>  +          ----- + ------ ---- ------- ---- ------- ---- ----- ----- -----
#>  |          Total |     12  4.8     120 48.4     116 46.8   248 100.0      
#>  + -------- ----- + ------ ---- ------- ---- ------- ---- ----- ----- -----
#>  ----- +
#>  exact |
#>  ----- +
#>        |
#>  0.000 |
#>        |
#>        |
#>        |
#>        |
#>        |
#>  ----- +
#>        |
#>  ----- +
#>        |
#>  0.012 |
#>        |
#>        |
#>  ----- +
#>        |
#>  ----- +
#>        |
#>  1.000 |
#>        |
#>  ----- +
#>        |
#>  ----- +
}\if{html}{\out{</div>}}
}

\concept{statistical analysis\preformatted{}}
