% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{tab}
\alias{tab}
\title{Tabulate or cross-tabulate}
\usage{
tab(data, ..., by = NULL, row.pct = TRUE, na.rm = FALSE, digits = 1)
}
\arguments{
\item{data}{A data.frame}

\item{...}{One or more unquoted expressions separated by commas.
Variable names can be used as if they were positions in the data
frame, so expressions like \code{x:y} can be used to select a range of
variables.}

\item{by}{Stratification Variable}

\item{row.pct}{This allows you to control how relative frequency are
calculated or none at all. The value can be:
\itemize{
\item \code{TRUE} reports relative frequency within its row.
\item \code{FALSE} reports relative frequency within its column.
\item \code{NULL} to hide relative frequency.
}}

\item{na.rm}{logical: if \code{TRUE}, it removes observations with missing values.}

\item{digits}{specify rounding of numbers.}
}
\description{
\Sexpr[results=rd]{lifecycle::badge("stable")}

\code{tab()} attaches a text label to each variable specified.
If a label is provided without a variable being specified, it will be
attached to the dataset.
}
\section{Examples}{


Tabulate variable by name:\if{html}{\out{<div class="r">}}\preformatted{tab(infert, education)
#>  One-way tabulation
#>  + --------- ------- + ----- ------- ----- +
#>  |  Variable   Level | Freq. Percent  Cum. |
#>  + --------- ------- + ----- ------- ----- +
#>  | education         |                     |
#>  |            0-5yrs |    12     4.8   4.8 |
#>  |           6-11yrs |   120    48.4  53.2 |
#>  |           12+ yrs |   116    46.8 100.0 |
#>  +           ------- + ----- ------- ----- +
#>  |             Total |   248   100.0 100.0 |
#>  + --------- ------- + ----- ------- ----- +
}\if{html}{\out{</div>}}

Tabulate multiple variables by separating them with commas:\if{html}{\out{<div class="r">}}\preformatted{tab(infert, education, parity, induced)
#>  One-way tabulation
#>  + --------- ------- + ----- ------- ----- +
#>  |  Variable   Level | Freq. Percent  Cum. |
#>  + --------- ------- + ----- ------- ----- +
#>  | education         |                     |
#>  |            0-5yrs |    12     4.8   4.8 |
#>  |           6-11yrs |   120    48.4  53.2 |
#>  |           12+ yrs |   116    46.8 100.0 |
#>  +           ------- + ----- ------- ----- +
#>  |             Total |   248   100.0 100.0 |
#>  + --------- ------- + ----- ------- ----- +
#>  |    parity         |                     |
#>  |                 1 |    99    39.9  39.9 |
#>  |                 2 |    81    32.7  72.6 |
#>  |                 3 |    36    14.5  87.1 |
#>  |                 4 |    18     7.3  94.4 |
#>  |                 5 |     6     2.4  96.8 |
#>  |                 6 |     8     3.2 100.0 |
#>  +           ------- + ----- ------- ----- +
#>  |             Total |   248   100.0 100.0 |
#>  + --------- ------- + ----- ------- ----- +
#>  |   induced         |                     |
#>  |                 0 |   143    57.7  57.7 |
#>  |                 1 |    68    27.4  85.1 |
#>  |                 2 |    37    14.9 100.0 |
#>  +           ------- + ----- ------- ----- +
#>  |             Total |   248   100.0 100.0 |
#>  + --------- ------- + ----- ------- ----- +
}\if{html}{\out{</div>}}

The \code{:} operator selects a range of consecutive variables:\if{html}{\out{<div class="r">}}\preformatted{tab(infert, education, parity:spontaneous)
#>  One-way tabulation
#>  + ----------- ------- + ----- ------- ----- +
#>  |    Variable   Level | Freq. Percent  Cum. |
#>  + ----------- ------- + ----- ------- ----- +
#>  |   education         |                     |
#>  |              0-5yrs |    12     4.8   4.8 |
#>  |             6-11yrs |   120    48.4  53.2 |
#>  |             12+ yrs |   116    46.8 100.0 |
#>  +             ------- + ----- ------- ----- +
#>  |               Total |   248   100.0 100.0 |
#>  + ----------- ------- + ----- ------- ----- +
#>  |      parity         |                     |
#>  |                   1 |    99    39.9  39.9 |
#>  |                   2 |    81    32.7  72.6 |
#>  |                   3 |    36    14.5  87.1 |
#>  |                   4 |    18     7.3  94.4 |
#>  |                   5 |     6     2.4  96.8 |
#>  |                   6 |     8     3.2 100.0 |
#>  +             ------- + ----- ------- ----- +
#>  |               Total |   248   100.0 100.0 |
#>  + ----------- ------- + ----- ------- ----- +
#>  |     induced         |                     |
#>  |                   0 |   143    57.7  57.7 |
#>  |                   1 |    68    27.4  85.1 |
#>  |                   2 |    37    14.9 100.0 |
#>  +             ------- + ----- ------- ----- +
#>  |               Total |   248   100.0 100.0 |
#>  + ----------- ------- + ----- ------- ----- +
#>  |        case         |                     |
#>  |                   0 |   165    66.5  66.5 |
#>  |                   1 |    83    33.5 100.0 |
#>  +             ------- + ----- ------- ----- +
#>  |               Total |   248   100.0 100.0 |
#>  + ----------- ------- + ----- ------- ----- +
#>  | spontaneous         |                     |
#>  |                   0 |   141    56.9  56.9 |
#>  |                   1 |    71    28.6  85.5 |
#>  |                   2 |    36    14.5 100.0 |
#>  +             ------- + ----- ------- ----- +
#>  |               Total |   248   100.0 100.0 |
#>  + ----------- ------- + ----- ------- ----- +
}\if{html}{\out{</div>}}

Tabulate all categorical variables without specifying any variables:\if{html}{\out{<div class="r">}}\preformatted{tab(infert)
#>  One-way tabulation
#>  + --------- ------- + ----- ------- ----- +
#>  |  Variable   Level | Freq. Percent  Cum. |
#>  + --------- ------- + ----- ------- ----- +
#>  | education         |                     |
#>  |            0-5yrs |    12     4.8   4.8 |
#>  |           6-11yrs |   120    48.4  53.2 |
#>  |           12+ yrs |   116    46.8 100.0 |
#>  +           ------- + ----- ------- ----- +
#>  |             Total |   248   100.0 100.0 |
#>  + --------- ------- + ----- ------- ----- +
}\if{html}{\out{</div>}}

Tabulate variables with a stratification variable \code{by}:\if{html}{\out{<div class="r">}}\preformatted{tab(infert, education, parity:spontaneous, by = case)
#>  Two-way tabulation
#>                                  case
#>  + ----------- ------- + --- ----- -- ----- ----- ----- ----- ----- +
#>  |    Variable   Level |   0  r(\%)  1  r(\%) Total  r(\%)  chi2 exact |
#>  + ----------- ------- + --- ----- -- ----- ----- ----- ----- ----- +
#>  |   education         |                                            |
#>  |              0-5yrs |   8  66.7  4  33.3    12 100.0 0.999 1.000 |
#>  |             6-11yrs |  80  66.7 40  33.3   120 100.0             |
#>  |             12+ yrs |  77  66.4 39  33.6   116 100.0             |
#>  +             ------- + --- ----- -- ----- ----- ----- ----- ----- +
#>  |               Total | 165  66.5 83  33.5   248 100.0             |
#>  + ----------- ------- + --- ----- -- ----- ----- ----- ----- ----- +
#>  |      parity         |                                            |
#>  |                   1 |  66  66.7 33  33.3    99 100.0 1.000 1.000 |
#>  |                   2 |  54  66.7 27  33.3    81 100.0             |
#>  |                   3 |  24  66.7 12  33.3    36 100.0             |
#>  |                   4 |  12  66.7  6  33.3    18 100.0             |
#>  |                   5 |   4  66.7  2  33.3     6 100.0             |
#>  |                   6 |   5  62.5  3  37.5     8 100.0             |
#>  +             ------- + --- ----- -- ----- ----- ----- ----- ----- +
#>  |               Total | 165  66.5 83  33.5   248 100.0             |
#>  + ----------- ------- + --- ----- -- ----- ----- ----- ----- ----- +
#>  |     induced         |                                            |
#>  |                   0 |  96  67.1 47  32.9   143 100.0 0.964 0.969 |
#>  |                   1 |  45  66.2 23  33.8    68 100.0             |
#>  |                   2 |  24  64.9 13  35.1    37 100.0             |
#>  +             ------- + --- ----- -- ----- ----- ----- ----- ----- +
#>  |               Total | 165  66.5 83  33.5   248 100.0             |
#>  + ----------- ------- + --- ----- -- ----- ----- ----- ----- ----- +
#>  |        case         |                                            |
#>  |                   0 | 165 100.0  0   0.0   165 100.0 0.000 0.000 |
#>  |                   1 |   0   0.0 83 100.0    83 100.0             |
#>  +             ------- + --- ----- -- ----- ----- ----- ----- ----- +
#>  |               Total | 165  66.5 83  33.5   248 100.0             |
#>  + ----------- ------- + --- ----- -- ----- ----- ----- ----- ----- +
#>  | spontaneous         |                                            |
#>  |                   0 | 113  80.1 28  19.9   141 100.0 0.000 0.000 |
#>  |                   1 |  40  56.3 31  43.7    71 100.0             |
#>  |                   2 |  12  33.3 24  66.7    36 100.0             |
#>  +             ------- + --- ----- -- ----- ----- ----- ----- ----- +
#>  |               Total | 165  66.5 83  33.5   248 100.0             |
#>  + ----------- ------- + --- ----- -- ----- ----- ----- ----- ----- +
}\if{html}{\out{</div>}}
}

\concept{statistical analysis}
