% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{tab}
\alias{tab}
\alias{tab.default}
\alias{tab.character}
\alias{tab.list}
\alias{tab.data.frame}
\title{Tabulation}
\usage{
tab(x, data = NULL, na.rm = FALSE, rnd = 1)

\method{tab}{default}(...)

\method{tab}{character}(x, data = NULL, na.rm = FALSE, rnd = 1)

\method{tab}{list}(x, data = NULL, na.rm = FALSE, rnd = 1)

\method{tab}{data.frame}(x, data = NULL, na.rm = FALSE, rnd = 1)
}
\arguments{
\item{x}{a character, factor or logical object OR list OR data frame}

\item{data}{a data frame object (Optional)}

\item{na.rm}{A logical value to specify missing values, <NA> in the table}

\item{rnd}{specify rounding of numbers. See \code{\link{round}}.}

\item{...}{optional arguments}
}
\description{
\code{tab} generates frequency distribution of a categorical variable.
}
\details{
\code{tab}

produces one-way table of frequencies.
Exploring data before jumping into complex analysis is always a necessity.
The first step of an analysis is always to summarize and display data.

If variables are specified, all data types are accepted, except for dataframe,
it tabulates all variables with three data types: 1) character, 2) factor, and
3) logical.

\strong{References:}
\enumerate{
  \item Essential Medical Statistics, Betty R. Kirkwood & Jonathan A.C. Sterne,
  Second Edition. Chapter 3
  \item An Introduction to MEdical Statistics, Martin Bland, Thrid Edition,
  Chapter 4
}
}
\examples{
\dontrun{
# factor
tab(infert$education)
# numeric
tab(infert$case)

# multiple variables of mixed types
tab(c(infert$case, infert$education, infert$induced))
tab(c(case, parity, induced), infert)

# dataframe
tab(infert)
tab(iris)

# errors: no categorical variables
tab(mtcars)
}
}
\seealso{
\code{\link{xtab}}, \code{\link{summ}}, \code{\link{summBy}}
}
\author{
Myo Minn Oo (Email: \email{dr.myominnoo@gmail.com} |
Website: \url{https://myominnoo.github.io/})
}
\keyword{descriptive}
\keyword{distribution,}
\keyword{frequency}
\keyword{one-way}
\keyword{statistics,}
\keyword{table,}
\keyword{tabulation,}
