% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{tab}
\alias{tab}
\title{Tabulation}
\usage{
tab(data, ..., by = NULL, row.pct = TRUE, na.rm = FALSE, rnd = 1)
}
\arguments{
\item{data}{Dataset}

\item{...}{Variable or multiple variables
Colon separator \code{:} can be used to specify multiple variables.}

\item{by}{Varaiable for cross-tabulation}

\item{row.pct}{\code{TRUE}, \code{FALSE} or \code{NULL}:

\code{TRUE} shows row percentages.

\code{FALSE} shows column percentages.

\code{NULL} shows no percentages.}

\item{na.rm}{A logical value to specify missing values,}

\item{rnd}{specify rounding of numbers. See \code{\link{round}}.}
}
\value{
tabulation as \code{list}
}
\description{
\code{tab()} generates one-way or two-way tabulation of variables.
}
\details{
\strong{One-way tabulation}

If \code{by} is not specified, \code{tab} generates one-way tabulation of
a variable or multiple variables. \code{...} accepts multiple variables and
produces corresponding tabulations.

Tabulation is displayed in \code{Freq.} (frequency), \code{Percent.}
(Relative Frequency) and \code{Cum.Percent.} (Cumulative Relative frequency).

\preformatted{tab(data, var1)}

\preformatted{tab(data, var1, var2, var3:var5, var10)}

\strong{Two-way tabulation}

Specifying \code{by} produces two-way tables. P-values from
Chi-squared and Fisher's Exact tests are also shown.

\strong{Data type}

Tabulation of the whole dataset requires variables to be in either of
these data types: \code{character}, \code{factor}, \verb{order factor}, \code{logical}.
\subsection{if \code{...} is not specified, tabulation of the whole dataset is produced.}{

\preformatted{tab(data)}

\strong{Using colon \code{:} spearator}

Colon separator \code{:} can be used to indicate sequence of variables.

\preformatted{tab(data, var1, var2, var3:var5, var10)}
}
}
\examples{
\dontrun{

## UCLA IDRE Example
## Website:
path <- "https://stats.idre.ucla.edu/stat/data/hsbdemo.dta"
hsb <- haven::read_dta(path)
codebook(hsb)


## to use piping function
library(magrittr)


## one variable
tab(hsb, female)
tab(hsb, schtyp)
tab(hsb, prog)


## multiple variables
tab(hsb, female, ses, schtyp, prog)
tab(hsb, female:prog)


## the whole dataset
## all variables are numeric so this does not work.
tab(hsb)

## some recoding for hsb
hsb \%>\%
    recode(female, 0:1, c("male", "female")) \%>\%
    recode(ses, 1:3, c("low", "mid", "high")) \%>\%
    recode(schtyp, 1:2, c("public", "private")) \%>\%
    recode(prog, 1:3, c("general", "academic", "vocation")) \%>\%
    tab




# Example from IDRE UCLA
path <- "https://stats.idre.ucla.edu/stat/data/patient_pt1_stata_dm.dta"
hosp <- haven::read_dta(path)
codebook(hosp)


## to use piping function
library(magrittr)


## tabulation of hosp
tab(hosp)


hosp \%>\%
    recode(married, 0:1, c("No", "Yes")) \%>\%
    # recode(sex, 12.2, NA) \%>\% ## same statement as the command below
    replace(sex, NA, sex == 12.2) \%>\% ## same here
    replace(wbc, as.numeric(wbc)) \%>\%
    tab
}

}
\references{
\enumerate{
\item Essential Medical Statistics, Betty R. Kirkwood & Jonathan
A.C. Sterne,
Second Edition.
\item An Introduction to MEdical Statistics, Martin Bland,
Thrid Edition,
\item STATA DATA MANAGEMENT. UCLA: Statistical Consulting Group.
from https://stats.idre.ucla.edu/stata/seminars/stata-data-management/
(accessed Febrary 25, 2020).
}
}
\author{
For any feedback, please contact \code{Myo Minn Oo} via:

Email: \email{dr.myominnoo@gmail.com}

Website: \url{https://myominnoo.github.io/}
}
