% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableToDF.R
\name{tableToDF}
\alias{tableToDF}
\title{Table to Data Frame (Reverse Table)}
\usage{
tableToDF(table, var.names = NULL, x.name = NULL, y.name = NULL)
}
\arguments{
\item{table}{a table input}

\item{var.names}{specify names of two variables}

\item{x.name}{a vector corresponding to the levels in Column variable}

\item{y.name}{a vector corresponding to the levels in Row variable}
}
\description{
\code{\link{tableToDF}} transforms a contigency table into individual data frame.
}
\details{
by default, it takes a table and returns a data frame with values corresponding to the
numbers in that table. This is only for categorical data. If names are not specified,
it returns data that contains "x" and "y" variables. Names and levels can be specified. See the following examples.
}
\examples{
t <- table(infert$education, infert$case)
t

## get raw data from table
tableToDF(t)

## change variables' names
tableToDF(t, var.names = c("edu", "case"))

## specify names of levels of variables
tableToDF(t, x.name = c("No", "Yes"))

str(tableToDF(t, var.names = c("edu", "case"), x.name = c("No", "Yes")))
table(tableToDF(t, var.names = c("edu", "case"), x.name = c("No", "Yes")))
}
\seealso{
\code{\link{isum}}, \code{\link{itab}}, \code{\link{inumsum}},
\code{\link{ibarplot}}, \code{\link{iboxplot}}, \code{\link{ikdplot}},
\code{\link{plotOR}}
}
\author{
Myo Minn Oo (Email: \email{dr.myominnoo@gmail.com} |
Website: \url{https://myominnoo.github.io/})
}
\keyword{data}
\keyword{frame,}
\keyword{get}
\keyword{raw}
\keyword{table}
\keyword{to}
