% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabOdds.R
\name{tabOdds}
\alias{tabOdds}
\title{Calculating Odds}
\usage{
tabOdds(
  data,
  ...,
  by,
  exp_value = NULL,
  case_value = NULL,
  plot = TRUE,
  na.rm = FALSE,
  rnd = 3
)
}
\arguments{
\item{data}{Dataset}

\item{...}{Variable or multiple variables
Colon separator \code{:} can be used to specify multiple variables.}

\item{by}{Varaiable for cross-tabulation}

\item{exp_value}{value for exposure as reference}

\item{case_value}{value for outcome as reference}

\item{plot}{logical value to display plots of rates across a categorical
variable}

\item{na.rm}{A logical value to specify missing values, \code{NA} in the table}

\item{rnd}{specify rounding of numbers. See \code{\link{round}}.}
}
\description{
\code{tabOdds} generates cross-tabulation between two variables and
display odds of failure \code{var_case} among exposure variable
\code{var_exp}. It is used in case-control studies.
}
\details{
A table tabulating odds and corresponding 95\\% confidence interval
is generated.

\strong{Formula for calculating Odds}

\deqn{OR = d1 x h0 / d0 x h1}

\strong{Error Factor} (EF)

\deqn{EF = exp(1.96 x SE(log odds))}

\deqn{SE(log odds) = \sqrt{1/d + 1/h}}
}
\examples{
\dontrun{


#' ## Example from Altman Statistics with Confidence, Chapter 7, Page 62
abo <- expandTables(c(54, 60, 89, 245),
                       exp_name = "study",
                       exp_lvl = c("cases", "controls"),
                       case_name = "state",
                       case_lvl = c("Yes", "No")) \%>\%
    labelVar(c(study, state),
             c("study group", "ABO non-secretor state")) \%>\%
    labelData(paste0("ABO non-secretor state for 114 patients with",
                     " spondyloarthropathies and 334 controls"))


## check dataset
codebook(abo)


## tabulate odds
tabOdds(abo, study, by = state)

## change case value to "Yes"
tabOdds(abo, study, by = state, case_value = "Yes")

## without plot
tabOdds(abo, study, by = state, case_value = "Yes", plot = FALSE)
}

}
\references{
\enumerate{
\item Essential Medical Statistics, Betty R. Kirwood, Second
Edition\preformatted{\\item Statistics Notes: The odds ratio; J Martin Bland, Douglas G Altman
}

BMJ 2000;320:1468

\item Altman, Statistics with confidence
}
}
\author{
For any feedback, please contact \code{Myo Minn Oo} via:

Email: \email{dr.myominnoo@gmail.com}

Website: \url{https://myominnoo.github.io/}
}
