% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabRisks.R
\name{tabRisks}
\alias{tabRisks}
\alias{tblRowColOrder}
\alias{plotRisks}
\title{Calculating Risks}
\usage{
tabRisks(
  data = NULL,
  var_exp,
  var_case,
  ref_exp = NULL,
  ref_case = NULL,
  na.rm = FALSE,
  rnd = 3,
  plot = TRUE,
  print.table = TRUE
)

tblRowColOrder(t, ref_row = NULL, ref_col = NULL)

plotRisks(t, exp_name, case_name, ylab = "Risks")
}
\arguments{
\item{data}{a data frame object (Optional)}

\item{var_exp}{Exposure variable.}

\item{var_case}{Case or outcome variable should be binary vector.}

\item{ref_exp}{Exposed level}

\item{ref_case}{outcome}

\item{na.rm}{A logical value to specify missing values, <NA> in the table}

\item{rnd}{specify rounding of numbers. See \code{\link{round}}.}

\item{plot}{logical value to display plots of rates across a categorical
variable}

\item{print.table}{logical value to display formatted outputs}

\item{t}{table}

\item{ref_row}{name of row levels}

\item{ref_col}{name of column levels}

\item{exp_name}{name of exposure variable}

\item{case_name}{name of outcome variable}

\item{ylab}{name of y axis}
}
\description{
\code{tabRisks()} cross-tabulates two variables and
reports risks of failure \code{var_case} among exposed and
unexposed levels of explanatory variable \code{var_exp}.
It is used with cross-sectional data.

\code{tblRowColOrder()} changes the order of levels in rows or
columns by specifying names of levels.

\code{plotRisks()} produces a bar plot with estimates and
corresponding CIs.
}
\details{
The variable \code{var_case} typically should be coded
1 for case and 0 for non-case. A tabulation with Risks and
corresponding 95\% Confidence Intervals (CI) is displayed.

Risks are sometimes called proportions, incidence proportions or
prevalence.

95\% CI is calculated by large sample method.

\strong{Calculating Risks}

\deqn{Risks = Outcome of Interest (A) / Sample Size (n)}

\deqn{Standard Error, SE = \sqrt(p x (1 - p) / n)}

\deqn{95\% CI = Risks +/- (1.96 x SE)}

This function is used in calculating risks and odds.
}
\section{Functions}{
\itemize{
\item \code{tblRowColOrder}: 

\item \code{plotRisks}: 
}}

\note{
This method should be avoided in small samples. Quadratic or
exact binomial methods are preferred in this regard.
}
\examples{
\dontrun{
### Simpson's Paradox: Example from Burt Gerstman's Epidemiology
### Chapter 14, Page 326

library(magrittr)
simpson <- expandTables(
    "1" = c(1000, 9000, 50, 950),
    "2" = c(95, 5, 5000, 5000),
    exp_name = "treatment",
    exp_lvl = c("new", "standard"),
    case_name = "outcome",
    case_lvl = c("alive", "dead"),
    strata_name = "clinic") \%>\%
    labelVars(c(treatment, outcome, clinic),
              c("Treatment: new or standard",
                "Outcome: alive or dead", "clinic: 1 or 2")) \%>\%
    labelData("Example of Simpson's Paradox")

## checking structure
codebook(simpson)

tabRisks(simpson, treatment, outcome)
tabRisks(simpson, treatment, outcome, ref_exp = "standard")
tabRisks(simpson, treatment, outcome, ref_case = "dead")
tabRisks(simpson, treatment, outcome, "standard", "dead")

## Risks for Clinic 1
simpson \%>\%
    pick(clinic == 1) \%>\%
    tabRisks(treatment, outcome)

## Risks for Clinic 2
simpson \%>\%
    pick(clinic == 2) \%>\%
    tabRisks(treatment, outcome)





## Asthma Example from Essential Medical Statistics
## page 160

library(magrittr)
asthma <- expandTables(c(81, 995, 57, 867),
              exp_name = "sex",
              exp_lvl = c("Women", "Man"),
              case_name = "asthma",
              case_lvl = c("Yes", "No")) \%>\%
          labelData("Hypothetical Data of Asthma Prevalence") \%>\%
          labelVars(c(sex, asthma),
              c("Man or Woman", "Asthma or No Asthma"))

## Checking codebook
codebook(asthma)

tabRisks(asthma, sex, asthma)






## The odds ratio, J Martin Bland, Douglas G Altman, BMJ 2000;320:1468
hay <- expandTables(c(141, 420, 928, 13525),
                    exp_name = "eczema",
                    exp_lvl = c("Yes", "No"),
                    case_name = "hayFever",
                    case_lvl = c("Yes", "No")) \%>\%
       labelData("hay fever and eczema in 11 year old children") \%>\%
       labelVars(c(eczema, hayFever),
                c("prevalence of eczema", "prevalence of hay fever"))

codebook(hay)

tabRisks(hay, eczema, hayFever)
}
}
\references{
B. Burt Gerstman. Epidemology Kept Simple. Wiley-Blackwell, 2013:
Chapter 13, Confidence Intervals and p-values
}
\seealso{
\code{\link{tabOdds}}, \code{\link{mhodds}}

\code{\link{tabRisks}}, \code{\link{tabOdds}}

\code{\link{tabRisks}}, \code{\link{tabOdds}}
}
\author{
For any feedback, please contact \code{Myo Minn Oo} via:

Email: \email{dr.myominnoo@gmail.com}

Website: \url{https://myominnoo.github.io/}
}
\keyword{Risks,}
\keyword{descriptive}
\keyword{frequency}
\keyword{incidence,}
\keyword{prevalence,}
\keyword{ratio,}
\keyword{risk}
\keyword{statistics,}
\keyword{table,}
