% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtab.R
\name{xtab}
\alias{xtab}
\alias{xtab.default}
\alias{xtab.character}
\alias{xtab.list}
\alias{xtab.data.frame}
\title{Cross tabulation, contigency table or two-by-two table}
\usage{
xtab(
  data = NULL,
  ...,
  by,
  row.pct = TRUE,
  na.rm = FALSE,
  rnd = 1,
  print.table = TRUE
)

\method{xtab}{default}(
  data = NULL,
  ...,
  by,
  row.pct = TRUE,
  na.rm = FALSE,
  rnd = 1,
  print.table = TRUE
)

\method{xtab}{character}(
  data = NULL,
  x,
  by,
  row.pct = TRUE,
  na.rm = FALSE,
  rnd = 1,
  print.table = TRUE
)

\method{xtab}{list}(
  data = NULL,
  ...,
  by,
  row.pct = TRUE,
  na.rm = FALSE,
  rnd = 1,
  print.table = TRUE
)

\method{xtab}{data.frame}(
  data = NULL,
  ...,
  by = NULL,
  row.pct = TRUE,
  na.rm = FALSE,
  rnd = 1,
  print.table = TRUE
)
}
\arguments{
\item{data}{dataset (optional)}

\item{...}{a variable or variables. Colon separator \code{:} can be
used for multiple variables. See details.}

\item{by}{outcome variable for cross-tabulation}

\item{row.pct}{TRUE or FALSE: If \code{TRUE}, row percentages
are shown and if \code{FALSE}, column percentages.
if \code{NULL}, a table without any row or column percentages
is produced.}

\item{na.rm}{A logical value to specify missing values,
\code{<NA>} in the table}

\item{rnd}{specify rounding of numbers. See \code{\link{round}}.}

\item{print.table}{logical value to display formatted outputs}

\item{x}{exposure variable}
}
\value{
\enumerate{
   \item \code{vector} or \code{list}: corresponding unformatted table(s)
   \item output: formatted texts
}
}
\description{
\code{xtab()} generates cross tabulation of two variables.
}
\details{
Exploring data before jumping into complex analysis is always a
necessity.
The first step of an analysis is always to summarize and display data.

\code{xtab()} produces \code{2x2} or contigency tables.


\strong{Two-way tabulation}

When \code{by} is specified, \code{tab()} or \code{xtab()} generates
two-way tabulation of a variable or multiple variables.
THere is no argument for \code{x}. Instead, \code{...} accepts
multiple variables and produces corresponding tabulations.

Tabulation is displayed in number for corresponding categories
and corresponding row \code{(r\%)} or column \code{(c\%)}
percentages. Totals for both direction are also produced along
with different tests as follows:

\enumerate{
   \item Pearson Chi-squared test of Independence \code{Chi.Square}
   \item Fisher's Exact Test \code{F.Exact}
}

\strong{Row or column percentages}

If \code{row.pct} is \code{TRUE}, row percentages are displayed.
If not, column percentages will be tabulated. If put as \code{NULL},
a default table without percentages will be displayed.
}
\examples{
\dontrun{


## Cross-tabulation
## using infert dataset
# one variable
xtab(infert, education, by = case)

# multiple variables
xtab(infert, education, age:induced, spontaneous, by = case)

# the whole dataset
xtab(infert, by = case)



## IDRE UCLA Example 1
## using hosp dataset
path <- "https://stats.idre.ucla.edu/stat/data/patient_pt1_stata_dm.dta"
hosp <- haven::read_dta(path)
codebook(hosp)

# one variable
xtab(hosp, sex, by = remission)

# multiple variables
xtab(hosp, hospital, married:lengthofstay, by = remission)

# the whole dataset
xtab(hosp, by = remission)


## Cross-tabulation
## using infert dataset
# one variable
tab(infert, education, by = case)

# multiple variables
tab(infert, education, age:induced, spontaneous, by = case)

# the whole dataset
tab(infert, by = case)




## IDRE UCLA Example 1
## using hosp dataset
path <- "https://stats.idre.ucla.edu/stat/data/patient_pt1_stata_dm.dta"
hosp <- haven::read_dta(path)
codebook(hosp)

# one variable
tab(hosp, sex, by = remission)

# multiple variables
tab(hosp, hospital, married:lengthofstay, by = remission)

# the whole dataset
tab(hosp, by = remission)

}
}
\references{
\enumerate{
  \item Essential Medical Statistics, Betty R. Kirkwood & Jonathan
  A.C. Sterne,
  Second Edition.
  \item An Introduction to Medical Statistics, Martin Bland,
  Thrid Edition,
  \item STATA DATA MANAGEMENT. UCLA: Statistical Consulting Group.
   from https://stats.idre.ucla.edu/stata/seminars/stata-data-management/
   (accessed Febrary 25, 2020).
}
}
\seealso{
\code{\link{tab}}
}
\author{
For any feedback, please contact \code{Myo Minn Oo} via:

Email: \email{dr.myominnoo@gmail.com}

Website: \url{https://myominnoo.github.io/}
}
\keyword{2x2}
\keyword{descriptive}
\keyword{statistics,}
\keyword{table,}
\keyword{two-by-two}
\keyword{two-way}
